Revert changes in storefront-cart dropin
