+++ scripts/__dropins__/storefront-cart/chunks/CartSummaryGrid.js
@@ -1,4 +1,4 @@
 /*! Copyright 2026 Adobe
 All Rights Reserved. */
-import{jsx as t,Fragment as u}from"@dropins/tools/preact-jsx-runtime.js";import{useState as h,useEffect as C}from"@dropins/tools/preact-compat.js";import{b,E as y}from"./components.js";import{Slot as D}from"@dropins/tools/lib.js";import{Image as I}from"@dropins/tools/components.js";import"@dropins/tools/preact-hooks.js";import{events as S}from"@dropins/tools/event-bus.js";import{getCartDataFromCache as j}from"../api.js";const L=({children:O,initialData:s=null,routeProduct:n,routeEmptyCartCTA:i,slots:e,...d})=>{const[m,l]=h(s);C(()=>{const a=S.on("cart/data",r=>{l(r)},{eager:!0});return()=>{a==null||a.off()}},[]);const g=(a,r)=>{const f=a.selectedOptions?`${a.name}: ${Object.entries(a.selectedOptions).join("; ")}`:a.name,o={loading:r<4?"eager":"lazy",src:a.image.src,alt:a.image.alt},c=t(I,{"data-testid":"cart-grid-item-image","aria-label":f,width:"100%",...o});return t("div",{className:"cart-cart-summary-grid__item-container",children:t(D,{name:"Thumbnail",slot:e==null?void 0:e.Thumbnail,context:{item:a,defaultImageProps:o},children:n?t("a",{href:n(a),children:c}):c})},r)},p=m&&t(u,{children:m.items.map((a,r)=>g(a,r))});return t(b,{...d,emptyCart:t(y,{ctaLinkURL:i?i():void 0}),products:p})};L.getInitialData=async function(){return j()};export{L as C};
+import{jsx as a,Fragment as g}from"@dropins/tools/preact-jsx-runtime.js";import{useState as _,useEffect as C}from"@dropins/tools/preact-compat.js";import{E as b}from"./EmptyCart2.js";import{classes as l,VComponent as v,Slot as I}from"@dropins/tools/lib.js";/* empty css                */import{Image as N}from"@dropins/tools/components.js";import"@dropins/tools/preact-hooks.js";import{events as S}from"@dropins/tools/event-bus.js";import{g as j}from"./persisted-data.js";const D=({className:c,children:i,emptyCart:m,products:e,...n})=>a("div",{...n,className:l(["cart-cart-summary-grid",c]),children:a(g,{children:a("div",{className:l(["cart-cart-summary-grid__content",["cart-cart-summary-grid__content--empty",!e]]),children:e||a(v,{node:m,className:"cart-cart-summary-grid__empty-cart"})})})}),L=({children:c,initialData:i=null,routeProduct:m,routeEmptyCartCTA:e,slots:n,...p})=>{const[o,u]=_(i);C(()=>{const r=S.on("cart/data",t=>{u(t)},{eager:!0});return()=>{r==null||r.off()}},[]);const f=(r,t)=>{const y=r.selectedOptions?`${r.name}: ${Object.entries(r.selectedOptions).join("; ")}`:r.name,s={loading:t<4?"eager":"lazy",src:r.image.src,alt:r.image.alt},d=a(N,{"data-testid":"cart-grid-item-image","aria-label":y,width:"100%",...s});return a("div",{className:"cart-cart-summary-grid__item-container",children:a(I,{name:"Thumbnail",slot:n==null?void 0:n.Thumbnail,context:{item:r,defaultImageProps:s},children:m?a("a",{href:m(r),children:d}):d})},t)},h=o&&a(g,{children:o.items.map((r,t)=>f(r,t))});return a(D,{...p,emptyCart:a(b,{ctaLinkURL:e?e():void 0}),products:h})};L.getInitialData=async function(){return j()};export{L as C};
 //# sourceMappingURL=CartSummaryGrid.js.map
