+++ scripts/__dropins__/storefront-cart/chunks/CartSummaryList.js
@@ -1,4 +1,4 @@
 /*! Copyright 2026 Adobe
 All Rights Reserved. */
-import{jsx as r,Fragment as qt,jsxs as X}from"@dropins/tools/preact-jsx-runtime.js";import{forwardRef as dn,useState as _,useEffect as z,useCallback as Gt,useImperativeHandle as sn}from"@dropins/tools/preact-compat.js";import{Slot as m,classes as j}from"@dropins/tools/lib.js";import{f as Ut,E as ln,g as mn,h as fn,i as Xt,j as gn,k as vn}from"./components.js";import{InLineAlert as et,CartList as rt,Icon as W,Button as V,Accordion as yn,AccordionSection as hn,CartItem as pn,Price as $,Incrementer as Cn,Image as In}from"@dropins/tools/components.js";import"@dropins/tools/preact-hooks.js";import{s as kn,addProductsToCart as Sn,getCartDataFromCache as Wt,updateProductsFromCart as zt}from"../api.js";import{events as jt}from"@dropins/tools/event-bus.js";import{useText as An}from"@dropins/tools/i18n.js";const k=R=>new DOMParser().parseFromString(R,"text/html").body.textContent||"",wn=dn(({initialData:R=null,hideHeading:it,hideFooter:Z,routeProduct:B,routeEmptyCartCTA:J,routeCart:S,onItemUpdate:L,onItemRemove:K,maxItems:C,slots:i,attributesToHide:l=[],enableRemoveItem:N,enableUpdateItemQuantity:A,onItemsErrorsChange:Y,accordion:Vt=!1,variant:M="primary",isLoading:Zt,showMaxItems:at,showDiscount:ct,showSavings:ot,quantityType:ut,dropdownOptions:dt,undo:Q=!1,confirmBeforeDelete:H=!1,includeOutOfStockItems:F=!1,__testRecentlyRemovedItems:st,...lt},Jt)=>{var Ft;const[E,Kt]=_(!R),[a,Yt]=_(R),[v,Mt]=_(new Set),[D,mt]=_(new Map),[ft,tt]=_(new Set),[w,x]=_(st||[]);z(()=>{},[st,w]);const nt=typeof A=="object"?A.removeOnZero??!1:!1,f=(Ft=kn.config)==null?void 0:Ft.shoppingCartDisplaySetting,[T,Ht]=_(at?!0:!C&&!at),c=An({file:"Cart.CartItem.file",files:"Cart.CartItem.files",heading:"Cart.Cart.heading",message:"Cart.CartItem.message",recipient:"Cart.CartItem.recipient",regularPrice:"Cart.CartItem.regularPrice",discountedPrice:"Cart.CartItem.discountedPrice",sender:"Cart.CartItem.sender",lowInventory:"Cart.CartItem.lowInventory",insufficientQuantity:"Cart.CartItem.insufficientQuantity",insufficientQuantityGeneral:"Cart.CartItem.insufficientQuantityGeneral",outOfStockHeading:"Cart.OutOfStockMessage.heading",outOfStockDescription:"Cart.OutOfStockMessage.message",outOfStockAlert:"Cart.OutOfStockMessage.alert",removeAction:"Cart.OutOfStockMessage.action",notAvailableMessage:"Cart.CartItem.notAvailableMessage",viewMore:"Cart.Cart.viewMore",viewAll:"Cart.Cart.viewAll",discountPercent:"Cart.CartItem.discountPercentage",savingsAmount:"Cart.CartItem.savingsAmount",itemBeingRemoved:"Cart.CartItem.itemBeingRemoved",itemRemoved:"Cart.CartItem.itemRemoved",itemRemovedDescription:"Cart.CartItem.itemRemovedDescription",undoAction:"Cart.CartItem.undoAction",dismissAction:"Cart.CartItem.dismissAction",confirmDeleteHeading:"Cart.CartItem.confirmDeleteHeading",confirmAction:"Cart.CartItem.confirmAction",cancelAction:"Cart.CartItem.cancelAction",remove:"Dropin.CartItem.remove.label",removeDefault:"Dropin.CartItem.removeDefault.label",quantity:"Dropin.CartItem.quantity.label"}),O=(t,n)=>{Mt(e=>(n?e.add(t):e.delete(t),new Set(e)))},q=(t,n)=>{mt(e=>(n?e.set(t,n):e.delete(t),new Map(e)))},gt=t=>{H?tt(n=>new Set(n).add(t.uid)):P(t,0,!0)};function tn(t){const n=Wt(),e=n==null?void 0:n.items.find(o=>o.uid===t);e&&x(o=>{const d=n.items.findIndex(s=>s.uid===t);return d!==-1&&!o.some(s=>s.item.uid===t)?(jt.emit("cart/product/removed",{item:e,index:d}),[...o,{item:e,index:d,error:void 0,beingRemoved:!0}]):o})}const P=(t,n,e=!1)=>{if(O(t.uid,!0),mt(new Map),N&&n===0&&(nt||e))tn(t.uid),zt([{uid:t.uid,quantity:n}]).then(()=>{K==null||K({item:t})}).finally(()=>{O(t.uid,!1),x(o=>o.map(d=>d.item.uid===t.uid?{...d,beingRemoved:!1}:d))}).catch(o=>{console.warn(o)});else{if(n===0&&!nt){O(t.uid,!1);return}A&&zt([{uid:t.uid,quantity:n}]).then(()=>{L==null||L({item:t})}).finally(()=>{O(t.uid,!1)}).catch(o=>{q(t.uid,o.message)})}},vt=t=>{x(n=>n.filter(e=>e.item.uid!==t))},nn=Gt(()=>{Ht(t=>!t)},[]);z(()=>{const t=jt.on("cart/data",n=>{Yt(n),Kt(!!Zt)},{eager:!0});return()=>{t==null||t.off()}},[]),z(()=>{Q||x([])},[Q]),z(()=>{Y&&Y(D)},[D,Y]);const yt=(t,n)=>{if(l.includes("image"))return;const e={loading:n<4?"eager":"lazy",src:t.image.src,alt:t.image.alt,width:"300",height:"300",params:{width:300}},o=r(In,{"data-testid":"cart-list-item-image",...e});return r(m,{name:"Thumbnail",slot:i==null?void 0:i.Thumbnail,context:{item:t,defaultImageProps:e},children:B?r("a",{href:B(t),children:o}):o})},ht=t=>{if(!l.includes("name"))return r("span",{"data-testid":"cart-list-item-title",children:B?r("a",{href:B(t),children:k(t.name)}):k(t.name)})},pt=t=>{if(l.includes("configurations"))return;const n={...t.bundleOptions,...t.selectedOptions,...t.customizableOptions,...t.recipient?{[c.recipient]:t.recipient}:null,...t.recipientEmail&&t.recipient?{[c.recipient]:`${t.recipient} (${t.recipientEmail})`}:null,...t.sender?{[c.sender]:t.sender}:null,...t.senderEmail&&t.sender?{[c.sender]:`${t.sender} (${t.senderEmail})`}:{},...t.message?{[c.message]:t.message}:null,...t.links&&t.links.count?t.links.count>1?{[c.files.replace("{count}",t.links.count.toString())]:t.links.result}:{[c.file.replace("{count}",t.links.count.toString())]:t.links.result}:null};if(Object.keys(n).length!==0)return n},Ct=t=>{var n,e,o,d;return(f==null?void 0:f.price)==="INCLUDING_TAX"?t.discounted?{amount:t.regularPrice.value,currency:t.regularPrice.currency,style:{font:"inherit"},"data-testid":"including-tax-item-price"}:{amount:(n=t.taxedPrice)==null?void 0:n.value,currency:(e=t.taxedPrice)==null?void 0:e.currency,style:{font:"inherit"},"data-testid":"including-tax-item-price"}:{amount:(o=t.regularPrice)==null?void 0:o.value,currency:(d=t.regularPrice)==null?void 0:d.currency,style:{font:"inherit"},"data-testid":"regular-item-price"}},It=t=>{var n,e;return{amount:(n=t.savingsAmount)==null?void 0:n.value,currency:(e=t.savingsAmount)==null?void 0:e.currency,style:{font:"inherit"},"data-testid":"item-savings-amount"}},kt=t=>(f==null?void 0:f.price)==="INCLUDING_EXCLUDING_TAX"?r($,{amount:t.rowTotal.value,currency:t.rowTotal.currency,"data-testid":"excluding-tax-total","aria-label":c.regularPrice}):void 0,St=t=>{var o,d,s,u,g,y,p,I,h,Dt;const n={"aria-label":c.regularPrice},e=t.discounted?{}:null;return["INCLUDING_TAX","INCLUDING_EXCLUDING_TAX"].includes(f==null?void 0:f.price)?(n.amount=(o=t.rowTotalIncludingTax)==null?void 0:o.value,n.currency=(d=t.rowTotalIncludingTax)==null?void 0:d.currency,n.variant=t.discounted?"strikethrough":"default",n["data-testid"]="including-tax-item-total",e&&(n.amount=(s=t.total)==null?void 0:s.value,n.currency=(u=t.total)==null?void 0:u.currency,e.amount=(g=t.rowTotalIncludingTax)==null?void 0:g.value,e.currency=(y=t.rowTotalIncludingTax)==null?void 0:y.currency,e.sale=!0,e["aria-label"]=c.discountedPrice,e["data-testid"]="discount-total")):(n.amount=(p=t.total)==null?void 0:p.value,n.currency=(I=t.total)==null?void 0:I.currency,n.variant=t.discounted?"strikethrough":"default",n["data-testid"]="regular-item-total",e&&(e.amount=(h=t.discountedTotal)==null?void 0:h.value,e.currency=(Dt=t.discountedTotal)==null?void 0:Dt.currency,e.sale=!0,e["aria-label"]=c.regularPrice,e["data-testid"]="discount-total")),{totalProps:n,discountProps:e}},At=t=>{var g,y,p;if(l.includes("warning"))return;const n=D.get(t.uid),e=(g=D.get(t.uid))==null?void 0:g.includes("The requested qty is not available"),o=v.has(t.uid),d=t.insufficientQuantity&&t.stockLevel?t.stockLevel==="noNumber"?c.insufficientQuantityGeneral:c.insufficientQuantity.replace("{inventory}",(y=t.stockLevel)==null?void 0:y.toString()).replace("{count}",t.quantity.toString()):"",s=t.lowInventory&&t.onlyXLeftInStock&&c.lowInventory.replace("{count}",(p=t.onlyXLeftInStock)==null?void 0:p.toString()),u=!t.outOfStock&&n&&e?c.notAvailableMessage:n;return!o&&(n||t.insufficientQuantity||t.lowInventory)?X("span",{"data-testid":"item-warning",children:[r(W,{source:gn,size:"16"}),u||d||s]}):void 0},wt=t=>l!=null&&l.includes("alert")?void 0:!v.has(t.uid)&&t.outOfStock?X("span",{"data-testid":"item-alert",children:[r(W,{source:Ut,size:"16"}),c.outOfStockAlert]}):void 0,xt=t=>r(m,{name:"ProductAttributes",slot:i==null?void 0:i.ProductAttributes,context:{item:t}}),Ot=t=>{if(!l.includes("sku"))return r("span",{"data-testid":"cart-list-item-sku",children:t.sku})},Pt=t=>r(m,{name:"Footer",slot:i==null?void 0:i.Footer,context:{item:t,handleItemsLoading:O,handleItemsError:q,onItemUpdate:L}}),bt=t=>r(m,{name:"RowTotalFooter",slot:i==null?void 0:i.RowTotalFooter,context:{item:t}}),en=t=>{var n;if(N)return r(m,{name:"ItemRemoveAction",slot:i==null?void 0:i.ItemRemoveAction,context:{item:t,enableRemoveItem:N,handleItemQuantityUpdate:P,handleItemsError:q,handleItemsLoading:O,onItemUpdate:L,itemsLoading:v},children:r(V,{"data-testid":"cart-item-remove-button",className:j(["dropin-cart-item__remove"]),variant:"tertiary",onClick:()=>gt(t),disabled:v.has(t.uid),icon:r(W,{"data-testid":"cart-item-remove-icon",source:vn,size:"24",stroke:"2",viewBox:"0 0 24 24","aria-label":((n=c.remove)==null?void 0:n.replace("{product}",k(t.name)))||c.removeDefault})})})},rn=t=>{if(!l.includes("quantity")&&A)return r(m,{name:"ItemQuantity",slot:i==null?void 0:i.ItemQuantity,context:{item:t,enableUpdateItemQuantity:A,handleItemQuantityUpdate:P,itemsLoading:v,handleItemsError:q,handleItemsLoading:O,onItemUpdate:L},children:r(Cn,{value:t.quantity,min:nt?0:1,onValue:n=>P(t,Number(n)),name:"quantity",disabled:v.has(t.uid),"aria-label":c.quantity})})},_t=t=>a!=null&&a.totalQuantity?a.items.filter(t).map((n,e)=>{var u;const{totalProps:o,discountProps:d}=St(n),s=r(pn,{updating:v==null?void 0:v.has(n.uid),loading:E,"data-testid":`cart-list-item-entry-${n.uid}`,image:yt(n,e),title:r(m,{name:"ItemTitle",slot:i==null?void 0:i.ItemTitle,context:{item:n},children:ht(n)}),sku:l.includes("sku")?void 0:r(m,{name:"ItemSku",slot:i==null?void 0:i.ItemSku,context:{item:n},children:Ot(n)}),price:l.includes("price")?void 0:r(m,{name:"ItemPrice",slot:i==null?void 0:i.ItemPrice,context:{item:n},slotTag:"span",children:r($,{...Ct(n)})}),quantity:l.includes("quantity")?void 0:n.quantity,quantityContent:rn(n),total:X(qt,{children:[l.includes("total")?void 0:r(m,{name:"ItemTotal",slot:i==null?void 0:i.ItemTotal,context:{item:n},children:r($,{...o})}),l.includes("totalDiscount")?void 0:d&&r(m,{name:"ItemTotal",slot:i==null?void 0:i.ItemTotal,context:{item:n},children:r($,{...d})})]}),attributes:xt(n),configurations:pt(n),totalExcludingTax:l.includes("totalExcludingTax")?void 0:kt(n),taxIncluded:(f==null?void 0:f.price)==="INCLUDING_TAX",taxExcluded:!l.includes("totalExcludingTax")&&(f==null?void 0:f.price)==="INCLUDING_EXCLUDING_TAX",warning:At(n),alert:wt(n),quantityType:ut,dropdownOptions:dt,onQuantity:A&&!v.has(n.uid)?g=>{P(n,g)}:void 0,onRemove:N&&!v.has(n.uid)?()=>gt(n):void 0,discount:ct&&n.discounted&&n.discountPercentage?r("div",{"data-testid":"item-discount-percent",children:c.discountPercent.replace("{discount}",((u=n.discountPercentage)==null?void 0:u.toString())??"")}):void 0,savings:ot&&n.discounted&&n.savingsAmount?X("div",{children:[r("span",{children:r($,{...It(n)})})," ",c.savingsAmount]}):void 0,rowTotalFooter:bt(n),footer:Pt(n),removeContent:en(n)},n.uid);return r(m,{name:"CartItem",slot:i==null?void 0:i.CartItem,context:{item:n,index:e,enableUpdateItemQuantity:A,enableRemoveItem:N,itemsLoading:v,loading:E,attributesToHide:l,cartTaxesConfig:f,quantityType:ut,dropdownOptions:dt,showDiscount:ct,showSavings:ot,dictionary:c,getImage:yt,getTitle:ht,getSku:Ot,getPriceProps:Ct,getTotalPriceProps:St,getProductAttributes:xt,getConfiguration:pt,getTotalExcludingTax:kt,getWarning:At,getAlert:wt,getRowTotalFooter:bt,getFooter:Pt,getSavingsAmount:It},children:s},n.uid)}):null,Lt=r(m,{name:"Heading",slot:i==null?void 0:i.Heading,context:{count:a==null?void 0:a.totalQuantity},children:r("div",{"data-testid":"default-cart-heading",children:c.heading.replace("({count})",a!=null&&a.totalQuantity?`(${a==null?void 0:a.totalQuantity.toString()})`:"")})}),an=Lt.props.children.props.children,cn=()=>{const t=a==null?void 0:a.items.filter(n=>n.outOfStock);t==null||t.forEach(n=>{P(n,0,!0)})},on=_t(t=>t.outOfStock||t.insufficientQuantity||!1),Nt=a!=null&&a.hasOutOfStockItems?r(et,{"data-testid":"cart-out-of-stock-message",icon:r(W,{source:Ut,size:"16"}),itemList:r(rt,{className:"cart-cart-summary-list__out-of-stock-items","data-testid":"out-of-stock-cart-items","aria-hidden":F||void 0,children:on}),type:"warning",heading:c.outOfStockHeading,description:c.outOfStockDescription,variant:"primary",actionButtonPosition:"bottom",additionalActions:a!=null&&a.hasFullyOutOfStockItems&&N?[{label:c.removeAction,onClick:cn}]:void 0}):void 0,b=_t(F?()=>!0:t=>!t.outOfStock&&!t.insufficientQuantity),Qt=T?Math.max(C||5,5):Math.min((a==null?void 0:a.totalQuantity)||5,5),Et=(a==null?void 0:a.totalQuantity)>Qt,un=Et&&!T&&Qt!=C,G=a!=null&&a.totalQuantity&&b?r(m,{name:"Footer",slot:i==null?void 0:i.CartSummaryFooter,context:{displayMaxItems:T,routeCart:S},"data-testid":"cart-cart-summary-footer-slot",className:"cart-cart-summary-footer",children:Et?un?r(V,{className:"cart-cart-summary-list-footer__action",onClick:nn,"data-testid":"view-more-items-button",variant:"tertiary",children:c.viewMore}):S&&r(V,{className:"cart-cart-summary-list-footer__action",href:S(),variant:"tertiary","data-testid":"view-cart-or-less-items-button",children:c.viewAll}):S&&r(V,{className:"cart-cart-summary-list-footer__action",href:S(),variant:"tertiary","data-testid":"view-cart-button",children:c.viewAll})}):null,Tt=()=>{var d;if(!(a!=null&&a.items))return null;const t=[],n=a.items,e=i==null?void 0:i.UndoBanner,o=T?Math.max(C||(a==null?void 0:a.totalQuantity),5):Math.min(C??5,5);for(let s=0;s<=o;s++)if(Q&&w.filter(u=>u.index===s).forEach(({item:u,error:g,beingRemoved:y})=>{var I,h;const p=r(et,{type:g?"error":"info",heading:y?(I=c.itemBeingRemoved)==null?void 0:I.replace("{product}",k(u.name)):(h=c.itemRemoved)==null?void 0:h.replace("{product}",k(u.name)),description:g||c.itemRemovedDescription,variant:"primary",actionButtonPosition:"bottom",additionalActions:y?[]:[{label:c.undoAction,onClick:()=>U(u.uid),"aria-label":`${c.undoAction} remove ${k(u.name)}`},{label:c.dismissAction,onClick:()=>vt(u.uid),"aria-label":`${c.dismissAction} undo for ${k(u.name)}`}]},`undo-banner-${u.uid}`);t.push(e?r(m,{name:"UndoBanner",slot:e,context:{item:u,loading:E,error:g,onUndo:()=>U(u.uid),onDismiss:()=>vt(u.uid)},className:"cart-summary-list__banner cart-summary-list__banner--undo",children:p},`undo-banner-${u.uid}`):r("div",{className:"cart-summary-list__banner cart-summary-list__banner--undo",children:p},`undo-banner-${u.uid}`))}),s<o&&s<n.length){const u=n[s];if(H&&ft.has(u.uid)){const g=()=>{tt(I=>{const h=new Set(I);return h.delete(u.uid),h}),P(u,0,!0)},y=()=>{tt(I=>{const h=new Set(I);return h.delete(u.uid),h})},p=r(et,{type:"warning",heading:(d=c.confirmDeleteHeading)==null?void 0:d.replace("{product}",k(u.name)),variant:"primary",actionButtonPosition:"bottom",additionalActions:[{label:c.confirmAction,onClick:g},{label:c.cancelAction,onClick:y}]},`confirm-delete-banner-${u.uid}`);t.push(r(m,{name:"ConfirmDeleteBanner",slot:i==null?void 0:i.ConfirmDeleteBanner,context:{item:u,loading:v.has(u.uid),onConfirm:g,onCancel:y},className:"cart-summary-list__banner cart-summary-list__banner--confirm-delete",children:p},`confirm-delete-banner-${u.uid}`))}t.push((b==null?void 0:b[s])||null)}return r(rt,{"data-testid":"cart-list",children:t})},U=Gt(async t=>{x(e=>e.map(o=>o.item.uid===t?{...o,loading:!0,error:void 0}:o));const n=w.find(e=>e.item.uid===t);if(n)try{await Sn([{sku:n.item.sku,parentSku:n.item.topLevelSku,quantity:n.item.quantity,optionsUIDs:n.item.bundleOptionsUIDs||(n.item.selectedOptionsUIDs?Object.values(n.item.selectedOptionsUIDs):void 0),enteredOptions:n.item.customizableOptions?Object.entries(n.item.customizableOptions).map(([e,o])=>({uid:e,value:o})):void 0}]),x(e=>e.filter(o=>o.item.uid!==t))}catch(e){x(o=>o.map(d=>d.item.uid===t?{...d,error:(e==null?void 0:e.message)||"Failed to restore item."}:d))}},[w]);sn(Jt,()=>({handleUndo:U}),[U]);const $t=r(qt,{children:Q&&w.length>0?Tt():r(m,{name:"EmptyCart",slot:i==null?void 0:i.EmptyCart,context:{},children:r(ln,{"data-testid":"empty-cart",ctaLinkURL:J==null?void 0:J()})})}),Rt=r(rt,{"data-testid":"cart-list",children:b==null?void 0:b.slice(0,T?Math.max(C||(a==null?void 0:a.totalQuantity),5):Math.min(C??5,5))}),Bt=a!=null&&a.totalQuantity||Q&&w.length>0?(Q||H)&&(w.length>0||ft.size>0)?Tt():Rt:null;return Vt?r(yn,{"data-testid":"cart-summary-list-accordion",className:j(["cart-cart-summary-list-accordion",`cart-cart-summary-list__background--${M}`]),iconOpen:fn,iconClose:mn,children:r(hn,{title:an,"data-testid":"cart-summary-list-accordion__section",open:!0,renderContentWhenClosed:!0,children:r(Xt,{...lt,"aria-expanded":!0,"aria-label":"TEST",className:j(["cart-cart-summary-list-accordion__list",["cart-cart-summary-list--include-out-of-stock",F]]),loading:E,footer:Z?void 0:G||(S?G:void 0),emptyCart:$t,products:Bt,outOfStockMessage:Nt,variant:M})})}):r(Xt,{...lt,className:j([["cart-cart-summary-list--include-out-of-stock",F]]),heading:it?void 0:Lt,footer:Z?void 0:G||(S?G:void 0),loading:E,emptyCart:$t,products:Bt,outOfStockMessage:Nt,variant:M})});wn.getInitialData=async function(){return Wt()};export{wn as C};
+import{jsx as r,jsxs as S,Fragment as z}from"@dropins/tools/preact-jsx-runtime.js";import*as U from"@dropins/tools/preact-compat.js";import{forwardRef as ie,useState as T,useEffect as K,useCallback as Bt,useImperativeHandle as ce}from"@dropins/tools/preact-compat.js";import{classes as I,VComponent as Y,Slot as m}from"@dropins/tools/lib.js";import{E as oe}from"./EmptyCart2.js";/* empty css                */import{Divider as $t,Skeleton as se,SkeletonRow as ue,InLineAlert as Gt,CartList as Z,Icon as M,Button as H,Accordion as de,AccordionSection as le,CartItem as me,Price as X,Incrementer as ge,Image as fe}from"@dropins/tools/components.js";import"@dropins/tools/preact-hooks.js";import{g as zt}from"./persisted-data.js";import{events as qt}from"@dropins/tools/event-bus.js";import{s as he}from"./resetCart.js";import{a as ve}from"./createGuestCart.js";import{u as Ft}from"./updateProductsFromCart.js";import{S as Ut,a as ye,b as pe}from"./WarningWithCircle.js";import{S as Ce}from"./ChevronDown.js";import{useText as Ie}from"@dropins/tools/i18n.js";const ke=k=>U.createElement("svg",{id:"Icon_Chevron_right_Base","data-name":"Icon \\u2013 Chevron right \\u2013 Base",xmlns:"http://www.w3.org/2000/svg",width:24,height:24,viewBox:"0 0 24 24",...k},U.createElement("g",{id:"Large"},U.createElement("rect",{id:"Placement_area","data-name":"Placement area",width:24,height:24,fill:"#fff",opacity:0}),U.createElement("g",{id:"Chevron_right_icon","data-name":"Chevron right icon"},U.createElement("path",{vectorEffect:"non-scaling-stroke",id:"chevron",d:"M199.75,367.5l4.255,-4.255-4.255,-4.255",transform:"translate(-189.25 -351.0)",fill:"none",stroke:"currentColor"})))),Xt=({className:k,children:tt,heading:x,footer:A,emptyCart:B,products:h,outOfStockMessage:y,variant:$="primary",loading:g=!0,...a})=>r("div",{...a,className:I(["cart-cart-summary-list",k,`cart-cart-summary-list__background--${$}`]),children:g?r(we,{}):S(z,{children:[(x||y)&&S("div",{"data-testid":"cart-summary-list-heading-wrapper",className:I(["cart-cart-summary-list__heading",["cart-cart-summary-list__heading--full-width",!h]]),children:[x&&S(z,{children:[r(Y,{node:x,className:"cart-cart-summary-list__heading-text"}),r($t,{variant:"primary",className:I(["cart-cart-summary-list__heading-divider"])})]}),y&&r(Y,{node:y,className:"cart-cart-summary-list__out-of-stock-message"})]}),r("div",{className:I(["cart-cart-summary-list__content",["cart-cart-summary-list__content--empty",!h]]),children:h||r(Y,{node:B,className:"cart-cart-summary-list__empty-cart"})}),A&&S(z,{children:[r($t,{variant:"primary",className:I(["cart-cart-summary-list__footer-divider"])}),r(Y,{node:A,className:"cart-cart-summary-list__footer-text"})]})]})}),we=()=>r(se,{"data-testid":"cart-summary-list-skeleton",className:"cart-cart-summary-list__skeleton",rowGap:"medium",children:r(ue,{variant:"row",size:"xlarge",fullWidth:!0,lines:3,multilineGap:"small"})}),L=k=>new DOMParser().parseFromString(k,"text/html").body.textContent||"",_e=ie(({initialData:k=null,hideHeading:tt,hideFooter:x,routeProduct:A,routeEmptyCartCTA:B,routeCart:h,onItemUpdate:y,onItemRemove:$,maxItems:g,slots:a,attributesToHide:d=[],enableRemoveItem:b,enableUpdateItemQuantity:G,onItemsErrorsChange:et,accordion:Wt=!1,variant:nt="primary",isLoading:jt,showMaxItems:rt,showDiscount:at,showSavings:it,quantityType:ct,dropdownOptions:ot,undo:Q=!1,includeOutOfStockItems:W=!1,__testRecentlyRemovedItems:st,...ut},Dt)=>{var Et;const[q,Vt]=T(!k),[i,Jt]=T(k),[f,Kt]=T(new Set),[j,dt]=T(new Map),[N,P]=T(st||[]);K(()=>{},[st,N]);const l=(Et=he.config)==null?void 0:Et.shoppingCartDisplaySetting,[E,Yt]=T(rt?!0:!g&&!rt),c=Ie({file:"Cart.CartItem.file",files:"Cart.CartItem.files",heading:"Cart.Cart.heading",message:"Cart.CartItem.message",recipient:"Cart.CartItem.recipient",regularPrice:"Cart.CartItem.regularPrice",discountedPrice:"Cart.CartItem.discountedPrice",sender:"Cart.CartItem.sender",lowInventory:"Cart.CartItem.lowInventory",insufficientQuantity:"Cart.CartItem.insufficientQuantity",insufficientQuantityGeneral:"Cart.CartItem.insufficientQuantityGeneral",outOfStockHeading:"Cart.OutOfStockMessage.heading",outOfStockDescription:"Cart.OutOfStockMessage.message",outOfStockAlert:"Cart.OutOfStockMessage.alert",removeAction:"Cart.OutOfStockMessage.action",notAvailableMessage:"Cart.CartItem.notAvailableMessage",viewMore:"Cart.Cart.viewMore",viewAll:"Cart.Cart.viewAll",discountPercent:"Cart.CartItem.discountPercentage",savingsAmount:"Cart.CartItem.savingsAmount",itemBeingRemoved:"Cart.CartItem.itemBeingRemoved",itemRemoved:"Cart.CartItem.itemRemoved",itemRemovedDescription:"Cart.CartItem.itemRemovedDescription",undoAction:"Cart.CartItem.undoAction",dismissAction:"Cart.CartItem.dismissAction",remove:"Dropin.CartItem.remove.label",removeDefault:"Dropin.CartItem.removeDefault.label",quantity:"Dropin.CartItem.quantity.label"}),R=(t,e)=>{Kt(n=>(e?n.add(t):n.delete(t),new Set(n)))},D=(t,e)=>{dt(n=>(e?n.set(t,e):n.delete(t),new Map(n)))};function Zt(t){const e=zt(),n=e==null?void 0:e.items.find(o=>o.uid===t);n&&P(o=>{const s=e.items.findIndex(u=>u.uid===t);return s!==-1&&!o.some(u=>u.item.uid===t)?(qt.emit("cart/product/removed",{item:n,index:s}),[...o,{item:n,index:s,error:void 0,beingRemoved:!0}]):o})}const O=(t,e)=>{R(t.uid,!0),dt(new Map),b&&e===0?(Zt(t.uid),Ft([{uid:t.uid,quantity:e}]).then(()=>{$==null||$({item:t})}).finally(()=>{R(t.uid,!1),P(n=>n.map(o=>o.item.uid===t.uid?{...o,beingRemoved:!1}:o))}).catch(n=>{console.warn(n)})):G&&Ft([{uid:t.uid,quantity:e}]).then(()=>{y==null||y({item:t})}).finally(()=>{R(t.uid,!1)}).catch(n=>{D(t.uid,n.message)})},lt=t=>{P(e=>e.filter(n=>n.item.uid!==t))},Mt=Bt(()=>{Yt(t=>!t)},[]);K(()=>{const t=qt.on("cart/data",e=>{Jt(e),Vt(!!jt)},{eager:!0});return()=>{t==null||t.off()}},[]),K(()=>{Q||P([])},[Q]),K(()=>{et&&et(j)},[j,et]);const mt=(t,e)=>{if(d.includes("image"))return;const n={loading:e<4?"eager":"lazy",src:t.image.src,alt:t.image.alt,width:"300",height:"300",params:{width:300}},o=r(fe,{"data-testid":"cart-list-item-image",...n});return r(m,{name:"Thumbnail",slot:a==null?void 0:a.Thumbnail,context:{item:t,defaultImageProps:n},children:A?r("a",{href:A(t),children:o}):o})},gt=t=>{if(!d.includes("name"))return r("span",{"data-testid":"cart-list-item-title",children:A?r("a",{href:A(t),children:L(t.name)}):L(t.name)})},ft=t=>{if(d.includes("configurations"))return;const e={...t.bundleOptions,...t.selectedOptions,...t.customizableOptions,...t.recipient?{[c.recipient]:t.recipient}:null,...t.recipientEmail&&t.recipient?{[c.recipient]:`${t.recipient} (${t.recipientEmail})`}:null,...t.sender?{[c.sender]:t.sender}:null,...t.senderEmail&&t.sender?{[c.sender]:`${t.sender} (${t.senderEmail})`}:{},...t.message?{[c.message]:t.message}:null,...t.links&&t.links.count?t.links.count>1?{[c.files.replace("{count}",t.links.count.toString())]:t.links.result}:{[c.file.replace("{count}",t.links.count.toString())]:t.links.result}:null};if(Object.keys(e).length!==0)return e},ht=t=>{var e,n,o,s;return(l==null?void 0:l.price)==="INCLUDING_TAX"?t.discounted?{amount:t.regularPrice.value,currency:t.regularPrice.currency,style:{font:"inherit"},"data-testid":"including-tax-item-price"}:{amount:(e=t.taxedPrice)==null?void 0:e.value,currency:(n=t.taxedPrice)==null?void 0:n.currency,style:{font:"inherit"},"data-testid":"including-tax-item-price"}:{amount:(o=t.regularPrice)==null?void 0:o.value,currency:(s=t.regularPrice)==null?void 0:s.currency,style:{font:"inherit"},"data-testid":"regular-item-price"}},vt=t=>{var e,n;return{amount:(e=t.savingsAmount)==null?void 0:e.value,currency:(n=t.savingsAmount)==null?void 0:n.currency,style:{font:"inherit"},"data-testid":"item-savings-amount"}},yt=t=>(l==null?void 0:l.price)==="INCLUDING_EXCLUDING_TAX"?r(X,{amount:t.rowTotal.value,currency:t.rowTotal.currency,"data-testid":"excluding-tax-total","aria-label":c.regularPrice}):void 0,pt=t=>{var o,s,u,v,C,w,_,F,Rt,Tt;const e={"aria-label":c.regularPrice},n=t.discounted?{}:null;return["INCLUDING_TAX","INCLUDING_EXCLUDING_TAX"].includes(l==null?void 0:l.price)?(e.amount=(o=t.rowTotalIncludingTax)==null?void 0:o.value,e.currency=(s=t.rowTotalIncludingTax)==null?void 0:s.currency,e.variant=t.discounted?"strikethrough":"default",e["data-testid"]="including-tax-item-total",n&&(e.amount=(u=t.total)==null?void 0:u.value,e.currency=(v=t.total)==null?void 0:v.currency,n.amount=(C=t.rowTotalIncludingTax)==null?void 0:C.value,n.currency=(w=t.rowTotalIncludingTax)==null?void 0:w.currency,n.sale=!0,n["aria-label"]=c.discountedPrice,n["data-testid"]="discount-total")):(e.amount=(_=t.total)==null?void 0:_.value,e.currency=(F=t.total)==null?void 0:F.currency,e.variant=t.discounted?"strikethrough":"default",e["data-testid"]="regular-item-total",n&&(n.amount=(Rt=t.discountedTotal)==null?void 0:Rt.value,n.currency=(Tt=t.discountedTotal)==null?void 0:Tt.currency,n.sale=!0,n["aria-label"]=c.regularPrice,n["data-testid"]="discount-total")),{totalProps:e,discountProps:n}},Ct=t=>{var C,w,_;if(d.includes("warning"))return;const e=j.get(t.uid),n=(C=j.get(t.uid))==null?void 0:C.includes("The requested qty is not available"),o=f.has(t.uid),s=t.insufficientQuantity&&t.stockLevel?t.stockLevel==="noNumber"?c.insufficientQuantityGeneral:c.insufficientQuantity.replace("{inventory}",(w=t.stockLevel)==null?void 0:w.toString()).replace("{count}",t.quantity.toString()):"",u=t.lowInventory&&t.onlyXLeftInStock&&c.lowInventory.replace("{count}",(_=t.onlyXLeftInStock)==null?void 0:_.toString()),v=!t.outOfStock&&e&&n?c.notAvailableMessage:e;return!o&&(e||t.insufficientQuantity||t.lowInventory)?S("span",{"data-testid":"item-warning",children:[r(M,{source:ye,size:"16"}),v||s||u]}):void 0},It=t=>d!=null&&d.includes("alert")?void 0:!f.has(t.uid)&&t.outOfStock?S("span",{"data-testid":"item-alert",children:[r(M,{source:Ut,size:"16"}),c.outOfStockAlert]}):void 0,kt=t=>r(m,{name:"ProductAttributes",slot:a==null?void 0:a.ProductAttributes,context:{item:t}}),wt=t=>{if(!d.includes("sku"))return r("span",{"data-testid":"cart-list-item-sku",children:t.sku})},_t=t=>r(m,{name:"Footer",slot:a==null?void 0:a.Footer,context:{item:t,handleItemsLoading:R,handleItemsError:D,onItemUpdate:y}}),St=t=>r(m,{name:"RowTotalFooter",slot:a==null?void 0:a.RowTotalFooter,context:{item:t}}),Ht=t=>{var e;if(b)return r(m,{name:"ItemRemoveAction",slot:a==null?void 0:a.ItemRemoveAction,context:{item:t,enableRemoveItem:b,handleItemQuantityUpdate:O,handleItemsError:D,handleItemsLoading:R,onItemUpdate:y,itemsLoading:f},children:r(H,{"data-testid":"cart-item-remove-button",className:I(["dropin-cart-item__remove"]),variant:"tertiary",onClick:()=>O(t,0),disabled:f.has(t.uid),icon:r(M,{"data-testid":"cart-item-remove-icon",source:pe,size:"24",stroke:"2",viewBox:"0 0 24 24","aria-label":((e=c.remove)==null?void 0:e.replace("{product}",L(t.name)))||c.removeDefault})})})},te=t=>{if(!d.includes("quantity")&&G)return r(m,{name:"ItemQuantity",slot:a==null?void 0:a.ItemQuantity,context:{item:t,enableUpdateItemQuantity:G,handleItemQuantityUpdate:O,itemsLoading:f,handleItemsError:D,handleItemsLoading:R,onItemUpdate:y},children:r(ge,{value:t.quantity,min:1,onValue:e=>O(t,Number(e)),name:"quantity",disabled:f.has(t.uid),"aria-label":c.quantity})})},xt=t=>i!=null&&i.totalQuantity?i.items.filter(t).map((e,n)=>{var v;const{totalProps:o,discountProps:s}=pt(e),u=r(me,{updating:f==null?void 0:f.has(e.uid),loading:q,"data-testid":`cart-list-item-entry-${e.uid}`,image:mt(e,n),title:r(m,{name:"ItemTitle",slot:a==null?void 0:a.ItemTitle,context:{item:e},children:gt(e)}),sku:d.includes("sku")?void 0:r(m,{name:"ItemSku",slot:a==null?void 0:a.ItemSku,context:{item:e},children:wt(e)}),price:d.includes("price")?void 0:r(m,{name:"ItemPrice",slot:a==null?void 0:a.ItemPrice,context:{item:e},slotTag:"span",children:r(X,{...ht(e)})}),quantity:d.includes("quantity")?void 0:e.quantity,quantityContent:te(e),total:S(z,{children:[d.includes("total")?void 0:r(m,{name:"ItemTotal",slot:a==null?void 0:a.ItemTotal,context:{item:e},children:r(X,{...o})}),d.includes("totalDiscount")?void 0:s&&r(m,{name:"ItemTotal",slot:a==null?void 0:a.ItemTotal,context:{item:e},children:r(X,{...s})})]}),attributes:kt(e),configurations:ft(e),totalExcludingTax:d.includes("totalExcludingTax")?void 0:yt(e),taxIncluded:(l==null?void 0:l.price)==="INCLUDING_TAX",taxExcluded:!d.includes("totalExcludingTax")&&(l==null?void 0:l.price)==="INCLUDING_EXCLUDING_TAX",warning:Ct(e),alert:It(e),quantityType:ct,dropdownOptions:ot,onQuantity:G&&!f.has(e.uid)?C=>{O(e,C)}:void 0,onRemove:b&&!f.has(e.uid)?()=>O(e,0):void 0,discount:at&&e.discounted&&e.discountPercentage?r("div",{"data-testid":"item-discount-percent",children:c.discountPercent.replace("{discount}",((v=e.discountPercentage)==null?void 0:v.toString())??"")}):void 0,savings:it&&e.discounted&&e.savingsAmount?S("div",{children:[r("span",{children:r(X,{...vt(e)})})," ",c.savingsAmount]}):void 0,rowTotalFooter:St(e),footer:_t(e),removeContent:Ht(e)},e.uid);return r(m,{name:"CartItem",slot:a==null?void 0:a.CartItem,context:{item:e,index:n,enableUpdateItemQuantity:G,enableRemoveItem:b,itemsLoading:f,loading:q,attributesToHide:d,cartTaxesConfig:l,quantityType:ct,dropdownOptions:ot,showDiscount:at,showSavings:it,dictionary:c,getImage:mt,getTitle:gt,getSku:wt,getPriceProps:ht,getTotalPriceProps:pt,getProductAttributes:kt,getConfiguration:ft,getTotalExcludingTax:yt,getWarning:Ct,getAlert:It,getRowTotalFooter:St,getFooter:_t,getSavingsAmount:vt},children:u},e.uid)}):null,At=r(m,{name:"Heading",slot:a==null?void 0:a.Heading,context:{count:i==null?void 0:i.totalQuantity},children:r("div",{"data-testid":"default-cart-heading",children:c.heading.replace("({count})",i!=null&&i.totalQuantity?`(${i==null?void 0:i.totalQuantity.toString()})`:"")})}),ee=At.props.children.props.children,ne=()=>{const t=i==null?void 0:i.items.filter(e=>e.outOfStock);t==null||t.forEach(e=>{O(e,0)})},re=xt(t=>t.outOfStock||t.insufficientQuantity||!1),Nt=i!=null&&i.hasOutOfStockItems?r(Gt,{"data-testid":"cart-out-of-stock-message",icon:r(M,{source:Ut,size:"16"}),itemList:r(Z,{className:"cart-cart-summary-list__out-of-stock-items","data-testid":"out-of-stock-cart-items","aria-hidden":W||void 0,children:re}),type:"warning",heading:c.outOfStockHeading,description:c.outOfStockDescription,variant:"primary",actionButtonPosition:"bottom",additionalActions:i!=null&&i.hasFullyOutOfStockItems&&b?[{label:c.removeAction,onClick:ne}]:void 0}):void 0,p=xt(W?()=>!0:t=>!t.outOfStock&&!t.insufficientQuantity),Pt=E?Math.max(g||5,5):Math.min((i==null?void 0:i.totalQuantity)||5,5),Ot=(i==null?void 0:i.totalQuantity)>Pt,ae=Ot&&!E&&Pt!=g,V=i!=null&&i.totalQuantity&&p?r(m,{name:"Footer",slot:a==null?void 0:a.CartSummaryFooter,context:{displayMaxItems:E,routeCart:h},"data-testid":"cart-cart-summary-footer-slot",className:"cart-cart-summary-footer",children:Ot?ae?r(H,{className:"cart-cart-summary-list-footer__action",onClick:Mt,"data-testid":"view-more-items-button",variant:"tertiary",children:c.viewMore}):h&&r(H,{className:"cart-cart-summary-list-footer__action",href:h(),variant:"tertiary","data-testid":"view-cart-or-less-items-button",children:c.viewAll}):h&&r(H,{className:"cart-cart-summary-list-footer__action",href:h(),variant:"tertiary","data-testid":"view-cart-button",children:c.viewAll})}):null,Lt=()=>{if(!(i!=null&&i.items))return null;const t=[],e=i.items,n=a==null?void 0:a.UndoBanner,o=E?Math.max(g||(i==null?void 0:i.totalQuantity),5):Math.min(g??5,5);for(let s=0;s<=o;s++)Q&&N.filter(u=>u.index===s).forEach(({item:u,error:v,beingRemoved:C})=>{var _,F;const w=r(Gt,{type:v?"error":"info",heading:C?(_=c.itemBeingRemoved)==null?void 0:_.replace("{product}",L(u.name)):(F=c.itemRemoved)==null?void 0:F.replace("{product}",L(u.name)),description:v||c.itemRemovedDescription,variant:"primary",actionButtonPosition:"bottom",additionalActions:C?[]:[{label:c.undoAction,onClick:()=>J(u.uid),"aria-label":`${c.undoAction} remove ${L(u.name)}`},{label:c.dismissAction,onClick:()=>lt(u.uid),"aria-label":`${c.dismissAction} undo for ${L(u.name)}`}]},`undo-banner-${u.uid}`);t.push(n?r(m,{name:"UndoBanner",slot:n,context:{item:u,loading:q,error:v,onUndo:()=>J(u.uid),onDismiss:()=>lt(u.uid)},children:w},`undo-banner-${u.uid}`):w)}),s<o&&s<e.length&&t.push((p==null?void 0:p[s])||null);return r(Z,{"data-testid":"cart-list",children:t})},J=Bt(async t=>{P(n=>n.map(o=>o.item.uid===t?{...o,loading:!0,error:void 0}:o));const e=N.find(n=>n.item.uid===t);if(e)try{await ve([{sku:e.item.sku,parentSku:e.item.topLevelSku,quantity:e.item.quantity,optionsUIDs:e.item.bundleOptionsUIDs||(e.item.selectedOptionsUIDs?Object.values(e.item.selectedOptionsUIDs):void 0),enteredOptions:e.item.customizableOptions?Object.entries(e.item.customizableOptions).map(([n,o])=>({uid:n,value:o})):void 0}]),P(n=>n.filter(o=>o.item.uid!==t))}catch(n){P(o=>o.map(s=>s.item.uid===t?{...s,error:(n==null?void 0:n.message)||"Failed to restore item."}:s))}},[N]);ce(Dt,()=>({handleUndo:J}),[J]);const bt=r(z,{children:Q&&N.length>0?Lt():r(m,{name:"EmptyCart",slot:a==null?void 0:a.EmptyCart,context:{},children:r(oe,{"data-testid":"empty-cart",ctaLinkURL:B==null?void 0:B()})})}),Qt=i!=null&&i.totalQuantity||Q&&N.length>0?Q?N.length>0?Lt():r(Z,{"data-testid":"cart-list",children:p==null?void 0:p.slice(0,E?Math.max(g||(i==null?void 0:i.totalQuantity),5):Math.min(g??5,5))}):r(Z,{"data-testid":"cart-list",children:p==null?void 0:p.slice(0,E?Math.max(g||(i==null?void 0:i.totalQuantity),5):Math.min(g??5,5))}):null;return Wt?r(de,{"data-testid":"cart-summary-list-accordion",className:I(["cart-cart-summary-list-accordion",`cart-cart-summary-list__background--${nt}`]),iconOpen:ke,iconClose:Ce,children:r(le,{title:ee,"data-testid":"cart-summary-list-accordion__section",open:!0,renderContentWhenClosed:!0,children:r(Xt,{...ut,"aria-expanded":!0,"aria-label":"TEST",className:I(["cart-cart-summary-list-accordion__list",["cart-cart-summary-list--include-out-of-stock",W]]),loading:q,footer:x?void 0:V||(h?V:void 0),emptyCart:bt,products:Qt,outOfStockMessage:Nt,variant:nt})})}):r(Xt,{...ut,className:I([["cart-cart-summary-list--include-out-of-stock",W]]),heading:tt?void 0:At,footer:x?void 0:V||(h?V:void 0),loading:q,emptyCart:bt,products:Qt,outOfStockMessage:Nt,variant:nt})});_e.getInitialData=async function(){return zt()};export{_e as C};
 //# sourceMappingURL=CartSummaryList.js.map
