--- scripts/__dropins__/storefront-cart/chunks/CartSummaryTable.js
@@ -1,4 +1,0 @@
-/*! Copyright 2026 Adobe
-All Rights Reserved. */
-import{jsxs as _,jsx as n}from"@dropins/tools/preact-jsx-runtime.js";import{debounce as De,classes as ee,VComponent as he,Slot as v}from"@dropins/tools/lib.js";import{Button as F,Icon as H,Image as Me,Price as D,Skeleton as Ee,SkeletonRow as Ae}from"@dropins/tools/components.js";import{S as Le,f as ze,j as Ge,k as We,I as Be,l as Y}from"./components.js";import{useState as Ce,useCallback as Qe,useMemo as Ve}from"@dropins/tools/preact-compat.js";import{useState as Z,useCallback as Xe,useEffect as je,useMemo as Re}from"@dropins/tools/preact-hooks.js";import{events as ye}from"@dropins/tools/event-bus.js";import{s as Fe,updateProductsFromCart as Ie,addProductsToCart as He,getCartDataFromCache as _e}from"../api.js";import{E as Je}from"./EmptyCart.js";import{useText as ve}from"@dropins/tools/i18n.js";const Ke=({dictionary:r,onQuantityUpdate:c,onItemRemove:s})=>{var T,x;const[e,u]=Ce(new Map),[o,f]=Ce(new Map),h=(x=(T=Fe)==null?void 0:T.config)==null?void 0:x.shoppingCartDisplaySetting,S=(h==null?void 0:h.price)==="INCLUDING_TAX",y=(h==null?void 0:h.price)==="INCLUDING_EXCLUDING_TAX",w=({item:t})=>{const i={...t.bundleOptions,...t.selectedOptions,...t.customizableOptions,...t.recipient?{[r.recipient]:t.recipient}:null,...t.recipientEmail&&t.recipient?{[r.recipient]:`${t.recipient} (${t.recipientEmail})`}:null,...t.sender?{[r.sender]:t.sender}:null,...t.senderEmail&&t.sender?{[r.sender]:`${t.sender} (${t.senderEmail})`}:{},...t.message?{[r.message]:t.message}:null,...t.links&&t.links.count?t.links.count>1?{[r.files.replace("{count}",t.links.count.toString())]:t.links.result}:{[r.file.replace("{count}",t.links.count.toString())]:t.links.result}:null};if(Object.keys(i).length!==0)return i},U=t=>{var i,l,d,m;return S?t.discounted?{amount:t.regularPrice.value,currency:t.regularPrice.currency,"data-testid":"discounted-regular-item-price"}:{amount:(i=t.taxedPrice)==null?void 0:i.value,currency:(l=t.taxedPrice)==null?void 0:l.currency,"data-testid":"taxed-item-price"}:{amount:(d=t.regularPrice)==null?void 0:d.value,currency:(m=t.regularPrice)==null?void 0:m.currency,"data-testid":"regular-item-price"}},q=t=>{var d,m,I,M,E,A,V,X,j,L;const i={"aria-label":r.regularPrice},l=t.discounted?{}:null;return["INCLUDING_TAX","INCLUDING_EXCLUDING_TAX"].includes(h==null?void 0:h.price)?(i.amount=(d=t.rowTotalIncludingTax)==null?void 0:d.value,i.currency=(m=t.rowTotalIncludingTax)==null?void 0:m.currency,i.variant=t.discounted?"strikethrough":"default",i["data-testid"]="including-tax-item-total",l&&(i.amount=(I=t.total)==null?void 0:I.value,i.currency=(M=t.total)==null?void 0:M.currency,l.amount=(E=t.rowTotalIncludingTax)==null?void 0:E.value,l.currency=(A=t.rowTotalIncludingTax)==null?void 0:A.currency,l.sale=!0,l["aria-label"]=r.discountedPrice,l["data-testid"]="discount-total")):(i.amount=(V=t.total)==null?void 0:V.value,i.currency=(X=t.total)==null?void 0:X.currency,i.variant=t.discounted?"strikethrough":"default",i["data-testid"]="regular-item-total",l&&(l.amount=(j=t.discountedTotal)==null?void 0:j.value,l.currency=(L=t.discountedTotal)==null?void 0:L.currency,l.sale=!0,l["aria-label"]=r.regularPrice,l["data-testid"]="discount-total")),{subtotalProps:i,subtotalDiscountProps:l}},C=t=>{var m,I;const i=o.get(t.uid),l=t.insufficientQuantity&&t.stockLevel?t.stockLevel==="noNumber"?r.insufficientQuantityGeneral:r.insufficientQuantity.replace("{inventory}",(m=t.stockLevel)==null?void 0:m.toString()).replace("{count}",t.quantity.toString()):void 0,d=t.lowInventory&&t.onlyXLeftInStock&&r.lowInventory.replace("{count}",(I=t.onlyXLeftInStock)==null?void 0:I.toString());return i||l||d||void 0},P=Qe(async(t,i)=>{const l=(d,m)=>Ie([{uid:d.uid,quantity:m}]);if(!(isNaN(i)||t.quantity===i)){f(new Map);try{u(d=>{const m=new Map(d);return m.set(t.uid,{isUpdating:!0,updatedValue:i}),m}),await l(t,i),i===0?s==null||s(t):c==null||c(t,i)}catch(d){u(m=>{const I=new Map(m);return I.delete(t.uid),I}),f(m=>{const I=new Map(m);return I.set(t.uid,d.message),I})}}},[s,c]),N=Ve(()=>De(P,500),[P]);return{showIncludedTaxPrice:S,showExcludingTaxPrice:y,itemsUpdating:e,itemUpdateErrors:o,getConfiguration:w,getPriceProps:U,getSubtotalProps:q,processQuantityChange:P,debouncedQuantityChange:N,setItemsUpdating:u,getWarningMessage:C,setItemUpdateError:(t,i)=>{f(l=>{const d=new Map(l);return i?d.set(t,i):d.delete(t),d})},handleRemoveItem:t=>(u(i=>{const l=new Map(i);return l.set(t.uid,{isUpdating:!0,updatedValue:0}),l}),Ie([{uid:t.uid,quantity:0}]).then(i=>(s==null||s(t),Promise.resolve(i)))),setItemUpdating:(t,i)=>{u(l=>{const d=new Map(l);return d.set(t,{isUpdating:i,updatedValue:i}),d})}}},Ye=({variant:r="primary",className:c,type:s="warning",additionalActions:e,onDismiss:u,heading:o,description:f,icon:h,itemList:S,actionButtonPosition:y,...w})=>{var q,C,P;const U=ve({dismiss:"Dropin.InlineAlert.dismissLabel"});return _("div",{...w,className:ee(["dropin-in-line-alert",`dropin-in-line-alert--${s}`,`dropin-in-line-alert--${r}`,c]),children:[_("div",{className:"dropin-in-line-alert__heading",children:[_("div",{className:"dropin-in-line-alert__title-container",children:[h&&n(he,{node:h,className:"dropin-in-line-alert__icon"}),n("span",{className:"dropin-in-line-alert__title",children:o})]}),_("div",{className:"dropin-in-line-alert__actions-container",children:[e&&(y==="top"||!y&&e.length<=1)&&n(F,{variant:"tertiary",className:"dropin-in-line-alert__additional-action",onClick:e.length>0?(q=e[0])==null?void 0:q.onClick:void 0,"aria-label":(C=e[0])==null?void 0:C.label,children:(P=e[0])==null?void 0:P.label}),u&&n(F,{icon:n(H,{source:Le,size:"24",stroke:"2"}),className:"dropin-in-line-alert__dismiss-button",variant:"tertiary",onClick:u,"aria-label":U.dismiss})]})]}),f&&n("p",{className:"dropin-in-line-alert__description",children:f}),n("div",{className:"dropin-in-line-alert__item-list-container",children:S&&n(he,{node:S,className:ee(["dropin-in-line-alert__item-list"])})}),e&&(y==="bottom"||!y&&e.length>1)&&n("div",{className:"dropin-in-line-alert__additional-actions-container",children:e.map(N=>n(F,{variant:"tertiary",className:"dropin-in-line-alert__additional-action",onClick:N.onClick,children:N.label},N.label))})]})},Ze=()=>n(Ee,{"data-testid":"cart-summary-table-skeleton",className:"cart-cart-summary-table__skeleton",rowGap:"medium",children:n(Ae,{variant:"row",size:"xlarge",fullWidth:!0,lines:4,multilineGap:"small"})}),et=({initialData:r=null,routeProduct:c,routeEmptyCartCTA:s,slots:e,className:u,allowQuantityUpdates:o=!0,allowRemoveItems:f=!0,onQuantityUpdate:h,onItemRemove:S,undo:y=!1,...w})=>{const[U,q]=Z(!r),[C,P]=Z(r),[N,W]=Z([]),k=ve({file:"Cart.CartItem.file",files:"Cart.CartItem.files",heading:"Cart.Cart.heading",message:"Cart.CartItem.message",recipient:"Cart.CartItem.recipient",regularPrice:"Cart.CartItem.regularPrice",discountedPrice:"Cart.CartItem.discountedPrice",sender:"Cart.CartItem.sender",lowInventory:"Cart.CartItem.lowInventory",insufficientQuantity:"Cart.CartItem.insufficientQuantity",insufficientQuantityGeneral:"Cart.CartItem.insufficientQuantityGeneral",outOfStockHeading:"Cart.OutOfStockMessage.heading",outOfStockDescription:"Cart.OutOfStockMessage.message",outOfStockAlert:"Cart.OutOfStockMessage.alert",removeAction:"Cart.OutOfStockMessage.action",notAvailableMessage:"Cart.CartItem.notAvailableMessage",viewMore:"Cart.Cart.viewMore",viewAll:"Cart.Cart.viewAll",discountPercent:"Cart.CartItem.discountPercentage",savingsAmount:"Cart.CartItem.savingsAmount",includingTax:"Cart.CartItem.includingTax",excludingTax:"Cart.CartItem.excludingTax",remove:"Dropin.CartItem.remove.label",removeDefault:"Dropin.CartItem.removeDefault.label",quantity:"Dropin.CartItem.quantity.label",itemBeingRemoved:"Cart.CartItem.itemBeingRemoved",itemRemoved:"Cart.CartItem.itemRemoved",itemRemovedDescription:"Cart.CartItem.itemRemovedDescription",undoAction:"Cart.CartItem.undoAction",dismissAction:"Cart.CartItem.dismissAction"}),{showIncludedTaxPrice:B,showExcludingTaxPrice:Q,itemsUpdating:T,itemUpdateErrors:x,getConfiguration:t,getPriceProps:i,getSubtotalProps:l,debouncedQuantityChange:d,setItemsUpdating:m,getWarningMessage:I,handleRemoveItem:M,setItemUpdating:E,setItemUpdateError:A}=Ke({dictionary:k,onQuantityUpdate:h,onItemRemove:S});function V(b){const a=_e(),p=a==null?void 0:a.items.find(g=>g.uid===b);p&&W(g=>{const O=a.items.findIndex($=>$.uid===b);return O!==-1&&!g.some($=>$.item.uid===b)?(ye.emit("cart/product/removed",{item:p,index:O}),[...g,{item:p,index:O,loading:!1,error:void 0}]):g})}const X=tt(N,W),j=at(W),L=Xe(b=>{y&&V(b.uid),M(b)},[y,M]);je(()=>{const b=ye.on("cart/data",a=>{q(!0),P(a),m(new Map),q(!1)},{eager:!0});return()=>{b==null||b.off()}},[q,P,m]);const z=Re(()=>{var b;return((b=C==null?void 0:C.items)==null?void 0:b.sort((a,p)=>x.has(a.uid)&&!x.has(p.uid)?-1:!x.has(a.uid)&&x.has(p.uid)?1:a.outOfStock&&!p.outOfStock?-1:!a.outOfStock&&p.outOfStock?1:0).map((a,p)=>{var ce,ue,le,se,oe,de,me,pe,ge,fe,be;const g=T.has(a.uid),O=I(a),$=n(v,{name:"ProductTitle",slot:e==null?void 0:e.ProductTitle,context:{item:a},children:c?n("a",{href:c(a),children:a.name}):a.name}),R={src:a.image.src,alt:a.image.alt,width:"300",height:"300",params:{width:300}},te=n(Me,{"data-testid":`cart-table-item-image-${a.sku}`,loading:p<4?"eager":"lazy",...R}),we=n(v,{name:"Thumbnail",slot:e==null?void 0:e.Thumbnail,className:"cart-cart-summary-table__item-image-wrapper",context:{item:a,defaultImageProps:{...R,loading:p<4?"eager":"lazy"},index:p},children:c?n("a",{href:c(a),children:te}):te}),Ne=n(v,{name:"Sku",slot:e==null?void 0:e.Sku,context:{item:a},children:a.sku}),Te=a.outOfStock||e!=null&&e.ItemAlert?n(v,{name:"ItemAlert",slot:e==null?void 0:e.ItemAlert,context:{item:a},children:a.outOfStock&&_("div",{className:"cart-cart-summary-table__item-quantity-alert-wrapper","data-testid":`cart-table-item-quantity-alert-${a.uid}`,children:[n(H,{className:"cart-cart-summary-table__item-quantity-alert-icon",source:ze,size:"16"}),n("span",{className:"cart-cart-summary-table__item-quantity-alert-text",children:k.outOfStockAlert})]})}):void 0,Se=O||e!=null&&e.ItemWarning?n(v,{name:"ItemWarning",slot:e==null?void 0:e.ItemWarning,context:{item:a},children:O&&_("div",{className:"cart-cart-summary-table__item-quantity-warning-wrapper","data-testid":`cart-table-item-quantity-warning-${a.uid}`,children:[n(H,{className:"cart-cart-summary-table__item-quantity-warning-icon",source:Ge,size:"16"}),n("span",{className:"cart-cart-summary-table__item-quantity-warning-text",children:O})]})}):void 0,J=t({item:a}),Pe=J||e!=null&&e.Configurations?n(v,{name:"Configurations",slot:e==null?void 0:e.Configurations,context:{item:a},children:J&&n("ul",{className:"cart-cart-summary-table__item-configurations","data-testid":`cart-table-item-configurations-${a.sku}`,children:Object.entries(J).map(([G,K])=>_("li",{className:"cart-cart-summary-table__item-configuration","data-testid":`cart-table-item-configuration-${G}`,children:[_("span",{className:"cart-cart-summary-table__item-configuration-label",children:[G,":"]}),n("span",{className:"cart-cart-summary-table__item-configuration-value",children:K})]},G))})}):void 0,{subtotalProps:Oe,subtotalDiscountProps:ae}=l(a),ne=(ce=T.get(a.uid))==null?void 0:ce.updatedValue,re=g&&!isNaN(ne)?ne:a.quantity,ie=G=>{const K=G.target,$e=parseInt(K.value,10);d(a,$e)},Ue=T.has(a.uid)&&((ue=T.get(a.uid))==null?void 0:ue.updatedValue)===0,qe=n(v,{name:"Actions",slot:e==null?void 0:e.Actions,context:{item:a,itemsUpdating:T,setItemUpdating:E,setItemUpdateError:A},children:f&&!Ue&&n(F,{variant:"tertiary",size:"medium",className:"cart-cart-summary-table__item-remove-button","data-testid":`cart-table-item-remove-${a.uid}`,icon:n(H,{source:We,size:"32"}),onClick:()=>L(a),"aria-label":((le=k.remove)==null?void 0:le.replace("{product}",a.name))||k.removeDefault})});return{key:a.uid,uid:a.uid,updating:g,hasError:a.outOfStock||x.has(a.uid),item:n(v,{name:"Item",slot:e==null?void 0:e.Item,context:{item:a},children:n(Be,{productTitle:$,image:we,sku:Ne,configurations:Pe,alert:Te,warning:Se})}),price:_(v,{name:"Price",className:"cart-cart-summary-table__item-price",slot:e==null?void 0:e.Price,context:{item:a},children:[n(D,{...i(a)}),B&&_("span",{className:"cart-cart-summary-table__item-price-tax-label",children:[k.includingTax,": ",n(D,{amount:(se=a.taxedPrice)==null?void 0:se.value,currency:(oe=a.taxedPrice)==null?void 0:oe.currency})]}),Q&&_("span",{className:"cart-cart-summary-table__item-price-tax-label",children:[k.excludingTax,": ",n(D,{amount:(de=a.regularPrice)==null?void 0:de.value,currency:(me=a.regularPrice)==null?void 0:me.currency})]})]}),quantity:n(v,{name:"Quantity",className:"cart-cart-summary-table__item-qty",slot:e==null?void 0:e.Quantity,context:{item:a,isUpdating:g,quantityInputValue:re,handleInputChange:ie,itemUpdateErrors:x},children:o?n("input",{id:`cart-table-item-quantity-${a.uid}`,"data-testid":`cart-table-item-quantity-${a.uid}`,type:"number",min:"1",value:re,disabled:g,"aria-label":k.quantity,className:ee(["cart-cart-summary-table__cell-qty-input","cart-cart-summary-table__cell-qty-updater",["cart-cart-summary-table__cell-qty-updater--disabled",g],["cart-cart-summary-table__cell-qty-updater--error",x.has(a.uid)]]),onChange:ie}):n("span",{children:a.quantity})}),subtotal:_(v,{name:"Subtotal",className:"cart-cart-summary-table__item-subtotal",slot:e==null?void 0:e.Subtotal,context:{item:a},children:[n(D,{...Oe}),ae&&n(D,{...ae}),B&&_("span",{className:"cart-cart-summary-table__item-subtotal-tax-label",children:[k.includingTax,": ",n(D,{amount:(pe=a.rowTotalIncludingTax)==null?void 0:pe.value,currency:(ge=a.rowTotalIncludingTax)==null?void 0:ge.currency})]}),Q&&_("span",{className:"cart-cart-summary-table__item-subtotal-tax-label",children:[k.excludingTax,": ",n(D,{amount:(fe=a.rowTotal)==null?void 0:fe.value,currency:(be=a.rowTotal)==null?void 0:be.currency})]})]}),actions:qe}}))||[]},[C,T,o,c,e,x,k,t,f,d,I,i,l,E,A,B,Q,L]);if(U)return n(Ze,{});const ke=()=>{if(!(C!=null&&C.items))return null;const b=[],a=C.items;for(let p=0;p<=z.length;p++)y&&N.filter(g=>g.index===p).forEach(({item:g,error:O})=>{const $=nt(g.uid,T),R=ut({item:g,error:O},g,$,k,X,j,e);b.push({uid:`undo-banner-${g.uid}`,updating:!1,hasError:!1,item:n("div",{}),price:n("div",{}),quantity:n("div",{}),subtotal:n("div",{}),actions:n("div",{}),undoBanner:R})}),p<z.length&&p<a.length&&b.push(z[p]);return n(Y,{entries:b,className:u,...w})},xe=n(v,{name:"EmptyCart",slot:e==null?void 0:e.EmptyCart,context:{},children:n(Je,{ctaLinkURL:s==null?void 0:s()})});return C!=null&&C.totalQuantity||y&&N.length>0?y?N.length>0?ke():n(Y,{entries:z,className:u,...w}):n(Y,{entries:z,className:u,...w}):xe};et.getInitialData=async function(){return _e()};const tt=(r,c)=>async s=>{c(u=>u.map(o=>o.item.uid===s?{...o,error:void 0}:o));const e=r.find(u=>u.item.uid===s);if(e)try{await He([{sku:e.item.sku,parentSku:e.item.topLevelSku,quantity:e.item.quantity,optionsUIDs:e.item.bundleOptionsUIDs||(e.item.selectedOptionsUIDs?Object.values(e.item.selectedOptionsUIDs):void 0),enteredOptions:e.item.customizableOptions?Object.entries(e.item.customizableOptions).map(([u,o])=>({uid:u,value:o})):void 0}]),c(u=>u.filter(o=>o.item.uid!==s))}catch(u){c(o=>o.map(f=>f.item.uid===s?{...f,error:(u==null?void 0:u.message)||"Failed to restore item."}:f))}},at=r=>c=>{r(s=>s.filter(e=>e.item.uid!==c))},nt=(r,c)=>{var s;return c.has(r)&&((s=c.get(r))==null?void 0:s.updatedValue)===0},rt=(r,c,s,e,u,o)=>r?[]:[{label:c.undoAction,onClick:()=>{s(u)},"aria-label":`${c.undoAction} remove ${o}`},{label:c.dismissAction,onClick:()=>{e(u)},"aria-label":`${c.dismissAction} undo for ${o}`}],it=(r,c)=>()=>r(c),ct=(r,c)=>()=>r(c),ut=(r,c,s,e,u,o,f)=>{var w,U;const h=rt(s,e,u,o,r.item.uid,r.item.name),S=it(u,r.item.uid),y=ct(o,r.item.uid);return n(v,{name:"UndoBanner",slot:f==null?void 0:f.UndoBanner,context:{item:r.item,loading:s,error:r.error,onUndo:S,onDismiss:y},children:n(Ye,{type:r.error?"error":"info",heading:s?(w=e.itemBeingRemoved)==null?void 0:w.replace("{product}",c.name):(U=e.itemRemoved)==null?void 0:U.replace("{product}",r.item.name),description:r.error||e.itemRemovedDescription,variant:"primary",actionButtonPosition:"bottom",additionalActions:h},`undo-banner-${r.item.uid}`)},`undo-banner-${r.item.uid}`)};export{et as C,at as a,rt as b,tt as c,it as d,ct as e,ut as f,nt as g};
-//# sourceMappingURL=CartSummaryTable.js.map
