+++ scripts/__dropins__/storefront-cart/chunks/CartSummaryTable2.js
@@ -1,0 +1,4 @@
+/*! Copyright 2026 Adobe
+All Rights Reserved. */
+import{jsxs as m,jsx as a}from"@dropins/tools/preact-jsx-runtime.js";import{classes as E,VComponent as I,debounce as $e,Slot as N}from"@dropins/tools/lib.js";import{Button as J,Icon as K,Image as De,Price as D,Skeleton as Ee,SkeletonRow as Me}from"@dropins/tools/components.js";/* empty css                */import{useState as fe,useCallback as Ae,useMemo as Qe}from"@dropins/tools/preact-compat.js";import{useState as ee,useCallback as ze,useEffect as Be,useMemo as Ge}from"@dropins/tools/preact-hooks.js";import{useText as te}from"@dropins/tools/i18n.js";import{events as Ce}from"@dropins/tools/event-bus.js";import{s as We}from"./resetCart.js";import{u as ve}from"./updateProductsFromCart.js";import{g as Ie}from"./persisted-data.js";import{a as Ve}from"./createGuestCart.js";import{E as Xe}from"./EmptyCart.js";import{S as je,a as Re,b as He}from"./WarningWithCircle.js";import{S as Fe}from"./Close.js";const ae=({entries:n,className:l,...s})=>{const e=te({itemLabel:"Cart.CartSummaryTable.item",priceLabel:"Cart.CartSummaryTable.price",qtyLabel:"Cart.CartSummaryTable.qty",subtotalLabel:"Cart.CartSummaryTable.subtotal",mobilePriceLabel:"Cart.CartSummaryTable.mobilePrice",mobileQtyLabel:"Cart.CartSummaryTable.mobileQty",mobileSubtotalLabel:"Cart.CartSummaryTable.mobileSubtotal"});return m("div",{...s,className:E(["cart-cart-summary-table",l]),children:[m("div",{className:"cart-cart-summary-table__header",children:[a("div",{className:"cart-cart-summary-table__header-item",children:e.itemLabel}),a("div",{className:"cart-cart-summary-table__header-price",children:e.priceLabel}),a("div",{className:"cart-cart-summary-table__header-qty",children:e.qtyLabel}),a("div",{className:"cart-cart-summary-table__header-subtotal",children:e.subtotalLabel})]}),a("div",{className:"cart-cart-summary-table__body",children:n.map(i=>i.undoBanner?a("div",{className:"cart-cart-summary-table__undo-banner",children:a(I,{node:i.undoBanner})},i.uid):m("div",{className:E(["cart-cart-summary-table__row",["cart-cart-summary-table__row--updating",i.updating],["cart-cart-summary-table__row--error",i.hasError]]),children:[a("div",{className:"cart-cart-summary-table__cell-item",children:a(I,{node:i.item})}),m("div",{className:"cart-cart-summary-table__cell-price",children:[a("span",{className:"cart-cart-summary-table__mobile-label",children:e.mobilePriceLabel}),a(I,{node:i.price})]}),m("div",{className:"cart-cart-summary-table__cell-qty",children:[a("span",{className:"cart-cart-summary-table__mobile-label",children:e.mobileQtyLabel}),a(I,{node:i.quantity})]}),m("div",{className:"cart-cart-summary-table__cell-subtotal",children:[a("span",{className:"cart-cart-summary-table__mobile-label",children:e.mobileSubtotalLabel}),a(I,{node:i.subtotal})]}),a("div",{className:"cart-cart-summary-table__item-footer",children:a(I,{className:"cart-cart-summary-table__item-actions",node:i.actions})})]},i.uid))})]})},Je=({className:n,productTitle:l,sku:s,image:e,configurations:i,alert:o,warning:b,...y})=>m("div",{...y,className:E(["cart-cart-summary-table__item",n]),children:[e&&a(I,{className:"cart-cart-summary-table__item-image-wrapper",node:e}),m("div",{className:"cart-cart-summary-table__item-details",children:[l&&a(I,{className:"cart-cart-summary-table__item-name",node:l}),s&&a(I,{className:"cart-cart-summary-table__sku",node:s}),o&&a(I,{className:"cart-cart-summary-table__item-quantity-alert-text",node:o}),b&&a(I,{className:"cart-cart-summary-table__item-quantity-warning-text",node:b}),i&&a(I,{className:"cart-cart-summary-table__item-configurations",node:i})]})]}),Ke=({dictionary:n,onQuantityUpdate:l,onItemRemove:s})=>{var T,k;const[e,i]=fe(new Map),[o,b]=fe(new Map),y=(k=(T=We)==null?void 0:T.config)==null?void 0:k.shoppingCartDisplaySetting,P=(y==null?void 0:y.price)==="INCLUDING_TAX",C=(y==null?void 0:y.price)==="INCLUDING_EXCLUDING_TAX",S=({item:t})=>{const c={...t.bundleOptions,...t.selectedOptions,...t.customizableOptions,...t.recipient?{[n.recipient]:t.recipient}:null,...t.recipientEmail&&t.recipient?{[n.recipient]:`${t.recipient} (${t.recipientEmail})`}:null,...t.sender?{[n.sender]:t.sender}:null,...t.senderEmail&&t.sender?{[n.sender]:`${t.sender} (${t.senderEmail})`}:{},...t.message?{[n.message]:t.message}:null,...t.links&&t.links.count?t.links.count>1?{[n.files.replace("{count}",t.links.count.toString())]:t.links.result}:{[n.file.replace("{count}",t.links.count.toString())]:t.links.result}:null};if(Object.keys(c).length!==0)return c},L=t=>{var c,u,d,p;return P?t.discounted?{amount:t.regularPrice.value,currency:t.regularPrice.currency,"data-testid":"discounted-regular-item-price"}:{amount:(c=t.taxedPrice)==null?void 0:c.value,currency:(u=t.taxedPrice)==null?void 0:u.currency,"data-testid":"taxed-item-price"}:{amount:(d=t.regularPrice)==null?void 0:d.value,currency:(p=t.regularPrice)==null?void 0:p.currency,"data-testid":"regular-item-price"}},U=t=>{var d,p,v,M,A,Q,j,R,H,z;const c={"aria-label":n.regularPrice},u=t.discounted?{}:null;return["INCLUDING_TAX","INCLUDING_EXCLUDING_TAX"].includes(y==null?void 0:y.price)?(c.amount=(d=t.rowTotalIncludingTax)==null?void 0:d.value,c.currency=(p=t.rowTotalIncludingTax)==null?void 0:p.currency,c.variant=t.discounted?"strikethrough":"default",c["data-testid"]="including-tax-item-total",u&&(c.amount=(v=t.total)==null?void 0:v.value,c.currency=(M=t.total)==null?void 0:M.currency,u.amount=(A=t.rowTotalIncludingTax)==null?void 0:A.value,u.currency=(Q=t.rowTotalIncludingTax)==null?void 0:Q.currency,u.sale=!0,u["aria-label"]=n.discountedPrice,u["data-testid"]="discount-total")):(c.amount=(j=t.total)==null?void 0:j.value,c.currency=(R=t.total)==null?void 0:R.currency,c.variant=t.discounted?"strikethrough":"default",c["data-testid"]="regular-item-total",u&&(u.amount=(H=t.discountedTotal)==null?void 0:H.value,u.currency=(z=t.discountedTotal)==null?void 0:z.currency,u.sale=!0,u["aria-label"]=n.regularPrice,u["data-testid"]="discount-total")),{subtotalProps:c,subtotalDiscountProps:u}},f=t=>{var p,v;const c=o.get(t.uid),u=t.insufficientQuantity&&t.stockLevel?t.stockLevel==="noNumber"?n.insufficientQuantityGeneral:n.insufficientQuantity.replace("{inventory}",(p=t.stockLevel)==null?void 0:p.toString()).replace("{count}",t.quantity.toString()):void 0,d=t.lowInventory&&t.onlyXLeftInStock&&n.lowInventory.replace("{count}",(v=t.onlyXLeftInStock)==null?void 0:v.toString());return c||u||d||void 0},q=Ae(async(t,c)=>{const u=(d,p)=>ve([{uid:d.uid,quantity:p}]);if(!(isNaN(c)||t.quantity===c)){b(new Map);try{i(d=>{const p=new Map(d);return p.set(t.uid,{isUpdating:!0,updatedValue:c}),p}),await u(t,c),c===0?s==null||s(t):l==null||l(t,c)}catch(d){i(p=>{const v=new Map(p);return v.delete(t.uid),v}),b(p=>{const v=new Map(p);return v.set(t.uid,d.message),v})}}},[s,l]),w=Qe(()=>$e(q,500),[q]);return{showIncludedTaxPrice:P,showExcludingTaxPrice:C,itemsUpdating:e,itemUpdateErrors:o,getConfiguration:S,getPriceProps:L,getSubtotalProps:U,processQuantityChange:q,debouncedQuantityChange:w,setItemsUpdating:i,getWarningMessage:f,setItemUpdateError:(t,c)=>{b(u=>{const d=new Map(u);return c?d.set(t,c):d.delete(t),d})},handleRemoveItem:t=>(i(c=>{const u=new Map(c);return u.set(t.uid,{isUpdating:!0,updatedValue:0}),u}),ve([{uid:t.uid,quantity:0}]).then(c=>(s==null||s(t),Promise.resolve(c)))),setItemUpdating:(t,c)=>{i(u=>{const d=new Map(u);return d.set(t,{isUpdating:c,updatedValue:c}),d})}}},Ye=({variant:n="primary",className:l,type:s="warning",additionalActions:e,onDismiss:i,heading:o,description:b,icon:y,itemList:P,actionButtonPosition:C,...S})=>{var U,f,q;const L=te({dismiss:"Dropin.InlineAlert.dismissLabel"});return m("div",{...S,className:E(["dropin-in-line-alert",`dropin-in-line-alert--${s}`,`dropin-in-line-alert--${n}`,l]),children:[m("div",{className:"dropin-in-line-alert__heading",children:[m("div",{className:"dropin-in-line-alert__title-container",children:[y&&a(I,{node:y,className:"dropin-in-line-alert__icon"}),a("span",{className:"dropin-in-line-alert__title",children:o})]}),m("div",{className:"dropin-in-line-alert__actions-container",children:[e&&(C==="top"||!C&&e.length<=1)&&a(J,{variant:"tertiary",className:"dropin-in-line-alert__additional-action",onClick:e.length>0?(U=e[0])==null?void 0:U.onClick:void 0,"aria-label":(f=e[0])==null?void 0:f.label,children:(q=e[0])==null?void 0:q.label}),i&&a(J,{icon:a(K,{source:Fe,size:"24",stroke:"2"}),className:"dropin-in-line-alert__dismiss-button",variant:"tertiary",onClick:i,"aria-label":L.dismiss})]})]}),b&&a("p",{className:"dropin-in-line-alert__description",children:b}),a("div",{className:"dropin-in-line-alert__item-list-container",children:P&&a(I,{node:P,className:E(["dropin-in-line-alert__item-list"])})}),e&&(C==="bottom"||!C&&e.length>1)&&a("div",{className:"dropin-in-line-alert__additional-actions-container",children:e.map(w=>a(J,{variant:"tertiary",className:"dropin-in-line-alert__additional-action",onClick:w.onClick,children:w.label},w.label))})]})},Ze=()=>a(Ee,{"data-testid":"cart-summary-table-skeleton",className:"cart-cart-summary-table__skeleton",rowGap:"medium",children:a(Me,{variant:"row",size:"xlarge",fullWidth:!0,lines:4,multilineGap:"small"})}),ea=({initialData:n=null,routeProduct:l,routeEmptyCartCTA:s,slots:e,className:i,allowQuantityUpdates:o=!0,allowRemoveItems:b=!0,onQuantityUpdate:y,onItemRemove:P,undo:C=!1,...S})=>{const[L,U]=ee(!n),[f,q]=ee(n),[w,W]=ee([]),x=te({file:"Cart.CartItem.file",files:"Cart.CartItem.files",heading:"Cart.Cart.heading",message:"Cart.CartItem.message",recipient:"Cart.CartItem.recipient",regularPrice:"Cart.CartItem.regularPrice",discountedPrice:"Cart.CartItem.discountedPrice",sender:"Cart.CartItem.sender",lowInventory:"Cart.CartItem.lowInventory",insufficientQuantity:"Cart.CartItem.insufficientQuantity",insufficientQuantityGeneral:"Cart.CartItem.insufficientQuantityGeneral",outOfStockHeading:"Cart.OutOfStockMessage.heading",outOfStockDescription:"Cart.OutOfStockMessage.message",outOfStockAlert:"Cart.OutOfStockMessage.alert",removeAction:"Cart.OutOfStockMessage.action",notAvailableMessage:"Cart.CartItem.notAvailableMessage",viewMore:"Cart.Cart.viewMore",viewAll:"Cart.Cart.viewAll",discountPercent:"Cart.CartItem.discountPercentage",savingsAmount:"Cart.CartItem.savingsAmount",includingTax:"Cart.CartItem.includingTax",excludingTax:"Cart.CartItem.excludingTax",remove:"Dropin.CartItem.remove.label",removeDefault:"Dropin.CartItem.removeDefault.label",quantity:"Dropin.CartItem.quantity.label",itemBeingRemoved:"Cart.CartItem.itemBeingRemoved",itemRemoved:"Cart.CartItem.itemRemoved",itemRemovedDescription:"Cart.CartItem.itemRemovedDescription",undoAction:"Cart.CartItem.undoAction",dismissAction:"Cart.CartItem.dismissAction"}),{showIncludedTaxPrice:V,showExcludingTaxPrice:X,itemsUpdating:T,itemUpdateErrors:k,getConfiguration:t,getPriceProps:c,getSubtotalProps:u,debouncedQuantityChange:d,setItemsUpdating:p,getWarningMessage:v,handleRemoveItem:M,setItemUpdating:A,setItemUpdateError:Q}=Ke({dictionary:x,onQuantityUpdate:y,onItemRemove:P});function j(h){const r=Ie(),g=r==null?void 0:r.items.find(_=>_.uid===h);g&&W(_=>{const O=r.items.findIndex($=>$.uid===h);return O!==-1&&!_.some($=>$.item.uid===h)?(Ce.emit("cart/product/removed",{item:g,index:O}),[..._,{item:g,index:O,loading:!1,error:void 0}]):_})}const R=aa(w,W),H=ta(W),z=ze(h=>{C&&j(h.uid),M(h)},[C,M]);Be(()=>{const h=Ce.on("cart/data",r=>{U(!0),q(r),p(new Map),U(!1)},{eager:!0});return()=>{h==null||h.off()}},[U,q,p]);const B=Ge(()=>{var h;return((h=f==null?void 0:f.items)==null?void 0:h.sort((r,g)=>k.has(r.uid)&&!k.has(g.uid)?-1:!k.has(r.uid)&&k.has(g.uid)?1:r.outOfStock&&!g.outOfStock?-1:!r.outOfStock&&g.outOfStock?1:0).map((r,g)=>{var se,ue,oe,de,me,be,pe,ge,ye,_e,he;const _=T.has(r.uid),O=v(r),$=a(N,{name:"ProductTitle",slot:e==null?void 0:e.ProductTitle,context:{item:r},children:l?a("a",{href:l(r),children:r.name}):r.name}),F={src:r.image.src,alt:r.image.alt,width:"300",height:"300",params:{width:300}},re=a(De,{"data-testid":`cart-table-item-image-${r.sku}`,loading:g<4?"eager":"lazy",...F}),ke=a(N,{name:"Thumbnail",slot:e==null?void 0:e.Thumbnail,className:"cart-cart-summary-table__item-image-wrapper",context:{item:r,defaultImageProps:{...F,loading:g<4?"eager":"lazy"},index:g},children:l?a("a",{href:l(r),children:re}):re}),Se=a(N,{name:"Sku",slot:e==null?void 0:e.Sku,context:{item:r},children:r.sku}),we=r.outOfStock||e!=null&&e.ItemAlert?a(N,{name:"ItemAlert",slot:e==null?void 0:e.ItemAlert,context:{item:r},children:r.outOfStock&&m("div",{className:"cart-cart-summary-table__item-quantity-alert-wrapper","data-testid":`cart-table-item-quantity-alert-${r.uid}`,children:[a(K,{className:"cart-cart-summary-table__item-quantity-alert-icon",source:je,size:"16"}),a("span",{className:"cart-cart-summary-table__item-quantity-alert-text",children:x.outOfStockAlert})]})}):void 0,Te=O||e!=null&&e.ItemWarning?a(N,{name:"ItemWarning",slot:e==null?void 0:e.ItemWarning,context:{item:r},children:O&&m("div",{className:"cart-cart-summary-table__item-quantity-warning-wrapper","data-testid":`cart-table-item-quantity-warning-${r.uid}`,children:[a(K,{className:"cart-cart-summary-table__item-quantity-warning-icon",source:Re,size:"16"}),a("span",{className:"cart-cart-summary-table__item-quantity-warning-text",children:O})]})}):void 0,Y=t({item:r}),Pe=Y||e!=null&&e.Configurations?a(N,{name:"Configurations",slot:e==null?void 0:e.Configurations,context:{item:r},children:Y&&a("ul",{className:"cart-cart-summary-table__item-configurations","data-testid":`cart-table-item-configurations-${r.sku}`,children:Object.entries(Y).map(([G,Z])=>m("li",{className:"cart-cart-summary-table__item-configuration","data-testid":`cart-table-item-configuration-${G}`,children:[m("span",{className:"cart-cart-summary-table__item-configuration-label",children:[G,":"]}),a("span",{className:"cart-cart-summary-table__item-configuration-value",children:Z})]},G))})}):void 0,{subtotalProps:qe,subtotalDiscountProps:ne}=u(r),ie=(se=T.get(r.uid))==null?void 0:se.updatedValue,ce=_&&!isNaN(ie)?ie:r.quantity,le=G=>{const Z=G.target,Ue=parseInt(Z.value,10);d(r,Ue)},Oe=T.has(r.uid)&&((ue=T.get(r.uid))==null?void 0:ue.updatedValue)===0,Le=a(N,{name:"Actions",slot:e==null?void 0:e.Actions,context:{item:r,itemsUpdating:T,setItemUpdating:A,setItemUpdateError:Q},children:b&&!Oe&&a(J,{variant:"tertiary",size:"medium",className:"cart-cart-summary-table__item-remove-button","data-testid":`cart-table-item-remove-${r.uid}`,icon:a(K,{source:He,size:"32"}),onClick:()=>z(r),"aria-label":((oe=x.remove)==null?void 0:oe.replace("{product}",r.name))||x.removeDefault})});return{key:r.uid,uid:r.uid,updating:_,hasError:r.outOfStock||k.has(r.uid),item:a(N,{name:"Item",slot:e==null?void 0:e.Item,context:{item:r},children:a(Je,{productTitle:$,image:ke,sku:Se,configurations:Pe,alert:we,warning:Te})}),price:m(N,{name:"Price",className:"cart-cart-summary-table__item-price",slot:e==null?void 0:e.Price,context:{item:r},children:[a(D,{...c(r)}),V&&m("span",{className:"cart-cart-summary-table__item-price-tax-label",children:[x.includingTax,": ",a(D,{amount:(de=r.taxedPrice)==null?void 0:de.value,currency:(me=r.taxedPrice)==null?void 0:me.currency})]}),X&&m("span",{className:"cart-cart-summary-table__item-price-tax-label",children:[x.excludingTax,": ",a(D,{amount:(be=r.regularPrice)==null?void 0:be.value,currency:(pe=r.regularPrice)==null?void 0:pe.currency})]})]}),quantity:a(N,{name:"Quantity",className:"cart-cart-summary-table__item-qty",slot:e==null?void 0:e.Quantity,context:{item:r,isUpdating:_,quantityInputValue:ce,handleInputChange:le,itemUpdateErrors:k},children:o?a("input",{id:`cart-table-item-quantity-${r.uid}`,"data-testid":`cart-table-item-quantity-${r.uid}`,type:"number",min:"1",value:ce,disabled:_,"aria-label":x.quantity,className:E(["cart-cart-summary-table__cell-qty-input","cart-cart-summary-table__cell-qty-updater",["cart-cart-summary-table__cell-qty-updater--disabled",_],["cart-cart-summary-table__cell-qty-updater--error",k.has(r.uid)]]),onChange:le}):a("span",{children:r.quantity})}),subtotal:m(N,{name:"Subtotal",className:"cart-cart-summary-table__item-subtotal",slot:e==null?void 0:e.Subtotal,context:{item:r},children:[a(D,{...qe}),ne&&a(D,{...ne}),V&&m("span",{className:"cart-cart-summary-table__item-subtotal-tax-label",children:[x.includingTax,": ",a(D,{amount:(ge=r.rowTotalIncludingTax)==null?void 0:ge.value,currency:(ye=r.rowTotalIncludingTax)==null?void 0:ye.currency})]}),X&&m("span",{className:"cart-cart-summary-table__item-subtotal-tax-label",children:[x.excludingTax,": ",a(D,{amount:(_e=r.rowTotal)==null?void 0:_e.value,currency:(he=r.rowTotal)==null?void 0:he.currency})]})]}),actions:Le}}))||[]},[f,T,o,l,e,k,x,t,b,d,v,c,u,A,Q,V,X,z]);if(L)return a(Ze,{});const Ne=()=>{if(!(f!=null&&f.items))return null;const h=[],r=f.items;for(let g=0;g<=B.length;g++)C&&w.filter(_=>_.index===g).forEach(({item:_,error:O})=>{const $=ra(_.uid,T),F=la({item:_,error:O},_,$,x,R,H,e);h.push({uid:`undo-banner-${_.uid}`,updating:!1,hasError:!1,item:a("div",{}),price:a("div",{}),quantity:a("div",{}),subtotal:a("div",{}),actions:a("div",{}),undoBanner:F})}),g<B.length&&g<r.length&&h.push(B[g]);return a(ae,{entries:h,className:i,...S})},xe=a(N,{name:"EmptyCart",slot:e==null?void 0:e.EmptyCart,context:{},children:a(Xe,{ctaLinkURL:s==null?void 0:s()})});return f!=null&&f.totalQuantity||C&&w.length>0?C?w.length>0?Ne():a(ae,{entries:B,className:i,...S}):a(ae,{entries:B,className:i,...S}):xe};ea.getInitialData=async function(){return Ie()};const aa=(n,l)=>async s=>{l(i=>i.map(o=>o.item.uid===s?{...o,error:void 0}:o));const e=n.find(i=>i.item.uid===s);if(e)try{await Ve([{sku:e.item.sku,parentSku:e.item.topLevelSku,quantity:e.item.quantity,optionsUIDs:e.item.bundleOptionsUIDs||(e.item.selectedOptionsUIDs?Object.values(e.item.selectedOptionsUIDs):void 0),enteredOptions:e.item.customizableOptions?Object.entries(e.item.customizableOptions).map(([i,o])=>({uid:i,value:o})):void 0}]),l(i=>i.filter(o=>o.item.uid!==s))}catch(i){l(o=>o.map(b=>b.item.uid===s?{...b,error:(i==null?void 0:i.message)||"Failed to restore item."}:b))}},ta=n=>l=>{n(s=>s.filter(e=>e.item.uid!==l))},ra=(n,l)=>{var s;return l.has(n)&&((s=l.get(n))==null?void 0:s.updatedValue)===0},na=(n,l,s,e,i,o)=>n?[]:[{label:l.undoAction,onClick:()=>{s(i)},"aria-label":`${l.undoAction} remove ${o}`},{label:l.dismissAction,onClick:()=>{e(i)},"aria-label":`${l.dismissAction} undo for ${o}`}],ia=(n,l)=>()=>n(l),ca=(n,l)=>()=>n(l),la=(n,l,s,e,i,o,b)=>{var S,L;const y=na(s,e,i,o,n.item.uid,n.item.name),P=ia(i,n.item.uid),C=ca(o,n.item.uid);return a(N,{name:"UndoBanner",slot:b==null?void 0:b.UndoBanner,context:{item:n.item,loading:s,error:n.error,onUndo:P,onDismiss:C},children:a(Ye,{type:n.error?"error":"info",heading:s?(S=e.itemBeingRemoved)==null?void 0:S.replace("{product}",l.name):(L=e.itemRemoved)==null?void 0:L.replace("{product}",n.item.name),description:n.error||e.itemRemovedDescription,variant:"primary",actionButtonPosition:"bottom",additionalActions:y},`undo-banner-${n.item.uid}`)},`undo-banner-${n.item.uid}`)};export{ea as C,ta as a,na as b,aa as c,ia as d,ca as e,la as f,ra as g};
+//# sourceMappingURL=CartSummaryTable2.js.map
