+++ scripts/__dropins__/storefront-cart/chunks/Coupons.js
@@ -1,0 +1,4 @@
+/*! Copyright 2026 Adobe
+All Rights Reserved. */
+import{jsx as t,jsxs as u}from"@dropins/tools/preact-jsx-runtime.js";import*as o from"@dropins/tools/preact-compat.js";import{useRef as k}from"@dropins/tools/preact-compat.js";import{classes as e,VComponent as r,getFormValues as x}from"@dropins/tools/lib.js";import{Accordion as L,AccordionSection as N}from"@dropins/tools/components.js";/* empty css                */import{S}from"./Coupon.js";import{useText as C}from"@dropins/tools/i18n.js";const I=n=>o.createElement("svg",{id:"Icon_Add_Base","data-name":"Icon \\u2013 Add \\u2013 Base",xmlns:"http://www.w3.org/2000/svg",width:24,height:24,viewBox:"0 0 24 24",...n},o.createElement("g",{id:"Large"},o.createElement("rect",{id:"Placement_area","data-name":"Placement area",width:24,height:24,fill:"#fff",opacity:0}),o.createElement("g",{id:"Add_icon","data-name":"Add icon",transform:"translate(9.734 9.737)"},o.createElement("line",{vectorEffect:"non-scaling-stroke",id:"Line_579","data-name":"Line 579",y2:12.7,transform:"translate(2.216 -4.087)",fill:"none",stroke:"currentColor"}),o.createElement("line",{vectorEffect:"non-scaling-stroke",id:"Line_580","data-name":"Line 580",x2:12.7,transform:"translate(-4.079 2.263)",fill:"none",stroke:"currentColor"})))),P=n=>o.createElement("svg",{width:24,height:24,viewBox:"0 0 24 24",fill:"none",xmlns:"http://www.w3.org/2000/svg",...n},o.createElement("path",{d:"M17.3332 11.75H6.6665",strokeWidth:1,strokeLinecap:"square",strokeLinejoin:"round",vectorEffect:"non-scaling-stroke",fill:"none",stroke:"currentColor"})),V=({accordionSectionTitle:n,accordionSectionIcon:h,className:p,children:b,couponCodeField:i,applyCouponsButton:s,appliedCoupons:d,error:m,onApplyCoupon:a,...g})=>{const c=k(null),_=C({couponTitle:"Cart.PriceSummary.coupon.title"}),v=w=>{var f;w.preventDefault();const E=x(c.current);a==null||a(E);const l=(f=c==null?void 0:c.current)==null?void 0:f.querySelector("input");l&&(l.value="")};return t("div",{...g,"data-testid":"cart-coupons",className:e(["cart-coupons",p]),children:t(L,{"data-testid":"coupon-code",className:e(["cart-coupons__accordion"]),actionIconPosition:"right",iconOpen:I,iconClose:P,children:u(N,{title:n??_.couponTitle,iconLeft:h??S,showIconLeft:!0,renderContentWhenClosed:!1,"data-testid":"coupon-code-accordion-section",className:e(["cart-coupons__accordion-section"]),children:[t("form",{"data-testid":"coupon-code-form",className:e(["coupon-code-form--edit"]),ref:c,children:u("div",{className:e(["coupon-code-form__action"]),children:[i&&t(r,{node:i,className:e(["coupon-code-form__codes"])}),s&&t(r,{node:s,className:e(["coupon-code-form--button"]),onClick:v,type:"submit"})]})}),m&&t(r,{node:m,className:e(["coupon-code-form__error"])}),d&&t(r,{node:d,className:e(["coupon-code-form__applied"])})]})})})};export{V as C};
+//# sourceMappingURL=Coupons.js.map
