+++ scripts/__dropins__/storefront-cart/chunks/EmptyCart.js
@@ -1,4 +1,4 @@
 /*! Copyright 2026 Adobe
 All Rights Reserved. */
-import{jsx as t}from"@dropins/tools/preact-jsx-runtime.js";import{E as p}from"./components.js";import"@dropins/tools/lib.js";import"@dropins/tools/preact-compat.js";import"@dropins/tools/components.js";import"@dropins/tools/preact-hooks.js";const n=({routeCTA:m})=>t(p,{ctaLinkURL:m==null?void 0:m()});export{n as E};
+import{jsx as t}from"@dropins/tools/preact-jsx-runtime.js";import{E as p}from"./EmptyCart2.js";import"@dropins/tools/lib.js";import"@dropins/tools/preact-compat.js";/* empty css                */import"@dropins/tools/components.js";import"@dropins/tools/preact-hooks.js";const c=({routeCTA:m})=>t(p,{ctaLinkURL:m==null?void 0:m()});export{c as E};
 //# sourceMappingURL=EmptyCart.js.map
