+++ scripts/__dropins__/storefront-cart/chunks/OrderSummary.js
@@ -1,4 +1,4 @@
 /*! Copyright 2026 Adobe
 All Rights Reserved. */
-import{jsx as i,jsxs as q}from"@dropins/tools/preact-jsx-runtime.js";import{useState as w,useEffect as H,useCallback as ar,Fragment as lr}from"@dropins/tools/preact-compat.js";import{Slot as ne}from"@dropins/tools/lib.js";import{events as ce}from"@dropins/tools/event-bus.js";import{getEstimatedTotals as dr,s as $,p as or,getCartDataFromCache as sr}from"../api.js";import"@dropins/tools/preact-hooks.js";import{Price as t,Button as mr,Icon as ir}from"@dropins/tools/components.js";import{c as xr,d as gr,a as Tr}from"./components.js";import{useText as pr,Text as tr}from"@dropins/tools/i18n.js";const Ir=()=>{const[ue,k]=w(!1),[x,o]=w();return{handleEstimateTotals:(K,s)=>{k(!0);const{shippingCountry:y,shippingState:R="",shippingStateId:M,shippingZip:e=""}=K,O={countryCode:y,postcode:e,region:{region:R,id:M},shipping_method:{carrier_code:(s==null?void 0:s.carrier_code)||"",method_code:(s==null?void 0:s.method_code)||""}};dr(O).then(l=>{var z,h,C,r,b,G,a,S,f,T,V,j,p,_,n,Q,Z,m,g,Y;l&&o({estimatedTaxTotal:{amount:(z=l.totalTax)==null?void 0:z.value,currency:(h=l.totalTax)==null?void 0:h.currency},estimatedSubTotal:{excludingTax:{amount:(r=(C=l.subtotal)==null?void 0:C.excludingTax)==null?void 0:r.value,currency:(G=(b=l.subtotal)==null?void 0:b.excludingTax)==null?void 0:G.currency},includingTax:{amount:(S=(a=l.subtotal)==null?void 0:a.includingTax)==null?void 0:S.value,currency:(T=(f=l.subtotal)==null?void 0:f.includingTax)==null?void 0:T.currency},includingDiscountOnly:{amount:(j=(V=l.subtotal)==null?void 0:V.includingDiscountOnly)==null?void 0:j.value,currency:(_=(p=l.subtotal)==null?void 0:p.includingDiscountOnly)==null?void 0:_.currency}},estimatedGrandTotalPrice:{includingTax:{amount:(n=l.total)==null?void 0:n.includingTax.value,currency:(Q=l.total)==null?void 0:Q.includingTax.currency},excludingTax:{amount:(Z=l.total)==null?void 0:Z.excludingTax.value,currency:(m=l.total)==null?void 0:m.excludingTax.currency}},estimatedAppliedTaxes:{taxes:(g=l.appliedTaxes)==null?void 0:g.map(d=>{var I,v;return{label:d.label,amount:{value:(I=d.amount)==null?void 0:I.value,currency:(v=d.amount)==null?void 0:v.currency}}})},estimatedItems:{items:(Y=l.items)==null?void 0:Y.map(d=>{var I,v,D,N,E,A,P,L,W,U;return{uid:d.uid,price:{amount:(I=d.price)==null?void 0:I.value,currency:(v=d.price)==null?void 0:v.currency},taxedPrice:{amount:(D=d.taxedPrice)==null?void 0:D.value,currency:(N=d.taxedPrice)==null?void 0:N.currency},rowTotal:{amount:(E=d.rowTotal)==null?void 0:E.value,currency:(A=d.rowTotal)==null?void 0:A.currency},rowTotalIncludingTax:{amount:(P=d.rowTotalIncludingTax)==null?void 0:P.value,currency:(L=d.rowTotalIncludingTax)==null?void 0:L.currency},regularPrice:{amount:(W=d.regularPrice)==null?void 0:W.value,currency:(U=d.regularPrice)==null?void 0:U.currency}}})}})}).finally(()=>{k(!1)})},estimatedTotals:x,setEstimatedTotals:o,loading:ue}},vr=100,hr=({children:ue,initialData:k=null,routeCheckout:x,slots:o,errors:J,showTotalSaved:K,enableCoupons:s,enableGiftCards:y,updateLineItems:R=e=>e,...M})=>{var D,N,E,A,P,L,W,U,ie,te,ae,le,de,oe,se,me,xe,ge,Te,pe,Ie,ve,he,ye,Ce,be,Se,fe,_e,De,Ne,Ee,Ae,Pe,Le,We,Ue,Fe,Xe,Be,we,ke,ze,Ge,Ve,je,Qe,Ze,Ye,qe,He,$e,Je,Ke,Re,Me,Oe,er,rr;const[e,O]=w(k),[l,z]=w(!1),h=e==null?void 0:e.isVirtual;s=s??!0,y=y??!0;const{handleEstimateTotals:C,estimatedTotals:r,setEstimatedTotals:b,loading:G}=Ir(),a=(D=$.config)==null?void 0:D.shoppingCartDisplaySetting,S=(a==null?void 0:a.subtotal)==="INCLUDING_TAX",f=(a==null?void 0:a.subtotal)==="INCLUDING_EXCLUDING_TAX",T=a==null?void 0:a.zeroTax,V=()=>(e==null?void 0:e.appliedGiftCards.reduce((u,c)=>u+c.appliedBalance.value,0))??0,j=u=>u!=null&&u.code?q(lr,{children:[q("span",{className:"cart-order-summary__coupon__code",children:[i(ir,{source:Tr,size:"16"}),i("span",{children:u==null?void 0:u.code})]}),q("span",{className:"cart-order-summary__caption",children:[i(tr,{id:"Cart.PriceSummary.giftCard.appliedGiftCards.remainingBalance"}),i(t,{className:"cart-order-summary__caption",weight:"normal",size:"small",amount:u.giftCardBalance.value,currency:u.giftCardBalance.currency})]})]},u.code):null,p=(N=$.config)==null?void 0:N.cartGiftWrapping,_=(E=$.config)==null?void 0:E.cartPrintedCard,n=e==null?void 0:e.totalGiftOptions;H(()=>{const u=ce.on("cart/data",c=>{var F,X,B;O(c);const ee=(F=c==null?void 0:c.addresses)==null?void 0:F.shipping,re=c==null?void 0:c.isVirtual;(ee||re)&&b(null),r==null&&b({estimatedTaxTotal:{amount:(X=c==null?void 0:c.totalTax)==null?void 0:X.value,currency:(B=c==null?void 0:c.totalTax)==null?void 0:B.currency}})},{eager:!0});return()=>{u==null||u.off()}},[]),H(()=>{z(J)},[J]),H(()=>{const u=ce.on("shipping/estimate",c=>{var F,X,B,nr,cr,ur;const ee={shippingCountry:(F=c==null?void 0:c.address)==null?void 0:F.countryCode,shippingState:(X=c==null?void 0:c.address)==null?void 0:X.region,shippingStateId:(B=c==null?void 0:c.address)==null?void 0:B.regionId,shippingZip:(nr=c==null?void 0:c.address)==null?void 0:nr.postCode},re={carrier_code:((cr=c==null?void 0:c.shippingMethod)==null?void 0:cr.carrierCode)||"",method_code:((ur=c==null?void 0:c.shippingMethod)==null?void 0:ur.methodCode)||""};C(ee,re)});return()=>{u==null||u.off()}},[C]);const[Q,Z]=w(!1);H(()=>{const u=ce.on("checkout/initialized",c=>{!c||(c==null?void 0:c.type)==="quote"||Z(!0)});return()=>{u==null||u.off()}},[]);const m=pr({checkout:"Cart.PriceSummary.checkout",free:"Cart.PriceSummary.total.free",orderSummary:"Cart.PriceSummary.orderSummary",taxToBeDetermined:"Cart.PriceSummary.taxToBeDetermined"}),g=(e==null?void 0:e.hasOutOfStockItems)||l,Y=ar(u=>{!g&&e&&(u.preventDefault(),or(e,$.locale),setTimeout(()=>{const c=x==null?void 0:x({cartId:e.id});c&&(window.location.href=c)},vr))},[g,e,x]),d=!h&&(o!=null&&o.EstimateShipping)?i(ne,{name:"EstimateShipping",slot:o.EstimateShipping},"estimateShippingId"):void 0;if(!Object.keys(e??{}).length||(e==null?void 0:e.totalQuantity)===0)return null;const I=s&&(o!=null&&o.Coupons)?i(ne,{name:"Coupons",slot:o.Coupons},"couponsId"):void 0,v=y&&(o!=null&&o.GiftCards)?i(ne,{name:"GiftCards",slot:o.GiftCards},"giftCardId"):void 0;return i(xr,{...M,"data-testid":"cart-order-summary",heading:i("div",{children:m.orderSummary}),shipping:d,coupons:I,giftCards:v,loading:G,updateLineItems:R,appliedGiftCards:(A=e==null?void 0:e.appliedGiftCards)!=null&&A.length?{label:i(tr,{id:"Cart.PriceSummary.giftCard.appliedGiftCards.label",plural:(P=e==null?void 0:e.appliedGiftCards)==null?void 0:P.length,fields:{count:(L=e==null?void 0:e.appliedGiftCards)==null?void 0:L.length}}),price:i(t,{className:"cart-order-summary__price",amount:-V(),currency:(U=(W=e==null?void 0:e.appliedGiftCards)==null?void 0:W[0])==null?void 0:U.appliedBalance.currency}),content:(ie=e==null?void 0:e.appliedGiftCards)==null?void 0:ie.map(j).filter(Boolean)}:void 0,printedCard:{renderContent:!!((te=n==null?void 0:n.printedCard)!=null&&te.value),taxIncluded:_==="INCLUDING_TAX",taxInclAndExcl:_==="INCLUDING_EXCLUDING_TAX",priceExclTax:i(t,{"data-testid":"printed-card",amount:(ae=n==null?void 0:n.printedCard)==null?void 0:ae.value,currency:(le=n==null?void 0:n.printedCard)==null?void 0:le.currency}),priceInclTax:i(t,{amount:(de=n==null?void 0:n.printedCardInclTax)==null?void 0:de.value,currency:(oe=n==null?void 0:n.printedCardInclTax)==null?void 0:oe.currency})},itemsGiftWrapping:{renderContent:!!((se=n==null?void 0:n.giftWrappingForItems)!=null&&se.value),taxIncluded:p==="INCLUDING_TAX",taxInclAndExcl:p==="INCLUDING_EXCLUDING_TAX",priceExclTax:i(t,{amount:(me=n==null?void 0:n.giftWrappingForItems)==null?void 0:me.value,currency:(xe=n==null?void 0:n.giftWrappingForItems)==null?void 0:xe.currency}),priceInclTax:i(t,{amount:(ge=n==null?void 0:n.giftWrappingForItemsInclTax)==null?void 0:ge.value,currency:(Te=n==null?void 0:n.giftWrappingForItemsInclTax)==null?void 0:Te.currency})},orderGiftWrapping:{renderContent:!!((pe=n==null?void 0:n.giftWrappingForOrder)!=null&&pe.value),taxIncluded:p==="INCLUDING_TAX",taxInclAndExcl:p==="INCLUDING_EXCLUDING_TAX",priceExclTax:i(t,{amount:(Ie=n==null?void 0:n.giftWrappingForOrder)==null?void 0:Ie.value,currency:(ve=n==null?void 0:n.giftWrappingForOrder)==null?void 0:ve.currency}),priceInclTax:i(t,{amount:(he=n==null?void 0:n.giftWrappingForOrderInclTax)==null?void 0:he.value,currency:(ye=n==null?void 0:n.giftWrappingForOrderInclTax)==null?void 0:ye.currency})},subTotal:{taxIncluded:S&&!T,taxExcluded:f,zeroTaxSubtotal:T,priceExcludingTax:(Ce=r==null?void 0:r.estimatedSubTotal)!=null&&Ce.excludingTax?i(t,{"data-testid":"subtotal",...(be=r==null?void 0:r.estimatedSubTotal)==null?void 0:be.excludingTax}):i(t,{"data-testid":"subtotal",amount:(fe=(Se=e==null?void 0:e.subtotal)==null?void 0:Se.excludingTax)==null?void 0:fe.value,currency:(De=(_e=e==null?void 0:e.subtotal)==null?void 0:_e.excludingTax)==null?void 0:De.currency}),price:!T&&S||!T&&f?(Ne=r==null?void 0:r.estimatedSubTotal)!=null&&Ne.includingTax?i(t,{"data-testid":"subtotal",...(Ee=r==null?void 0:r.estimatedSubTotal)==null?void 0:Ee.includingTax}):i(t,{"data-testid":"subtotal",amount:(Ae=e==null?void 0:e.subtotal.includingTax)==null?void 0:Ae.value,currency:(Pe=e==null?void 0:e.subtotal.includingTax)==null?void 0:Pe.currency}):i(t,{"data-testid":"subtotal",amount:(We=(Le=e==null?void 0:e.subtotal)==null?void 0:Le.excludingTax)==null?void 0:We.value,currency:(Fe=(Ue=e==null?void 0:e.subtotal)==null?void 0:Ue.excludingTax)==null?void 0:Fe.currency})},discounts:(Xe=e==null?void 0:e.appliedDiscounts)==null?void 0:Xe.map(u=>{var c;return{label:u.label,price:i(t,{"data-testid":"summary-discount-total",amount:-u.amount.value,currency:u.amount.currency,sale:!0}),coupon:u!=null&&u.coupon?q("span",{children:[i(ir,{source:gr,size:"16"}),(c=u==null?void 0:u.coupon)==null?void 0:c.code]}):void 0}}),taxTotal:h&&!(e!=null&&e.appliedTaxes.length)&&!Q||r&&r.estimatedTaxTotal==null?{price:i("span",{"data-testid":"tax-total-tbd",children:m.taxToBeDetermined})}:{price:r!=null&&r.estimatedTaxTotal?i(t,{"data-testid":"tax-total-estimated",...r==null?void 0:r.estimatedTaxTotal}):i(t,{"data-testid":"tax-total-actual",amount:(Be=e==null?void 0:e.totalTax)==null?void 0:Be.value,currency:(we=e==null?void 0:e.totalTax)==null?void 0:we.currency}),estimated:(!r||!r.estimatedTaxTotal)&&!((ke=e==null?void 0:e.addresses)!=null&&ke.shipping)},taxesApplied:h?void 0:a!=null&&a.fullSummary?(Ge=((ze=r==null?void 0:r.estimatedAppliedTaxes)==null?void 0:ze.taxes)||(e==null?void 0:e.appliedTaxes))==null?void 0:Ge.map(u=>({label:u.label,price:i(t,{"data-testid":"applied-taxes",amount:u.amount.value,currency:u.amount.currency})})):void 0,total:{price:r!=null&&r.estimatedGrandTotalPrice?((je=(Ve=r==null?void 0:r.estimatedGrandTotalPrice)==null?void 0:Ve.includingTax)==null?void 0:je.amount)===0?i("span",{"data-testid":"total-including-tax",children:m.free}):i(t,{"data-testid":"total-including-tax-estimated",...(Qe=r==null?void 0:r.estimatedGrandTotalPrice)==null?void 0:Qe.includingTax}):((Ze=e==null?void 0:e.total)==null?void 0:Ze.includingTax.value)===0?i("span",{"data-testid":"total-including-tax",children:m.free}):i(t,{"data-testid":"total-including-tax-actual",amount:(Ye=e==null?void 0:e.total)==null?void 0:Ye.includingTax.value,currency:(qe=e==null?void 0:e.total)==null?void 0:qe.includingTax.currency}),estimated:(!r||!!(r!=null&&r.estimatedTaxTotal))&&!((He=e==null?void 0:e.addresses)!=null&&He.shipping),priceWithoutTax:a!=null&&a.grandTotal?r!=null&&r.estimatedAppliedTaxes?((Je=($e=r==null?void 0:r.estimatedGrandTotalPrice)==null?void 0:$e.excludingTax)==null?void 0:Je.amount)===0?i("span",{"data-testid":"total-excluding-tax",children:m.free}):i(t,{"data-testid":"total-excluding-tax",...(Ke=r==null?void 0:r.estimatedGrandTotalPrice)==null?void 0:Ke.excludingTax}):((Re=e==null?void 0:e.total)==null?void 0:Re.excludingTax.value)===0?i("span",{"data-testid":"total-excluding-tax",children:m.free}):i(t,{"data-testid":"total-excluding-tax",amount:(Me=e==null?void 0:e.total)==null?void 0:Me.excludingTax.value,currency:(Oe=e==null?void 0:e.total)==null?void 0:Oe.excludingTax.currency}):void 0},primaryAction:x&&i(mr,{"data-testid":"checkout-button",variant:"primary",disabled:g,"aria-disabled":g,href:g?void 0:x({cartId:e.id}),onClick:Y,children:m.checkout}),totalSaved:K?i(t,{amount:(er=e==null?void 0:e.discount)==null?void 0:er.value,currency:(rr=e==null?void 0:e.total)==null?void 0:rr.includingTax.currency}):void 0})};hr.getInitialData=async function(){return sr()};export{hr as O};
+import{jsx as r,jsxs as p,Fragment as gr}from"@dropins/tools/preact-jsx-runtime.js";import{useState as V,useEffect as re,useCallback as Tr,Fragment as hr}from"@dropins/tools/preact-compat.js";import{VComponent as L,classes as lr,Slot as ce}from"@dropins/tools/lib.js";import{events as ie}from"@dropins/tools/event-bus.js";import{g as yr}from"./persisted-data.js";import{s as te}from"./resetCart.js";import{g as Cr}from"./getEstimatedTotals.js";import{a as Sr}from"./acdl.js";import"@dropins/tools/preact-hooks.js";import{Accordion as ur,AccordionSection as or,ProgressSpinner as Ir,Divider as br,Price as u,Button as fr,Icon as sr}from"@dropins/tools/components.js";/* empty css                */import{O as y}from"./OrderSummaryLine.js";import{S as mr}from"./ChevronUp.js";import{S as dr}from"./ChevronDown.js";import{useText as pr,Text as xr}from"@dropins/tools/i18n.js";import{S as _r}from"./Coupon.js";import{S as vr}from"./GiftCard.js";const Pr=({className:ee,children:Z,variant:C="primary",heading:d,loading:B=!0,subTotal:f,shipping:T,discounts:_,taxTotal:S,taxesApplied:k,total:e,primaryAction:F,coupons:o,giftCards:A,totalSaved:v,appliedGiftCards:m,printedCard:t,itemsGiftWrapping:x,orderGiftWrapping:g,updateLineItems:s=P=>P,...w})=>{const[P,E]=V(!1),l=pr({checkout:"Cart.PriceSummary.checkout",discountedPrice:"Cart.CartItem.discountedPrice",download:"Cart.CartItem.download",heading:"Cart.Cart.heading",message:"Cart.CartItem.message",regularPrice:"Cart.CartItem.regularPrice",recipient:"Cart.CartItem.recipient",sender:"Cart.CartItem.sender",file:"Cart.CartItem.file",files:"Cart.CartItem.files",orderSummary:"Cart.PriceSummary.orderSummary",taxesBreakdownTitle:"Cart.PriceSummary.taxes.breakdown",taxTotal:"Cart.PriceSummary.taxes.total",taxEstimated:"Cart.PriceSummary.taxes.estimated",taxTotalOnly:"Cart.PriceSummary.taxes.totalOnly",showTaxBreakdown:"Cart.PriceSummary.taxes.showBreakdown",hideTaxBreakdown:"Cart.PriceSummary.taxes.hideBreakdown",taxToBeDetermined:"Cart.PriceSummary.taxToBeDetermined",subtotalLabel:"Cart.PriceSummary.subTotal.label",subtotalWithTaxes:"Cart.PriceSummary.subTotal.withTaxes",subtotalWithoutTaxes:"Cart.PriceSummary.subTotal.withoutTaxes",totalEstimated:"Cart.PriceSummary.total.estimated",totalLabel:"Cart.PriceSummary.total.label",totalWithoutTax:"Cart.PriceSummary.total.withoutTax",totalSaved:"Cart.PriceSummary.total.saved",shippingLabel:"Cart.PriceSummary.shipping.label",zipPlaceholder:"Cart.PriceSummary.shipping.zipPlaceholder",editZipAction:"Cart.PriceSummary.shipping.editZipAction",shippingWithTaxes:"Cart.PriceSummary.shipping.withTaxes",shippingWithoutTaxes:"Cart.PriceSummary.shipping.withoutTaxes",shippingEstimated:"Cart.PriceSummary.shipping.estimated",shippingEstimatedByState:"Cart.PriceSummary.shipping.alternateField.state",shippingEstimatedByZip:"Cart.PriceSummary.shipping.alternateField.zip",destinationLinkAriaLabel:"Cart.PriceSummary.shipping.destinationLinkAriaLabel",applyButton:"Cart.PriceSummary.estimatedShippingForm.apply.label",countryField:"Cart.PriceSummary.estimatedShippingForm.country.placeholder",freeShipping:"Cart.PriceSummary.freeShipping",stateField:"Cart.PriceSummary.estimatedShippingForm.state.placeholder",zipField:"Cart.PriceSummary.estimatedShippingForm.zip.placeholder",printedCardTitle:"Cart.PriceSummary.giftOptionsTax.printedCard.title",printedCardInclTax:"Cart.PriceSummary.giftOptionsTax.printedCard.inclTax",printedCardExclTax:"Cart.PriceSummary.giftOptionsTax.printedCard.exclTax",itemGiftWrappingTitle:"Cart.PriceSummary.giftOptionsTax.itemGiftWrapping.title",itemGiftWrappingInclTax:"Cart.PriceSummary.giftOptionsTax.itemGiftWrapping.inclTax",itemGiftWrappingExclTax:"Cart.PriceSummary.giftOptionsTax.itemGiftWrapping.exclTax",orderGiftWrappingTitle:"Cart.PriceSummary.giftOptionsTax.orderGiftWrapping.title",orderGiftWrappingInclTax:"Cart.PriceSummary.giftOptionsTax.orderGiftWrapping.inclTax",orderGiftWrappingExclTax:"Cart.PriceSummary.giftOptionsTax.orderGiftWrapping.exclTax"}),O=m?r(y,{label:m==null?void 0:m.label,price:m==null?void 0:m.price,classSuffixes:["applied-gift-cards"],children:m==null?void 0:m.content},m==null?void 0:m.label):null,N=f&&p(y,{label:l.subtotalLabel,price:f.price,classSuffixes:["subTotal"],children:[f.taxIncluded&&r("div",{"data-testid":"sub-total-tax-caption",className:"cart-order-summary__caption",children:r("span",{children:l.subtotalWithTaxes})}),f.taxExcluded?r("div",{"data-testid":"sub-total-tax-caption-excluded",className:"cart-order-summary__caption",children:p("span",{children:[f.priceExcludingTax," ",l.subtotalWithoutTaxes]})}):void 0]},l.subtotalLabel),W=_&&_.length>0&&r(gr,{children:_.map(i=>p(y,{label:i.label,price:i.price,classSuffixes:["discount"],children:[i.coupon&&r(L,{node:i.coupon,className:"cart-order-summary__coupon__code"}),i.caption&&r(L,{node:i.caption,className:"cart-order-summary__caption"})]},i.label))}),n=k&&k.length>0?r(ur,{"data-testid":"tax-breakdown",className:"cart-order-summary__taxes",iconOpen:dr,iconClose:mr,children:p(or,{title:l.taxesBreakdownTitle,secondaryText:!P&&S?r(L,{node:S.price,className:"cart-order-summary__price"}):void 0,renderContentWhenClosed:!1,onStateChange:E,children:[r("div",{className:"cart-order-summary__appliedTaxes",children:k.map(i=>r(y,{label:i.label,price:i.price,classSuffixes:["taxEntry"],labelClassSuffix:"muted"},i.label))}),S&&r(y,{label:l.taxTotal,price:S.price,classSuffixes:["taxTotal"],labelClassSuffix:"muted"})]})}):S&&r(y,{label:S.estimated?l.taxEstimated:l.taxTotalOnly,price:S.price,classSuffixes:["taxTotal"],testId:"tax-total-only"}),U=t!=null&&t.renderContent?p(y,{label:l.printedCardTitle,price:t.taxInclAndExcl||t.taxIncluded?t.priceInclTax:t.priceExclTax,classSuffixes:["printedCardContent"],children:[t.taxIncluded&&r("div",{"data-testid":"printed-card-incl",className:"cart-order-summary__caption",children:r("span",{children:l.printedCardInclTax})}),t.taxInclAndExcl?r("div",{"data-testid":"printed-card-incl-excl",className:"cart-order-summary__caption",children:p("span",{children:[t.priceExclTax," ",l.printedCardExclTax]})}):void 0]},l.printedCardTitle):null,z=x!=null&&x.renderContent?p(y,{label:l.itemGiftWrappingTitle,price:x.taxInclAndExcl||x.taxIncluded?x.priceInclTax:x.priceExclTax,classSuffixes:["itemsGiftWrappingContent"],children:[x.taxIncluded?r("div",{"data-testid":"item-gift-wrapping-incl",className:"cart-order-summary__caption",children:r("span",{children:l.itemGiftWrappingInclTax})}):null,x.taxInclAndExcl?r("div",{"data-testid":"item-gift-wrapping-incl-excl",className:"cart-order-summary__caption",children:p("span",{children:[x.priceExclTax," ",l.itemGiftWrappingExclTax]})}):null]},l.itemGiftWrappingTitle):null,h=g!=null&&g.renderContent?p(y,{label:l.orderGiftWrappingTitle,price:g.taxInclAndExcl||g.taxIncluded?g.priceInclTax:g.priceExclTax,classSuffixes:["orderGiftWrappingContent"],children:[g.taxIncluded&&r("div",{"data-testid":"order-gift-wrapping-incl",className:"cart-order-summary__caption",children:r("span",{children:l.orderGiftWrappingInclTax})}),g.taxInclAndExcl?r("div",{"data-testid":"order-gift-wrapping-incl-excl",className:"cart-order-summary__caption",children:p("span",{children:[g.priceExclTax," ",l.orderGiftWrappingExclTax]})}):void 0]},l.orderGiftWrappingTitle):null,b=[{key:"subTotalContent",sortOrder:100,content:N},...T?[{key:"shippingContent",sortOrder:200,content:r(L,{node:T,className:"cart-order-summary__shipping"})}]:[],{key:"printedCard",sortOrder:300,content:U},{key:"itemsGiftWrappingContent",sortOrder:400,content:z},{key:"orderGiftWrappingContent",sortOrder:500,content:h},{key:"discountsContent",sortOrder:600,content:W},{key:"appliedGiftCardsContent",sortOrder:700,content:O},{key:"taxContent",sortOrder:800,content:n},...e?[{key:"taxContent",sortOrder:900,content:r(y,{label:e.estimated?l.totalEstimated:l.totalLabel,price:e.price,classSuffixes:["total","total--padded"],testId:"total-content",labelClassSuffix:"bold"})}]:[],...e!=null&&e.priceWithoutTax?[{key:"totalWithoutTaxContent",sortOrder:1e3,content:r(y,{label:l.totalWithoutTax,price:e.priceWithoutTax,classSuffixes:["totalWithoutTax"],testId:"total-without-tax",labelClassSuffix:"muted"})}]:[],...v?[{key:"totalSavedContent",sortOrder:1100,content:r(y,{label:l.totalSaved,price:v,classSuffixes:["saved"],testId:"total-saved",labelClassSuffix:"muted"})}]:[],...F?[{key:"primaryActionContent",sortOrder:1200,content:r("div",{className:lr(["cart-order-summary__entry","cart-order-summary__primaryAction"]),children:F})}]:[],...o?[{key:"couponsContent",sortOrder:1300,content:r(L,{node:o,className:"cart-order-summary__coupons"})}]:[],...A?[{key:"giftCardsContent",sortOrder:1400,content:r(L,{node:A,className:"cart-order-summary__gift-cards"})}]:[]],X=s(b).sort((i,I)=>i.sortOrder-I.sortOrder);return p("div",{...w,className:lr(["cart-order-summary",["cart-order-summary--loading",B],[`cart-order-summary__${C}`,C],ee]),children:[B&&r(Ir,{className:"cart-order-summary__spinner"}),p("div",{className:"cart-order-summary__heading",children:[d&&r(L,{node:d,className:"cart-order-summary__heading-text"}),r(br,{variant:"primary",className:"cart-order-summary__divider-primary"})]}),r("div",{className:"cart-order-summary__content",children:X.map(i=>Array.isArray(i.content)?r(ur,{className:i.className,actionIconPosition:"right",iconOpen:dr,iconClose:mr,children:r(or,{defaultOpen:!1,title:i.title,renderContentWhenClosed:!1,children:i.content.map(I=>I.content)})},i.key):i.content)})]})},Er=()=>{const[ee,Z]=V(!1),[C,d]=V();return{handleEstimateTotals:(f,T)=>{Z(!0);const{shippingCountry:_,shippingState:S="",shippingStateId:k,shippingZip:e=""}=f,F={countryCode:_,postcode:e,region:{region:S,id:k},shipping_method:{carrier_code:(T==null?void 0:T.carrier_code)||"",method_code:(T==null?void 0:T.method_code)||""}};Cr(F).then(o=>{var A,v,m,t,x,g,s,w,P,E,l,O,N,W,n,U,z,h,b,X;o&&d({estimatedTaxTotal:{amount:(A=o.totalTax)==null?void 0:A.value,currency:(v=o.totalTax)==null?void 0:v.currency},estimatedSubTotal:{excludingTax:{amount:(t=(m=o.subtotal)==null?void 0:m.excludingTax)==null?void 0:t.value,currency:(g=(x=o.subtotal)==null?void 0:x.excludingTax)==null?void 0:g.currency},includingTax:{amount:(w=(s=o.subtotal)==null?void 0:s.includingTax)==null?void 0:w.value,currency:(E=(P=o.subtotal)==null?void 0:P.includingTax)==null?void 0:E.currency},includingDiscountOnly:{amount:(O=(l=o.subtotal)==null?void 0:l.includingDiscountOnly)==null?void 0:O.value,currency:(W=(N=o.subtotal)==null?void 0:N.includingDiscountOnly)==null?void 0:W.currency}},estimatedGrandTotalPrice:{includingTax:{amount:(n=o.total)==null?void 0:n.includingTax.value,currency:(U=o.total)==null?void 0:U.includingTax.currency},excludingTax:{amount:(z=o.total)==null?void 0:z.excludingTax.value,currency:(h=o.total)==null?void 0:h.excludingTax.currency}},estimatedAppliedTaxes:{taxes:(b=o.appliedTaxes)==null?void 0:b.map(i=>{var I,D;return{label:i.label,amount:{value:(I=i.amount)==null?void 0:I.value,currency:(D=i.amount)==null?void 0:D.currency}}})},estimatedItems:{items:(X=o.items)==null?void 0:X.map(i=>{var I,D,j,Q,$,Y,q,H,J,K;return{uid:i.uid,price:{amount:(I=i.price)==null?void 0:I.value,currency:(D=i.price)==null?void 0:D.currency},taxedPrice:{amount:(j=i.taxedPrice)==null?void 0:j.value,currency:(Q=i.taxedPrice)==null?void 0:Q.currency},rowTotal:{amount:($=i.rowTotal)==null?void 0:$.value,currency:(Y=i.rowTotal)==null?void 0:Y.currency},rowTotalIncludingTax:{amount:(q=i.rowTotalIncludingTax)==null?void 0:q.value,currency:(H=i.rowTotalIncludingTax)==null?void 0:H.currency},regularPrice:{amount:(J=i.regularPrice)==null?void 0:J.value,currency:(K=i.regularPrice)==null?void 0:K.currency}}})}})}).finally(()=>{Z(!1)})},estimatedTotals:C,setEstimatedTotals:d,loading:ee}},Nr=100,kr=({children:ee,initialData:Z=null,routeCheckout:C,slots:d,errors:B,showTotalSaved:f,enableCoupons:T,enableGiftCards:_,updateLineItems:S=e=>e,...k})=>{var j,Q,$,Y,q,H,J,K,le,ue,oe,se,me,de,xe,pe,ge,Te,he,ye,Ce,Se,Ie,be,fe,_e,ve,Pe,Ee,Ne,ke,Ae,we,We,De,Le,Be,Fe,Oe,Ue,ze,Xe,Ve,Ze,je,Qe,$e,Ye,qe,He,Je,Ke,Re,Me,Ge,er,rr,tr,nr;const[e,F]=V(Z),[o,A]=V(!1),v=e==null?void 0:e.isVirtual;T=T??!0,_=_??!0;const{handleEstimateTotals:m,estimatedTotals:t,setEstimatedTotals:x,loading:g}=Er(),s=(j=te.config)==null?void 0:j.shoppingCartDisplaySetting,w=(s==null?void 0:s.subtotal)==="INCLUDING_TAX",P=(s==null?void 0:s.subtotal)==="INCLUDING_EXCLUDING_TAX",E=s==null?void 0:s.zeroTax,l=()=>(e==null?void 0:e.appliedGiftCards.reduce((c,a)=>c+a.appliedBalance.value,0))??0,O=c=>c!=null&&c.code?p(hr,{children:[p("span",{className:"cart-order-summary__coupon__code",children:[r(sr,{source:vr,size:"16"}),r("span",{children:c==null?void 0:c.code})]}),p("span",{className:"cart-order-summary__caption",children:[r(xr,{id:"Cart.PriceSummary.giftCard.appliedGiftCards.remainingBalance"}),r(u,{className:"cart-order-summary__caption",weight:"normal",size:"small",amount:c.giftCardBalance.value,currency:c.giftCardBalance.currency})]})]},c.code):null,N=(Q=te.config)==null?void 0:Q.cartGiftWrapping,W=($=te.config)==null?void 0:$.cartPrintedCard,n=e==null?void 0:e.totalGiftOptions;re(()=>{const c=ie.on("cart/data",a=>{var R,M,G;F(a);const ne=(R=a==null?void 0:a.addresses)==null?void 0:R.shipping,ae=a==null?void 0:a.isVirtual;(ne||ae)&&x(null),t==null&&x({estimatedTaxTotal:{amount:(M=a==null?void 0:a.totalTax)==null?void 0:M.value,currency:(G=a==null?void 0:a.totalTax)==null?void 0:G.currency}})},{eager:!0});return()=>{c==null||c.off()}},[]),re(()=>{A(B)},[B]),re(()=>{const c=ie.on("shipping/estimate",a=>{var R,M,G,ar,cr,ir;const ne={shippingCountry:(R=a==null?void 0:a.address)==null?void 0:R.countryCode,shippingState:(M=a==null?void 0:a.address)==null?void 0:M.region,shippingStateId:(G=a==null?void 0:a.address)==null?void 0:G.regionId,shippingZip:(ar=a==null?void 0:a.address)==null?void 0:ar.postCode},ae={carrier_code:((cr=a==null?void 0:a.shippingMethod)==null?void 0:cr.carrierCode)||"",method_code:((ir=a==null?void 0:a.shippingMethod)==null?void 0:ir.methodCode)||""};m(ne,ae)});return()=>{c==null||c.off()}},[m]);const[U,z]=V(!1);re(()=>{const c=ie.on("checkout/initialized",a=>{!a||(a==null?void 0:a.type)==="quote"||z(!0)});return()=>{c==null||c.off()}},[]);const h=pr({checkout:"Cart.PriceSummary.checkout",free:"Cart.PriceSummary.total.free",orderSummary:"Cart.PriceSummary.orderSummary",taxToBeDetermined:"Cart.PriceSummary.taxToBeDetermined"}),b=(e==null?void 0:e.hasOutOfStockItems)||o,X=Tr(c=>{!b&&e&&(c.preventDefault(),Sr(e,te.locale),setTimeout(()=>{const a=C==null?void 0:C({cartId:e.id});a&&(window.location.href=a)},Nr))},[b,e,C]),i=!v&&(d!=null&&d.EstimateShipping)?r(ce,{name:"EstimateShipping",slot:d.EstimateShipping},"estimateShippingId"):void 0;if(!Object.keys(e??{}).length||(e==null?void 0:e.totalQuantity)===0)return null;const I=T&&(d!=null&&d.Coupons)?r(ce,{name:"Coupons",slot:d.Coupons},"couponsId"):void 0,D=_&&(d!=null&&d.GiftCards)?r(ce,{name:"GiftCards",slot:d.GiftCards},"giftCardId"):void 0;return r(Pr,{...k,"data-testid":"cart-order-summary",heading:r("div",{children:h.orderSummary}),shipping:i,coupons:I,giftCards:D,loading:g,updateLineItems:S,appliedGiftCards:(Y=e==null?void 0:e.appliedGiftCards)!=null&&Y.length?{label:r(xr,{id:"Cart.PriceSummary.giftCard.appliedGiftCards.label",plural:(q=e==null?void 0:e.appliedGiftCards)==null?void 0:q.length,fields:{count:(H=e==null?void 0:e.appliedGiftCards)==null?void 0:H.length}}),price:r(u,{className:"cart-order-summary__price",amount:-l(),currency:(K=(J=e==null?void 0:e.appliedGiftCards)==null?void 0:J[0])==null?void 0:K.appliedBalance.currency}),content:(le=e==null?void 0:e.appliedGiftCards)==null?void 0:le.map(O).filter(Boolean)}:void 0,printedCard:{renderContent:!!((ue=n==null?void 0:n.printedCard)!=null&&ue.value),taxIncluded:W==="INCLUDING_TAX",taxInclAndExcl:W==="INCLUDING_EXCLUDING_TAX",priceExclTax:r(u,{"data-testid":"printed-card",amount:(oe=n==null?void 0:n.printedCard)==null?void 0:oe.value,currency:(se=n==null?void 0:n.printedCard)==null?void 0:se.currency}),priceInclTax:r(u,{amount:(me=n==null?void 0:n.printedCardInclTax)==null?void 0:me.value,currency:(de=n==null?void 0:n.printedCardInclTax)==null?void 0:de.currency})},itemsGiftWrapping:{renderContent:!!((xe=n==null?void 0:n.giftWrappingForItems)!=null&&xe.value),taxIncluded:N==="INCLUDING_TAX",taxInclAndExcl:N==="INCLUDING_EXCLUDING_TAX",priceExclTax:r(u,{amount:(pe=n==null?void 0:n.giftWrappingForItems)==null?void 0:pe.value,currency:(ge=n==null?void 0:n.giftWrappingForItems)==null?void 0:ge.currency}),priceInclTax:r(u,{amount:(Te=n==null?void 0:n.giftWrappingForItemsInclTax)==null?void 0:Te.value,currency:(he=n==null?void 0:n.giftWrappingForItemsInclTax)==null?void 0:he.currency})},orderGiftWrapping:{renderContent:!!((ye=n==null?void 0:n.giftWrappingForOrder)!=null&&ye.value),taxIncluded:N==="INCLUDING_TAX",taxInclAndExcl:N==="INCLUDING_EXCLUDING_TAX",priceExclTax:r(u,{amount:(Ce=n==null?void 0:n.giftWrappingForOrder)==null?void 0:Ce.value,currency:(Se=n==null?void 0:n.giftWrappingForOrder)==null?void 0:Se.currency}),priceInclTax:r(u,{amount:(Ie=n==null?void 0:n.giftWrappingForOrderInclTax)==null?void 0:Ie.value,currency:(be=n==null?void 0:n.giftWrappingForOrderInclTax)==null?void 0:be.currency})},subTotal:{taxIncluded:w&&!E,taxExcluded:P,zeroTaxSubtotal:E,priceExcludingTax:(fe=t==null?void 0:t.estimatedSubTotal)!=null&&fe.excludingTax?r(u,{"data-testid":"subtotal",...(_e=t==null?void 0:t.estimatedSubTotal)==null?void 0:_e.excludingTax}):r(u,{"data-testid":"subtotal",amount:(Pe=(ve=e==null?void 0:e.subtotal)==null?void 0:ve.excludingTax)==null?void 0:Pe.value,currency:(Ne=(Ee=e==null?void 0:e.subtotal)==null?void 0:Ee.excludingTax)==null?void 0:Ne.currency}),price:!E&&w||!E&&P?(ke=t==null?void 0:t.estimatedSubTotal)!=null&&ke.includingTax?r(u,{"data-testid":"subtotal",...(Ae=t==null?void 0:t.estimatedSubTotal)==null?void 0:Ae.includingTax}):r(u,{"data-testid":"subtotal",amount:(we=e==null?void 0:e.subtotal.includingTax)==null?void 0:we.value,currency:(We=e==null?void 0:e.subtotal.includingTax)==null?void 0:We.currency}):r(u,{"data-testid":"subtotal",amount:(Le=(De=e==null?void 0:e.subtotal)==null?void 0:De.excludingTax)==null?void 0:Le.value,currency:(Fe=(Be=e==null?void 0:e.subtotal)==null?void 0:Be.excludingTax)==null?void 0:Fe.currency})},discounts:(Oe=e==null?void 0:e.appliedDiscounts)==null?void 0:Oe.map(c=>{var a;return{label:c.label,price:r(u,{"data-testid":"summary-discount-total",amount:-c.amount.value,currency:c.amount.currency,sale:!0}),coupon:c!=null&&c.coupon?p("span",{children:[r(sr,{source:_r,size:"16"}),(a=c==null?void 0:c.coupon)==null?void 0:a.code]}):void 0}}),taxTotal:v&&!(e!=null&&e.appliedTaxes.length)&&!U||t&&t.estimatedTaxTotal==null?{price:r("span",{"data-testid":"tax-total-tbd",children:h.taxToBeDetermined})}:{price:t!=null&&t.estimatedTaxTotal?r(u,{"data-testid":"tax-total-estimated",...t==null?void 0:t.estimatedTaxTotal}):r(u,{"data-testid":"tax-total-actual",amount:(Ue=e==null?void 0:e.totalTax)==null?void 0:Ue.value,currency:(ze=e==null?void 0:e.totalTax)==null?void 0:ze.currency}),estimated:(!t||!t.estimatedTaxTotal)&&!((Xe=e==null?void 0:e.addresses)!=null&&Xe.shipping)},taxesApplied:v?void 0:s!=null&&s.fullSummary?(Ze=((Ve=t==null?void 0:t.estimatedAppliedTaxes)==null?void 0:Ve.taxes)||(e==null?void 0:e.appliedTaxes))==null?void 0:Ze.map(c=>({label:c.label,price:r(u,{"data-testid":"applied-taxes",amount:c.amount.value,currency:c.amount.currency})})):void 0,total:{price:t!=null&&t.estimatedGrandTotalPrice?((Qe=(je=t==null?void 0:t.estimatedGrandTotalPrice)==null?void 0:je.includingTax)==null?void 0:Qe.amount)===0?r("span",{"data-testid":"total-including-tax",children:h.free}):r(u,{"data-testid":"total-including-tax-estimated",...($e=t==null?void 0:t.estimatedGrandTotalPrice)==null?void 0:$e.includingTax}):((Ye=e==null?void 0:e.total)==null?void 0:Ye.includingTax.value)===0?r("span",{"data-testid":"total-including-tax",children:h.free}):r(u,{"data-testid":"total-including-tax-actual",amount:(qe=e==null?void 0:e.total)==null?void 0:qe.includingTax.value,currency:(He=e==null?void 0:e.total)==null?void 0:He.includingTax.currency}),estimated:(!t||!!(t!=null&&t.estimatedTaxTotal))&&!((Je=e==null?void 0:e.addresses)!=null&&Je.shipping),priceWithoutTax:s!=null&&s.grandTotal?t!=null&&t.estimatedAppliedTaxes?((Re=(Ke=t==null?void 0:t.estimatedGrandTotalPrice)==null?void 0:Ke.excludingTax)==null?void 0:Re.amount)===0?r("span",{"data-testid":"total-excluding-tax",children:h.free}):r(u,{"data-testid":"total-excluding-tax",...(Me=t==null?void 0:t.estimatedGrandTotalPrice)==null?void 0:Me.excludingTax}):((Ge=e==null?void 0:e.total)==null?void 0:Ge.excludingTax.value)===0?r("span",{"data-testid":"total-excluding-tax",children:h.free}):r(u,{"data-testid":"total-excluding-tax",amount:(er=e==null?void 0:e.total)==null?void 0:er.excludingTax.value,currency:(rr=e==null?void 0:e.total)==null?void 0:rr.excludingTax.currency}):void 0},primaryAction:C&&r(fr,{"data-testid":"checkout-button",variant:"primary",disabled:b,"aria-disabled":b,href:b?void 0:C({cartId:e.id}),onClick:X,children:h.checkout}),totalSaved:f?r(u,{amount:(tr=e==null?void 0:e.discount)==null?void 0:tr.value,currency:(nr=e==null?void 0:e.total)==null?void 0:nr.includingTax.currency}):void 0})};kr.getInitialData=async function(){return yr()};export{kr as O};
 //# sourceMappingURL=OrderSummary.js.map
