+++ scripts/__dropins__/storefront-cart/chunks/acdl.js
@@ -1,0 +1,4 @@
+/*! Copyright 2026 Adobe
+All Rights Reserved. */
+function C(o){const{cart:e,locale:t="en-US"}=o;return{id:e.id,items:g(e.items,t),prices:{subtotalExcludingTax:e.subtotal.excludingTax,subtotalIncludingTax:e.subtotal.includingTax},totalQuantity:e.totalUniqueItems,possibleOnepageCheckout:void 0,giftMessageSelected:void 0,giftWrappingSelected:void 0,source:void 0}}function g(o,e){return o.map(t=>({canApplyMsrp:!1,formattedPrice:b(e,t.price.currency,t.price.value),id:t.uid,prices:{price:t.price,discount:t.discount&&t.discount.value!==void 0?{value:t.discount.value,currency:t.discount.currency}:void 0},product:{productId:t.uid,name:t.name,sku:t.sku,topLevelSku:t.topLevelSku,specialToDate:void 0,specialFromDate:void 0,newToDate:void 0,newFromDate:void 0,createdAt:void 0,updatedAt:void 0,manufacturer:void 0,countryOfManufacture:void 0,categories:t.categories,productType:t.itemType,pricing:{regularPrice:t.regularPrice.value,minimalPrice:void 0,maximalPrice:void 0,specialPrice:P(t),tierPricing:void 0,currencyCode:t.regularPrice.currency},canonicalUrl:t.canonicalUrl,mainImageUrl:t.image.src,image:{src:t.image.src,alt:t.image.alt}},configurableOptions:t.selectedOptions?Object.entries(t.selectedOptions).map(([n,c],a)=>({id:a+1,optionLabel:n,valueId:a+1,valueLabel:c})):[],bundleOptions:t.bundleOptions?Object.entries(t.bundleOptions).map(([n,c],a)=>({id:(a+1).toString(),optionLabel:n,valueId:a+1,valueLabel:c})):[],customizableOptions:t.customizableOptions?Object.entries(t.customizableOptions).map(([n,c],a)=>({id:(a+1).toString(),optionLabel:n,valueId:a+1,valueLabel:c})):[],quantity:t.quantity,selectedOptions:(()=>{const n={...t.selectedOptions,...t.bundleOptions,...t.customizableOptions};return Object.keys(n).length>0?n:void 0})()}))}function b(o,e,t){const n=o.replace("_","-");return new Intl.NumberFormat(n,{style:"currency",currency:e}).format(t)}function P(o){var e;if(o.discounted)return(e=o.price)==null?void 0:e.value}const r={SHOPPING_CART_CONTEXT:"shoppingCartContext",PRODUCT_CONTEXT:"productContext",CHANGED_PRODUCTS_CONTEXT:"changedProductsContext",CHANNEL_CONTEXT:"channelContext"},u={OPEN_CART:"open-cart",ADD_TO_CART:"add-to-cart",REMOVE_FROM_CART:"remove-from-cart",SHOPPING_CART_VIEW:"shopping-cart-view",INITIATE_CHECKOUT:"initiate-checkout"};function _(){return window.adobeDataLayer=window.adobeDataLayer||[],window.adobeDataLayer}function i(o,e){const t=_();t.push({[o]:null}),t.push({[o]:e})}function T(o,e){_().push(n=>{const c=n.getState?n.getState():{};n.push({event:o,eventInfo:{...c,...e}})})}function h(){return{_id:"https://ns.adobe.com/xdm/channels/web",_type:"https://ns.adobe.com/xdm/channel-types/web"}}function l(){i(r.CHANNEL_CONTEXT,h())}function N(o,e,t){const n=C({cart:o,locale:t});l(),i(r.SHOPPING_CART_CONTEXT,{...n});const c=g(e,t);i(r.CHANGED_PRODUCTS_CONTEXT,{items:c}),T(u.OPEN_CART),c.forEach(a=>{i(r.PRODUCT_CONTEXT,a.product),d(n,[a],u.ADD_TO_CART)})}function D(o,e){const t=C({cart:o,locale:e});l(),i(r.SHOPPING_CART_CONTEXT,{...t}),T(u.SHOPPING_CART_VIEW)}function d(o,e,t){const n={items:e};l(),i(r.SHOPPING_CART_CONTEXT,{...o}),i(r.CHANGED_PRODUCTS_CONTEXT,{...n}),T(t)}function A(o,e,t){const n=C({cart:o,locale:t}),c=n.items,a=_(),v=a.getState?a.getState():{},{shoppingCartContext:{items:f=[]}={}}=v;e.forEach(E=>{const p=f.find(O=>O.id===E.uid),s=c.find(O=>O.id===E.uid);!s&&!p||(!p&&s?(i(r.PRODUCT_CONTEXT,s.product),d(n,[s],u.ADD_TO_CART)):p&&!s?(i(r.PRODUCT_CONTEXT,p.product),d(n,[p],u.REMOVE_FROM_CART)):s.quantity>p.quantity?(i(r.PRODUCT_CONTEXT,s.product),d(n,[s],u.ADD_TO_CART)):(i(r.PRODUCT_CONTEXT,s.product),d(n,[s],u.REMOVE_FROM_CART)))})}function I(o,e){const t=C({cart:o,locale:e});l(),i(r.SHOPPING_CART_CONTEXT,{...t}),T(u.INITIATE_CHECKOUT)}export{I as a,A as b,N as c,D as p};
+//# sourceMappingURL=acdl.js.map
