--- scripts/__dropins__/storefront-cart/chunks/components.js
@@ -1,4 +1,0 @@
-/*! Copyright 2026 Adobe
-All Rights Reserved. */
-import{classes as d,VComponent as h,getFormValues as ee}from"@dropins/tools/lib.js";import{useText as S,Text as G}from"@dropins/tools/i18n.js";import*as o from"@dropins/tools/preact-compat.js";import{useRef as te,useState as D}from"@dropins/tools/preact-compat.js";import{jsx as e,jsxs as a,Fragment as O}from"@dropins/tools/preact-jsx-runtime.js";import{IllustratedMessage as ue,Button as R,Icon as I,Divider as K,Skeleton as re,SkeletonRow as ae,Accordion as F,AccordionSection as Z,ProgressSpinner as ie,Modal as he,ContentGrid as fe,ImageSwatch as Ce,Price as U,Field as W,Checkbox as j,Input as X,TextArea as ge,Card as ne}from"@dropins/tools/components.js";import{useRef as ye,useEffect as Q,useState as _e,useId as ce,useCallback as xe}from"@dropins/tools/preact-hooks.js";import{s as Y}from"../api.js";const ve=t=>o.createElement("svg",{id:"Icon_Add_Base","data-name":"Icon \\u2013 Add \\u2013 Base",xmlns:"http://www.w3.org/2000/svg",width:24,height:24,viewBox:"0 0 24 24",...t},o.createElement("g",{id:"Large"},o.createElement("rect",{id:"Placement_area","data-name":"Placement area",width:24,height:24,fill:"#fff",opacity:0}),o.createElement("g",{id:"Add_icon","data-name":"Add icon",transform:"translate(9.734 9.737)"},o.createElement("line",{vectorEffect:"non-scaling-stroke",id:"Line_579","data-name":"Line 579",y2:12.7,transform:"translate(2.216 -4.087)",fill:"none",stroke:"currentColor"}),o.createElement("line",{vectorEffect:"non-scaling-stroke",id:"Line_580","data-name":"Line 580",x2:12.7,transform:"translate(-4.079 2.263)",fill:"none",stroke:"currentColor"})))),be=t=>o.createElement("svg",{width:24,height:24,viewBox:"0 0 24 24",fill:"none",xmlns:"http://www.w3.org/2000/svg",...t},o.createElement("g",{clipPath:"url(#clip0_102_196)"},o.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M18.3601 18.16H6.5601L4.8801 3H2.3501M19.6701 19.59C19.6701 20.3687 19.0388 21 18.2601 21C17.4814 21 16.8501 20.3687 16.8501 19.59C16.8501 18.8113 17.4814 18.18 18.2601 18.18C19.0388 18.18 19.6701 18.8113 19.6701 19.59ZM7.42986 19.59C7.42986 20.3687 6.79858 21 6.01986 21C5.24114 21 4.60986 20.3687 4.60986 19.59C4.60986 18.8113 5.24114 18.18 6.01986 18.18C6.79858 18.18 7.42986 18.8113 7.42986 19.59Z",stroke:"currentColor",strokeLinejoin:"round"}),o.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M5.25 6.37L20.89 8.06L20.14 14.8H6.19",stroke:"currentColor",strokeLinejoin:"round"})),o.createElement("defs",null,o.createElement("clipPath",{id:"clip0_102_196"},o.createElement("rect",{vectorEffect:"non-scaling-stroke",width:19.29,height:19.5,fill:"white",transform:"translate(2.3501 2.25)"})))),B=t=>o.createElement("svg",{width:24,height:24,viewBox:"0 0 24 24",fill:"none",xmlns:"http://www.w3.org/2000/svg",...t},o.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M0.75 12C0.75 5.78421 5.78421 0.75 12 0.75C18.2158 0.75 23.25 5.78421 23.25 12C23.25 18.2158 18.2158 23.25 12 23.25C5.78421 23.25 0.75 18.2158 0.75 12Z",stroke:"currentColor"}),o.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M6.75 12.762L10.2385 15.75L17.25 9",stroke:"currentColor"})),z=t=>o.createElement("svg",{width:24,height:24,viewBox:"0 0 24 24",fill:"none",xmlns:"http://www.w3.org/2000/svg",...t},o.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M7.74512 9.87701L12.0001 14.132L16.2551 9.87701",stroke:"currentColor",strokeWidth:1,strokeLinecap:"square",strokeLinejoin:"round"})),$e=t=>o.createElement("svg",{id:"Icon_Chevron_right_Base","data-name":"Icon \\u2013 Chevron right \\u2013 Base",xmlns:"http://www.w3.org/2000/svg",width:24,height:24,viewBox:"0 0 24 24",...t},o.createElement("g",{id:"Large"},o.createElement("rect",{id:"Placement_area","data-name":"Placement area",width:24,height:24,fill:"#fff",opacity:0}),o.createElement("g",{id:"Chevron_right_icon","data-name":"Chevron right icon"},o.createElement("path",{vectorEffect:"non-scaling-stroke",id:"chevron",d:"M199.75,367.5l4.255,-4.255-4.255,-4.255",transform:"translate(-189.25 -351.0)",fill:"none",stroke:"currentColor"})))),A=t=>o.createElement("svg",{width:24,height:24,viewBox:"0 0 24 24",fill:"none",xmlns:"http://www.w3.org/2000/svg",...t},o.createElement("path",{d:"M7.74512 14.132L12.0001 9.87701L16.2551 14.132",stroke:"currentColor",strokeWidth:1,strokeLinecap:"square",strokeLinejoin:"round"})),Be=t=>o.createElement("svg",{width:24,height:24,viewBox:"0 0 24 24",fill:"none",xmlns:"http://www.w3.org/2000/svg",...t},o.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M18.3599 5.64001L5.62988 18.37",stroke:"currentColor"}),o.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M18.3599 18.37L5.62988 5.64001",stroke:"currentColor"})),ke=t=>o.createElement("svg",{width:24,height:24,viewBox:"0 0 24 24",fill:"none",xmlns:"http://www.w3.org/2000/svg",...t},o.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M22 6.25H22.75C22.75 5.83579 22.4142 5.5 22 5.5V6.25ZM22 9.27L22.2514 9.97663C22.5503 9.87029 22.75 9.58731 22.75 9.27H22ZM20.26 12.92L19.5534 13.1714L19.5539 13.1728L20.26 12.92ZM22 14.66H22.75C22.75 14.3433 22.551 14.0607 22.2528 13.9539L22 14.66ZM22 17.68V18.43C22.4142 18.43 22.75 18.0942 22.75 17.68H22ZM2 17.68H1.25C1.25 18.0942 1.58579 18.43 2 18.43V17.68ZM2 14.66L1.74865 13.9534C1.44969 14.0597 1.25 14.3427 1.25 14.66H2ZM3.74 11.01L4.44663 10.7586L4.44611 10.7572L3.74 11.01ZM2 9.27H1.25C1.25 9.58675 1.44899 9.86934 1.7472 9.97611L2 9.27ZM2 6.25V5.5C1.58579 5.5 1.25 5.83579 1.25 6.25H2ZM21.25 6.25V9.27H22.75V6.25H21.25ZM21.7486 8.56337C19.8706 9.23141 18.8838 11.2889 19.5534 13.1714L20.9666 12.6686C20.5762 11.5711 21.1494 10.3686 22.2514 9.97663L21.7486 8.56337ZM19.5539 13.1728C19.9195 14.1941 20.7259 15.0005 21.7472 15.3661L22.2528 13.9539C21.6541 13.7395 21.1805 13.2659 20.9661 12.6672L19.5539 13.1728ZM21.25 14.66V17.68H22.75V14.66H21.25ZM22 16.93H2V18.43H22V16.93ZM2.75 17.68V14.66H1.25V17.68H2.75ZM2.25135 15.3666C4.12941 14.6986 5.11623 12.6411 4.44663 10.7586L3.03337 11.2614C3.42377 12.3589 2.85059 13.5614 1.74865 13.9534L2.25135 15.3666ZM4.44611 10.7572C4.08045 9.73588 3.27412 8.92955 2.2528 8.56389L1.7472 9.97611C2.34588 10.1905 2.81955 10.6641 3.03389 11.2628L4.44611 10.7572ZM2.75 9.27V6.25H1.25V9.27H2.75ZM2 7H22V5.5H2V7ZM7.31 6.74V18.17H8.81V6.74H7.31ZM17.0997 8.39967L11.0397 14.4597L12.1003 15.5203L18.1603 9.46033L17.0997 8.39967ZM12.57 9.67C12.57 9.87231 12.4159 10 12.27 10V11.5C13.2839 11.5 14.07 10.6606 14.07 9.67H12.57ZM12.27 10C12.1241 10 11.97 9.87231 11.97 9.67H10.47C10.47 10.6606 11.2561 11.5 12.27 11.5V10ZM11.97 9.67C11.97 9.46769 12.1241 9.34 12.27 9.34V7.84C11.2561 7.84 10.47 8.67938 10.47 9.67H11.97ZM12.27 9.34C12.4159 9.34 12.57 9.46769 12.57 9.67H14.07C14.07 8.67938 13.2839 7.84 12.27 7.84V9.34ZM17.22 14.32C17.22 14.5223 17.0659 14.65 16.92 14.65V16.15C17.9339 16.15 18.72 15.3106 18.72 14.32H17.22ZM16.92 14.65C16.7741 14.65 16.62 14.5223 16.62 14.32H15.12C15.12 15.3106 15.9061 16.15 16.92 16.15V14.65ZM16.62 14.32C16.62 14.1177 16.7741 13.99 16.92 13.99V12.49C15.9061 12.49 15.12 13.3294 15.12 14.32H16.62ZM16.92 13.99C17.0659 13.99 17.22 14.1177 17.22 14.32H18.72C18.72 13.3294 17.9339 12.49 16.92 12.49V13.99Z",fill:"currentColor"})),J=t=>o.createElement("svg",{width:20,height:23,viewBox:"0 0 20 23",fill:"none",xmlns:"http://www.w3.org/2000/svg",...t},o.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M10 6L10 21.5M10 6H12.25C13.4926 6 14.5 4.99264 14.5 3.75C14.5 2.50736 13.4926 1.5 12.25 1.5C11.0074 1.5 10 2.50736 10 3.75M10 6V3.75M10 6H7.75C6.50736 6 5.5 4.99264 5.5 3.75C5.5 2.50736 6.50736 1.5 7.75 1.5C8.99264 1.5 10 2.50736 10 3.75M3.25 10.75H16.75C17.9926 10.75 19 9.74264 19 8.5C19 7.25736 17.9926 6.25 16.75 6.25H3.25C2.00736 6.25 1 7.25736 1 8.5C1 9.74264 2.00736 10.75 3.25 10.75ZM4.75 21.5H15.25C16.4926 21.5 17.5 20.4926 17.5 19.25V13.25C17.5 12.0074 16.4926 11 15.25 11H4.75C3.50736 11 2.5 12.0074 2.5 13.25V19.25C2.5 20.4926 3.50736 21.5 4.75 21.5Z",stroke:"currentColor",strokeWidth:1})),Re=t=>o.createElement("svg",{width:24,height:24,viewBox:"0 0 24 24",fill:"none",xmlns:"http://www.w3.org/2000/svg",...t},o.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M22 12.8571V6.75774C22 6.33761 21.5848 6 21.0884 6H2.91155C2.40614 6 2 6.34511 2 6.75774V12.8571M22 12.8571H2M22 12.8571V17.0049C22 17.5468 21.5939 18 21.0884 18H2.91155C2.41516 18 2 17.5567 2 17.0049V12.8571M5.18182 16.0886L8.64545 12.8229M8.64545 12.8229L12.1087 16.0886M8.64545 12.8229C8.64545 12.8229 8.09964 10.4743 7.56328 9.96857C7.02692 9.46286 6.15419 9.46286 5.61783 9.96857C5.08146 10.4743 5.08146 11.2971 5.61783 11.8029C6.15419 12.3086 8.64545 12.8229 8.64545 12.8229ZM8.64545 12.8229C8.64545 12.8229 11.1542 12.2914 11.6905 11.7857C12.2269 11.28 12.2269 10.4571 11.6905 9.95141C11.1542 9.4457 10.2814 9.4457 9.74507 9.95141C9.2087 10.4571 8.64545 12.8229 8.64545 12.8229ZM8.64506 17.2886V6.70286",stroke:"currentColor",strokeWidth:1,strokeLinecap:"round"})),Te=t=>o.createElement("svg",{width:24,height:24,viewBox:"0 0 24 24",fill:"none",xmlns:"http://www.w3.org/2000/svg",...t},o.createElement("path",{d:"M17.3332 11.75H6.6665",strokeWidth:1,strokeLinecap:"square",strokeLinejoin:"round",vectorEffect:"non-scaling-stroke",fill:"none",stroke:"currentColor"})),ze=t=>o.createElement("svg",{width:24,height:24,viewBox:"0 0 24 24",fill:"none",xmlns:"http://www.w3.org/2000/svg",...t},o.createElement("g",{clipPath:"url(#clip0_4797_15331)"},o.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M10.25 20.91L1.5 17.55V6.51996L10.25 9.92996V20.91Z",stroke:"currentColor",strokeLinecap:"round",strokeLinejoin:"round",strokeWidth:1}),o.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M6.24023 4.64001L14.9902 8.06001V11.42",stroke:"currentColor",strokeLinecap:"round",strokeLinejoin:"round",strokeWidth:1}),o.createElement("path",{className:"error-icon",vectorEffect:"non-scaling-stroke",d:"M19 13.31L15.5 19.37H22.5L19 13.31Z",stroke:"currentColor",strokeLinecap:"round",strokeLinejoin:"round",strokeWidth:1}),o.createElement("path",{className:"error-icon",vectorEffect:"non-scaling-stroke",d:"M19.0202 17.11H18.9802L18.9502 15.56H19.0502L19.0202 17.11ZM18.9602 18.29V18.06H19.0502V18.29H18.9602Z",stroke:"currentColor",strokeLinecap:"round",strokeLinejoin:"round",strokeWidth:1}),o.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M19 12.16V6.51996L10.25 9.92996V20.91L14.27 19.37L14.4 19.32",stroke:"currentColor",strokeLinejoin:"round",strokeWidth:1}),o.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M1.5 6.51999L10.25 3.04999L19 6.51999L10.25 9.92999L1.5 6.51999Z",stroke:"currentColor",strokeLinecap:"round",strokeLinejoin:"round",strokeWidth:1})),o.createElement("defs",null,o.createElement("clipPath",{id:"clip0_4797_15331"},o.createElement("rect",{width:22,height:18.86,fill:"white",transform:"translate(1 2.54999)"})))),Ae=t=>o.createElement("svg",{xmlns:"http://www.w3.org/2000/svg",width:24,height:24,viewBox:"0 0 24 24",fill:"none",...t},o.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M1 5H23",stroke:"currentColor",strokeWidth:1,strokeMiterlimit:10}),o.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M17.3674 22H6.63446C5.67952 22 4.88992 21.2688 4.8379 20.3338L4 5H20L19.1621 20.3338C19.1119 21.2688 18.3223 22 17.3655 22H17.3674Z",stroke:"currentColor",strokeWidth:1,strokeMiterlimit:10}),o.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M9.87189 2H14.1281C14.6085 2 15 2.39766 15 2.88889V5H9V2.88889C9 2.39912 9.39006 2 9.87189 2Z",stroke:"currentColor",strokeWidth:1,strokeMiterlimit:10}),o.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M8.87402 8.58057L9.39348 17.682",stroke:"currentColor",strokeWidth:1,strokeMiterlimit:10}),o.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M14.6673 8.58057L14.146 17.682",stroke:"currentColor",strokeWidth:1,strokeMiterlimit:10})),qe=t=>o.createElement("svg",{width:24,height:24,viewBox:"0 0 24 24",fill:"none",xmlns:"http://www.w3.org/2000/svg",...t},o.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M0.75 12C0.75 5.78421 5.78421 0.75 12 0.75C18.2158 0.75 23.25 5.78421 23.25 12C23.25 18.2158 18.2158 23.25 12 23.25C5.78421 23.25 0.75 18.2158 0.75 12Z",stroke:"currentColor"}),o.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M11.75 5.88423V4.75H12.25V5.88423L12.0485 13.0713H11.9515L11.75 5.88423ZM11.7994 18.25V16.9868H12.2253V18.25H11.7994Z",stroke:"currentColor"})),je=({className:t,children:n,ctaLinkURL:c,...i})=>{const r=S({emptyCart:"Cart.EmptyCart.heading",cta:"Cart.EmptyCart.cta"});return e("div",{...i,className:d(["cart-empty-cart",t]),children:e(ue,{className:d(["cart-empty-cart__wrapper",t]),"data-testid":"cart-empty-cart",heading:r.emptyCart,icon:e(I,{className:"cart-empty-cart__icon",source:be}),action:c?e(R,{"data-testid":"cart-empty-cart-button",size:"medium",variant:"primary",type:"submit",href:c,children:r.cta},"routeHome"):void 0})})},De=({countryField:t,destinationText:n,estimateButton:c,estimated:i,loading:r,onEstimate:l,price:C,priceExcludingTax:f,priceIncludingTax:s,stateField:m,taxExcluded:x,taxIncluded:p,zipField:w})=>{const g=te(null),[k,y]=D(!0),[v,_]=D("zip"),T=S({editZipAction:"Cart.EstimateShipping.editZipAction",destinationLinkAriaLabel:"Cart.EstimateShipping.destinationLinkAriaLabel",shippingLabel:"Cart.EstimateShipping.label",zipPlaceholder:"Cart.EstimateShipping.zipPlaceholder"}),L=u=>{u.preventDefault(),y(P=>!P)},M=u=>{u.preventDefault(),y(!0),_(P=>P==="zip"?"state":"zip")},E=u=>{u.preventDefault(),y(!1);const P=ee(g.current);l==null||l(P)},V=T.destinationLinkAriaLabel.replace("{destination}",n);return a("div",{"data-testid":"estimate-shipping",className:d(["cart-estimate-shipping",["cart-estimate-shipping--loading",r]]),children:[e("span",{className:"cart-estimate-shipping__label",children:i?n?a(O,{children:[e(G,{id:"Cart.EstimateShipping.estimatedDestination"})," ",e("a",{className:"cart-estimate-shippingLink",role:"button",href:"",onClick:L,onKeyDown:u=>{(u.key==="Enter"||u.key===" ")&&L(u)},tabIndex:0,"aria-label":V,"data-testid":"shipping-destination-link",children:n})]}):e(G,{id:"Cart.EstimateShipping.estimated"}):e(G,{id:"Cart.EstimateShipping.label"})}),e(h,{node:C,className:"cart-estimate-shipping__price"}),i&&a(O,{children:[e("div",{className:d(["cart-estimate-shipping__caption"]),children:e("a",{href:"#",className:"cart-estimate-shipping__link",onClick:M,"data-testid":"shipping-alternate-field-link",children:v==="zip"?e(G,{id:"Cart.EstimateShipping.alternateField.state"}):e(G,{id:"Cart.EstimateShipping.alternateField.zip"})})}),a("form",{className:d(["cart-estimate-shipping--edit",["cart-estimate-shipping--hide",!k]]),ref:g,"data-testid":"shipping-estimate-form",children:[t&&e(h,{node:t,className:d(["cart-estimate-shipping--country"])}),v==="state"?m&&e(h,{node:m,className:d(["cart-estimate-shipping--state"])}):w&&e(h,{node:w,className:d(["cart-estimate-shipping--zip"])}),c&&e(h,{node:c,className:d(["cart-estimate-shipping--action"]),onClick:E,type:"submit"})]})]}),p&&e("div",{"data-testid":"shipping-tax-included",className:d(["cart-estimate-shipping__caption"]),children:a("span",{children:[s," ",e(G,{id:"Cart.EstimateShipping.withTaxes"})]})}),x?e("div",{"data-testid":"shipping-tax-included-excluded",className:d(["cart-estimate-shipping__caption"]),children:a("span",{children:[f," ",e(G,{id:"Cart.EstimateShipping.withoutTaxes"})]})}):void 0]})},Ke=({className:t,products:n,productListFooter:c,subtotal:i,subtotalExcludingTaxes:r,preCheckoutSection:l,ctas:C,...f})=>{const s=S({subtotal:"Cart.MiniCart.subtotal",subtotalExcludingTaxes:"Cart.MiniCart.subtotalExcludingTaxes"});return e("div",{...f,className:d(["cart-mini-cart",t]),children:n&&a(O,{children:[e("div",{className:"cart-mini-cart__products","data-testid":"mini-cart-products-wrapper",children:n}),c&&e("div",{className:d(["cart-mini-cart__productListFooter"]),"data-testid":"mini-cart-product-list-footer",children:e(h,{node:c})}),a("div",{className:"cart-mini-cart__footer","data-testid":"mini-cart-subtotals",children:[i&&a("div",{className:"cart-mini-cart__footer__estimated-total","data-testid":"mini-cart-subtotal",children:[s.subtotal,e(h,{node:i})]}),r&&a("div",{className:"cart-mini-cart__footer__estimated-total-excluding-taxes","data-testid":"mini-cart-subtotal-excluding-taxes",children:[s.subtotalExcludingTaxes,e(h,{node:r,className:d(["dropin-price-summary__price","dropin-price-summary__price--muted"])})]}),l&&e("div",{className:d(["cart-mini-cart__preCheckoutSection"]),"data-testid":"mini-cart-pre-checkout-section",children:e(h,{node:l})}),C&&e(h,{node:C,className:"cart-mini-cart__footer__ctas"})]})]})})},Qe=({className:t,children:n,emptyCart:c,products:i,...r})=>e("div",{...r,className:d(["cart-cart-summary-grid",t]),children:e(O,{children:e("div",{className:d(["cart-cart-summary-grid__content",["cart-cart-summary-grid__content--empty",!i]]),children:i||e(h,{node:c,className:"cart-cart-summary-grid__empty-cart"})})})}),Je=({className:t,children:n,heading:c,footer:i,emptyCart:r,products:l,outOfStockMessage:C,variant:f="primary",loading:s=!0,...m})=>e("div",{...m,className:d(["cart-cart-summary-list",t,`cart-cart-summary-list__background--${f}`]),children:s?e(we,{}):a(O,{children:[(c||C)&&a("div",{"data-testid":"cart-summary-list-heading-wrapper",className:d(["cart-cart-summary-list__heading",["cart-cart-summary-list__heading--full-width",!l]]),children:[c&&a(O,{children:[e(h,{node:c,className:"cart-cart-summary-list__heading-text"}),e(K,{variant:"primary",className:d(["cart-cart-summary-list__heading-divider"])})]}),C&&e(h,{node:C,className:"cart-cart-summary-list__out-of-stock-message"})]}),e("div",{className:d(["cart-cart-summary-list__content",["cart-cart-summary-list__content--empty",!l]]),children:l||e(h,{node:r,className:"cart-cart-summary-list__empty-cart"})}),i&&a(O,{children:[e(K,{variant:"primary",className:d(["cart-cart-summary-list__footer-divider"])}),e(h,{node:i,className:"cart-cart-summary-list__footer-text"})]})]})}),we=()=>e(re,{"data-testid":"cart-summary-list-skeleton",className:"cart-cart-summary-list__skeleton",rowGap:"medium",children:e(ae,{variant:"row",size:"xlarge",fullWidth:!0,lines:3,multilineGap:"small"})}),Ue=({className:t,children:n,variant:c="primary",heading:i,loading:r=!0,subTotal:l,shipping:C,discounts:f,taxTotal:s,taxesApplied:m,total:x,primaryAction:p,coupons:w,giftCards:g,totalSaved:k,appliedGiftCards:y,printedCard:v,itemsGiftWrapping:_,orderGiftWrapping:T,updateLineItems:L=E=>E,...M})=>{const[E,V]=D(!1),u=S({checkout:"Cart.PriceSummary.checkout",discountedPrice:"Cart.CartItem.discountedPrice",download:"Cart.CartItem.download",heading:"Cart.Cart.heading",message:"Cart.CartItem.message",regularPrice:"Cart.CartItem.regularPrice",recipient:"Cart.CartItem.recipient",sender:"Cart.CartItem.sender",file:"Cart.CartItem.file",files:"Cart.CartItem.files",orderSummary:"Cart.PriceSummary.orderSummary",taxesBreakdownTitle:"Cart.PriceSummary.taxes.breakdown",taxTotal:"Cart.PriceSummary.taxes.total",taxEstimated:"Cart.PriceSummary.taxes.estimated",taxTotalOnly:"Cart.PriceSummary.taxes.totalOnly",showTaxBreakdown:"Cart.PriceSummary.taxes.showBreakdown",hideTaxBreakdown:"Cart.PriceSummary.taxes.hideBreakdown",taxToBeDetermined:"Cart.PriceSummary.taxToBeDetermined",subtotalLabel:"Cart.PriceSummary.subTotal.label",subtotalWithTaxes:"Cart.PriceSummary.subTotal.withTaxes",subtotalWithoutTaxes:"Cart.PriceSummary.subTotal.withoutTaxes",totalEstimated:"Cart.PriceSummary.total.estimated",totalLabel:"Cart.PriceSummary.total.label",totalWithoutTax:"Cart.PriceSummary.total.withoutTax",totalSaved:"Cart.PriceSummary.total.saved",shippingLabel:"Cart.PriceSummary.shipping.label",zipPlaceholder:"Cart.PriceSummary.shipping.zipPlaceholder",editZipAction:"Cart.PriceSummary.shipping.editZipAction",shippingWithTaxes:"Cart.PriceSummary.shipping.withTaxes",shippingWithoutTaxes:"Cart.PriceSummary.shipping.withoutTaxes",shippingEstimated:"Cart.PriceSummary.shipping.estimated",shippingEstimatedByState:"Cart.PriceSummary.shipping.alternateField.state",shippingEstimatedByZip:"Cart.PriceSummary.shipping.alternateField.zip",destinationLinkAriaLabel:"Cart.PriceSummary.shipping.destinationLinkAriaLabel",applyButton:"Cart.PriceSummary.estimatedShippingForm.apply.label",countryField:"Cart.PriceSummary.estimatedShippingForm.country.placeholder",freeShipping:"Cart.PriceSummary.freeShipping",stateField:"Cart.PriceSummary.estimatedShippingForm.state.placeholder",zipField:"Cart.PriceSummary.estimatedShippingForm.zip.placeholder",printedCardTitle:"Cart.PriceSummary.giftOptionsTax.printedCard.title",printedCardInclTax:"Cart.PriceSummary.giftOptionsTax.printedCard.inclTax",printedCardExclTax:"Cart.PriceSummary.giftOptionsTax.printedCard.exclTax",itemGiftWrappingTitle:"Cart.PriceSummary.giftOptionsTax.itemGiftWrapping.title",itemGiftWrappingInclTax:"Cart.PriceSummary.giftOptionsTax.itemGiftWrapping.inclTax",itemGiftWrappingExclTax:"Cart.PriceSummary.giftOptionsTax.itemGiftWrapping.exclTax",orderGiftWrappingTitle:"Cart.PriceSummary.giftOptionsTax.orderGiftWrapping.title",orderGiftWrappingInclTax:"Cart.PriceSummary.giftOptionsTax.orderGiftWrapping.inclTax",orderGiftWrappingExclTax:"Cart.PriceSummary.giftOptionsTax.orderGiftWrapping.exclTax"}),P=y?e(N,{label:y==null?void 0:y.label,price:y==null?void 0:y.price,classSuffixes:["applied-gift-cards"],children:y==null?void 0:y.content},y==null?void 0:y.label):null,$=l&&a(N,{label:u.subtotalLabel,price:l.price,classSuffixes:["subTotal"],children:[l.taxIncluded&&e("div",{"data-testid":"sub-total-tax-caption",className:"cart-order-summary__caption",children:e("span",{children:u.subtotalWithTaxes})}),l.taxExcluded?e("div",{"data-testid":"sub-total-tax-caption-excluded",className:"cart-order-summary__caption",children:a("span",{children:[l.priceExcludingTax," ",u.subtotalWithoutTaxes]})}):void 0]},u.subtotalLabel),H=f&&f.length>0&&e(O,{children:f.map(b=>a(N,{label:b.label,price:b.price,classSuffixes:["discount"],children:[b.coupon&&e(h,{node:b.coupon,className:"cart-order-summary__coupon__code"}),b.caption&&e(h,{node:b.caption,className:"cart-order-summary__caption"})]},b.label))}),oe=m&&m.length>0?e(F,{"data-testid":"tax-breakdown",className:"cart-order-summary__taxes",iconOpen:z,iconClose:A,children:a(Z,{title:u.taxesBreakdownTitle,secondaryText:!E&&s?e(h,{node:s.price,className:"cart-order-summary__price"}):void 0,renderContentWhenClosed:!1,onStateChange:V,children:[e("div",{className:"cart-order-summary__appliedTaxes",children:m.map(b=>e(N,{label:b.label,price:b.price,classSuffixes:["taxEntry"],labelClassSuffix:"muted"},b.label))}),s&&e(N,{label:u.taxTotal,price:s.price,classSuffixes:["taxTotal"],labelClassSuffix:"muted"})]})}):s&&e(N,{label:s.estimated?u.taxEstimated:u.taxTotalOnly,price:s.price,classSuffixes:["taxTotal"],testId:"tax-total-only"}),le=v!=null&&v.renderContent?a(N,{label:u.printedCardTitle,price:v.taxInclAndExcl||v.taxIncluded?v.priceInclTax:v.priceExclTax,classSuffixes:["printedCardContent"],children:[v.taxIncluded&&e("div",{"data-testid":"printed-card-incl",className:"cart-order-summary__caption",children:e("span",{children:u.printedCardInclTax})}),v.taxInclAndExcl?e("div",{"data-testid":"printed-card-incl-excl",className:"cart-order-summary__caption",children:a("span",{children:[v.priceExclTax," ",u.printedCardExclTax]})}):void 0]},u.printedCardTitle):null,se=_!=null&&_.renderContent?a(N,{label:u.itemGiftWrappingTitle,price:_.taxInclAndExcl||_.taxIncluded?_.priceInclTax:_.priceExclTax,classSuffixes:["itemsGiftWrappingContent"],children:[_.taxIncluded?e("div",{"data-testid":"item-gift-wrapping-incl",className:"cart-order-summary__caption",children:e("span",{children:u.itemGiftWrappingInclTax})}):null,_.taxInclAndExcl?e("div",{"data-testid":"item-gift-wrapping-incl-excl",className:"cart-order-summary__caption",children:a("span",{children:[_.priceExclTax," ",u.itemGiftWrappingExclTax]})}):null]},u.itemGiftWrappingTitle):null,de=T!=null&&T.renderContent?a(N,{label:u.orderGiftWrappingTitle,price:T.taxInclAndExcl||T.taxIncluded?T.priceInclTax:T.priceExclTax,classSuffixes:["orderGiftWrappingContent"],children:[T.taxIncluded&&e("div",{"data-testid":"order-gift-wrapping-incl",className:"cart-order-summary__caption",children:e("span",{children:u.orderGiftWrappingInclTax})}),T.taxInclAndExcl?e("div",{"data-testid":"order-gift-wrapping-incl-excl",className:"cart-order-summary__caption",children:a("span",{children:[T.priceExclTax," ",u.orderGiftWrappingExclTax]})}):void 0]},u.orderGiftWrappingTitle):null,me=[{key:"subTotalContent",sortOrder:100,content:$},...C?[{key:"shippingContent",sortOrder:200,content:e(h,{node:C,className:"cart-order-summary__shipping"})}]:[],{key:"printedCard",sortOrder:300,content:le},{key:"itemsGiftWrappingContent",sortOrder:400,content:se},{key:"orderGiftWrappingContent",sortOrder:500,content:de},{key:"discountsContent",sortOrder:600,content:H},{key:"appliedGiftCardsContent",sortOrder:700,content:P},{key:"taxContent",sortOrder:800,content:oe},...x?[{key:"taxContent",sortOrder:900,content:e(N,{label:x.estimated?u.totalEstimated:u.totalLabel,price:x.price,classSuffixes:["total","total--padded"],testId:"total-content",labelClassSuffix:"bold"})}]:[],...x!=null&&x.priceWithoutTax?[{key:"totalWithoutTaxContent",sortOrder:1e3,content:e(N,{label:u.totalWithoutTax,price:x.priceWithoutTax,classSuffixes:["totalWithoutTax"],testId:"total-without-tax",labelClassSuffix:"muted"})}]:[],...k?[{key:"totalSavedContent",sortOrder:1100,content:e(N,{label:u.totalSaved,price:k,classSuffixes:["saved"],testId:"total-saved",labelClassSuffix:"muted"})}]:[],...p?[{key:"primaryActionContent",sortOrder:1200,content:e("div",{className:d(["cart-order-summary__entry","cart-order-summary__primaryAction"]),children:p})}]:[],...w?[{key:"couponsContent",sortOrder:1300,content:e(h,{node:w,className:"cart-order-summary__coupons"})}]:[],...g?[{key:"giftCardsContent",sortOrder:1400,content:e(h,{node:g,className:"cart-order-summary__gift-cards"})}]:[]],pe=L(me).sort((b,q)=>b.sortOrder-q.sortOrder);return a("div",{...M,className:d(["cart-order-summary",["cart-order-summary--loading",r],[`cart-order-summary__${c}`,c],t]),children:[r&&e(ie,{className:"cart-order-summary__spinner"}),a("div",{className:"cart-order-summary__heading",children:[i&&e(h,{node:i,className:"cart-order-summary__heading-text"}),e(K,{variant:"primary",className:"cart-order-summary__divider-primary"})]}),e("div",{className:"cart-order-summary__content",children:pe.map(b=>Array.isArray(b.content)?e(F,{className:b.className,actionIconPosition:"right",iconOpen:z,iconClose:A,children:e(Z,{defaultOpen:!1,title:b.title,renderContentWhenClosed:!1,children:b.content.map(q=>q.content)})},b.key):b.content)})]})},Xe=({accordionSectionTitle:t,accordionSectionIcon:n,className:c,children:i,couponCodeField:r,applyCouponsButton:l,appliedCoupons:C,error:f,onApplyCoupon:s,...m})=>{const x=te(null),p=S({couponTitle:"Cart.PriceSummary.coupon.title"}),w=g=>{var v;g.preventDefault();const k=ee(x.current);s==null||s(k);const y=(v=x==null?void 0:x.current)==null?void 0:v.querySelector("input");y&&(y.value="")};return e("div",{...m,"data-testid":"cart-coupons",className:d(["cart-coupons",c]),children:e(F,{"data-testid":"coupon-code",className:d(["cart-coupons__accordion"]),actionIconPosition:"right",iconOpen:ve,iconClose:Te,children:a(Z,{title:t??p.couponTitle,iconLeft:n??ke,showIconLeft:!0,renderContentWhenClosed:!1,"data-testid":"coupon-code-accordion-section",className:d(["cart-coupons__accordion-section"]),children:[e("form",{"data-testid":"coupon-code-form",className:d(["coupon-code-form--edit"]),ref:x,children:a("div",{className:d(["coupon-code-form__action"]),children:[r&&e(h,{node:r,className:d(["coupon-code-form__codes"])}),l&&e(h,{node:l,className:d(["coupon-code-form--button"]),onClick:w,type:"submit"})]})}),f&&e(h,{node:f,className:d(["coupon-code-form__error"])}),C&&e(h,{node:C,className:d(["coupon-code-form__applied"])})]})})})},Oe=({label:t,price:n,classSuffixes:c=[],labelClassSuffix:i,testId:r,children:l,...C})=>{const f="cart-order-summary__label",s="cart-order-summary__price";return a("div",{...C,...r?{"data-testid":r}:{},className:d(["cart-order-summary__entry",...c.map(m=>`cart-order-summary__${m}`)]),children:[e("span",{className:d([f,...i?[`${f}--${i}`]:[]]),children:t}),e(h,{node:n,className:d([s,...i?[`${s}--${i}`]:[]])}),l]})};function Le(t){const n=ye(null);return Q(()=>{if(!t||!n.current)return;const c=n.current,i=c.querySelectorAll('a, button, input, textarea, select, details, [tabindex]:not([tabindex="-1"])');if(i.length===0)return;const r=i[0],l=i[i.length-1];function C(f){f.key==="Tab"&&(f.shiftKey&&document.activeElement===r?(f.preventDefault(),l.focus()):!f.shiftKey&&document.activeElement===l&&(f.preventDefault(),r.focus()))}return c.addEventListener("keydown",C),()=>{c.removeEventListener("keydown",C)}},[t]),n}const Ee=({view:t,showModal:n,productName:c,giftWrappingConfig:i,imageSwatchImageNode:r,setShowModal:l,updateGiftOptions:C})=>{var w;const f=Le(n),s=S({modalTitle:"Cart.GiftOptions.modal.title",defaultTitle:"Cart.GiftOptions.modal.defaultTitle",modalWrappingText:"Cart.GiftOptions.modal.wrappingText",modalWrappingSubText:"Cart.GiftOptions.modal.wrappingSubText",modalConfirmButton:"Cart.GiftOptions.modal.modalConfirmButton",modalCancelButton:"Cart.GiftOptions.modal.modalCancelButton",ariaLabelModal:"Cart.GiftOptions.modal.ariaLabelModal",ariaLabelModalOpen:"Cart.GiftOptions.modal.ariaLabelModalOpen",ariaLabelModalClose:"Cart.GiftOptions.modal.ariaLabelModalClose",ariaLabelWrapping:"Cart.GiftOptions.modal.ariaLabelWrapping"}),[m,x]=_e();if(Q(()=>{const g=i.find(k=>k.selected)??i[0];x(g)},[i]),Q(()=>{const g=document==null?void 0:document.querySelector(".cart-gift-options-view__modal-wrapper");g&&(g==null||g.focus())},[]),!n||!i.length)return null;const p=c?`${s.modalTitle} ${c}`:s.defaultTitle;return e("div",{ref:f,className:"cart-gift-options-view__modal-wrapper",tabIndex:0,"aria-label":`${s.ariaLabelModal} ${n?s.ariaLabelModalOpen:s.ariaLabelModalClose} ${c}`,children:a(he,{"data-testid":`gift-option-modal-${t}`,className:"cart-gift-options-view__modal",size:"medium",title:a(O,{children:[e("span",{children:p}),m&&((w=m==null?void 0:m.price)==null?void 0:w.value)>0?e(U,{amount:m.price.value,currency:m.price.currency,weight:"normal"}):null]}),centered:!0,onClose:l,children:[a("div",{className:"cart-gift-options-view__modal-content",children:[e("span",{className:"cart-gift-options-view__modal-text",children:s.modalWrappingText}),e(fe,{emptyGridContent:e(O,{}),maxColumns:6,columnWidth:"100px",className:"cart-gift-options-view__modal-grid",children:i.map(g=>e(Ce,{groupAriaLabel:s.ariaLabelWrapping,selected:(m==null?void 0:m.uid)===(g==null?void 0:g.uid),onValue:()=>{x(g)},imageNode:r,name:"giftWrappingId",value:g.uid,src:g.image.url,alt:g.design,label:g.design,"data-testid":`gift-option-modal-image-${g.uid}`,className:"cart-gift-options-view__modal-grid-item"},g.uid))}),e("span",{className:"cart-gift-options-view__modal-sub-text",children:m==null?void 0:m.design})]}),e(R,{"data-testid":"gift-option-modal-confirm-button",type:"button",onClick:()=>{C("giftWrappingId",m==null?void 0:m.uid,{isGiftWrappingSelected:!0}),l()},children:s.modalConfirmButton}),e(R,{type:"button",variant:"secondary",onClick:l,"data-testid":"gift-option-modal-cancel-button",children:s.modalCancelButton})]})})},Ne=()=>e(re,{children:e(ae,{variant:"row",size:"small",fullWidth:!0,lines:1,multilineGap:"small"})}),Se=({className:t,view:n,item:c,giftOptions:i,disabled:r,cartData:l,giftWrappingConfig:C,setShowModal:f,onInputChange:s,areGiftOptionsVisible:m})=>{var V,u,P,$;const x=ce(),p=S({customize:`Cart.GiftOptions.${n}.customize`,giftReceiptIncludedTitle:`Cart.GiftOptions.${n}.giftReceiptIncluded.title`,giftReceiptIncludedText:`Cart.GiftOptions.${n}.giftReceiptIncluded.subtitle`,printedCardIncludedTitle:`Cart.GiftOptions.${n}.printedCardIncluded.title`,printedCardIncludedText:`Cart.GiftOptions.${n}.printedCardIncluded.subtitle`,giftOptionsWrapTitle:`Cart.GiftOptions.${n}.giftOptionsWrap.title`,giftOptionsWrapText:`Cart.GiftOptions.${n}.giftOptionsWrap.subtitle`,requiredFieldError:"Cart.GiftOptions.formText.requiredFieldError"}),w=c==null?void 0:c.productGiftWrapping,g=l==null?void 0:l.cartGiftWrapping,k=n==="product"?w:g,y=k==null?void 0:k.find(H=>H.uid===i.giftWrappingId),v=(y==null?void 0:y.design)??"",_=c==null?void 0:c.giftWrappingPrice,T=v?`${p.giftOptionsWrapText} ${v}`:"",L=+(((u=(V=Y.config)==null?void 0:V.printedCardPrice)==null?void 0:u.value)??0)>0?(P=Y.config)==null?void 0:P.printedCardPrice:null;let M=null;_!=null&&_.value?M=_:($=y==null?void 0:y.price)!=null&&$.value&&(M=y.price);const E=H=>H?a("span",{children:[" (+",e(U,{amount:H.value,currency:H.currency,weight:"normal"}),")"]}):null;return a("div",{className:d([t,[`${t}--hidden`,!m.isGiftOptionsVisible]]),children:[m.isGiftReceiptVisible?e(W,{disabled:r,className:"cart-gift-options-view__field-gift-receipt",children:e(j,{id:`giftReceiptIncluded-${x}`,disabled:r,name:"giftReceiptIncluded",checked:i.giftReceiptIncluded,placeholder:p.giftReceiptIncludedTitle,label:p.giftReceiptIncludedTitle,description:p.giftReceiptIncludedText,onChange:s})}):null,m.isPrintedCartVisible?e(W,{disabled:r,className:"cart-gift-options-view__field-printed-card",children:e(j,{id:`printedCardIncluded-${x}`,disabled:r,name:"printedCardIncluded",checked:i.printedCardIncluded,placeholder:p.printedCardIncludedTitle,label:a(O,{children:[p.printedCardIncludedTitle,E(L)]}),description:p.printedCardIncludedText,onChange:s})}):null,m.isGiftWrappingVisible?a(O,{children:[e(W,{disabled:r,className:"cart-gift-options-view__field-gift-wrap",children:e(j,{id:`giftOptionsWrap-${x}`,disabled:r,name:"isGiftWrappingSelected",checked:i.isGiftWrappingSelected,placeholder:p.giftOptionsWrapTitle,label:a(O,{children:[p.giftOptionsWrapTitle,E(M)]}),description:T,onChange:s})}),e(R,{disabled:r||!C.length,type:"button","data-testid":`gift-option-customize-${n}`,variant:"tertiary",onClick:()=>f(!0),children:p.customize})]}):null]})},Me=({view:t,giftOptions:n,disabled:c,errorMessage:i,onInputChange:r,onBlur:l,isGiftMessageVisible:C})=>{const f=ce(),s=S({formTitle:`Cart.GiftOptions.${t}.formContent.formTitle`,formTo:`Cart.GiftOptions.${t}.formContent.formTo`,formFrom:`Cart.GiftOptions.${t}.formContent.formFrom`,giftMessageTitle:`Cart.GiftOptions.${t}.formContent.giftMessageTitle`,formToPlaceholder:`Cart.GiftOptions.${t}.formContent.formToPlaceholder`,formFromPlaceholder:`Cart.GiftOptions.${t}.formContent.formFromPlaceholder`,formMessagePlaceholder:`Cart.GiftOptions.${t}.formContent.formMessagePlaceholder`});return C?a(O,{children:[e("span",{children:s.formTitle}),a("div",{children:[e("span",{children:s.formTo}),e(W,{disabled:c,error:i.recipientName,children:e(X,{id:`recipientName-${f}`,disabled:c,type:"text",name:"recipientName",value:n.recipientName,placeholder:s.formToPlaceholder,onChange:r,onBlur:l})})]}),a("div",{children:[e("span",{children:s.formFrom}),e(W,{disabled:c,error:i.senderName,children:e(X,{id:`senderName-${f}`,disabled:c,type:"text",name:"senderName",value:n.senderName,placeholder:s.formFromPlaceholder,onChange:r,onBlur:l})})]}),a("div",{children:[e("span",{children:s.giftMessageTitle}),e(W,{disabled:c,children:e(ge,{id:`message-${f}`,errorMessage:i.message,disabled:c,name:"message",value:n.message,label:s.formMessagePlaceholder,onChange:r,onBlur:l})})]})]}):null},Pe=({view:t,giftOptions:n,giftWrappingConfig:c,readOnlyFormOrderView:i})=>{const r=S({readOnlyProductTitle:"Cart.GiftOptions.product.readOnlyFormView.title",wrapping:"Cart.GiftOptions.product.readOnlyFormView.wrapping",recipient:"Cart.GiftOptions.product.readOnlyFormView.recipient",sender:"Cart.GiftOptions.product.readOnlyFormView.sender",message:"Cart.GiftOptions.product.readOnlyFormView.message",readOnlyOrderTitle:"Cart.GiftOptions.order.readOnlyFormView.title",readOnlyOrderGiftReceiptTitle:"Cart.GiftOptions.order.readOnlyFormView.giftReceipt",readOnlyOrderGiftReceiptText:"Cart.GiftOptions.order.readOnlyFormView.giftReceiptText",readOnlyOrderGiftPrintCardTitle:"Cart.GiftOptions.order.readOnlyFormView.printCard",readOnlyOrderGiftPrintCardText:"Cart.GiftOptions.order.readOnlyFormView.printCardText",readOnlyOrderGiftWrapTitle:"Cart.GiftOptions.order.readOnlyFormView.giftWrap",readOnlyOrderGiftWrapOptionsText:"Cart.GiftOptions.order.readOnlyFormView.giftWrapOptions",readOnlyOrderFormTitle:"Cart.GiftOptions.order.readOnlyFormView.formTitle",readOnlyOrderFormTo:"Cart.GiftOptions.order.readOnlyFormView.formTo",readOnlyOrderFormFrom:"Cart.GiftOptions.order.readOnlyFormView.formFrom",readOnlyOrderFormMessageTitle:"Cart.GiftOptions.order.readOnlyFormView.formMessageTitle"}),{recipientName:l,senderName:C,message:f,giftReceiptIncluded:s,printedCardIncluded:m,isGiftWrappingSelected:x}=n,p=c==null?void 0:c.find(({uid:_})=>_===(n==null?void 0:n.giftWrappingId)),w=!!l||!!C||!!f,g=p==null?void 0:p.design,k=s||m||(p==null?void 0:p.selected),y=[{id:1,title:r.wrapping,message:x?g:""},{id:2,title:r.recipient,message:l},{id:3,title:r.sender,message:C},{id:4,title:r.message,message:f}],v=y.every(({message:_})=>!_);if(t==="product"&&!v)return e(F,{"data-testid":"gift-options-product",iconClose:A,iconOpen:z,actionIconPosition:"right",children:e(Z,{title:r.readOnlyProductTitle,showIconLeft:!0,iconLeft:J,defaultOpen:!1,renderContentWhenClosed:!1,children:e("div",{children:y.filter(_=>_.message).map(_=>a("p",{children:[_.title," ",_.message]},_.id))})})});if(t==="order"&&(w||k)){const _=a("div",{className:"cart-gift-options-readonly__header",children:[e(I,{source:J,size:"24"}),e("span",{children:r.readOnlyOrderTitle})]}),T=a(O,{children:[s?a("div",{className:"cart-gift-options-readonly__checkboxes cart-gift-options-readonly__checkboxes--gift-receipt",children:[e(I,{source:B,size:"16"}),e("p",{children:r.readOnlyOrderGiftReceiptTitle}),e("p",{children:r.readOnlyOrderGiftReceiptText})]}):null,m?a("div",{className:"cart-gift-options-readonly__checkboxes cart-gift-options-readonly__checkboxes--print-card",children:[e(I,{source:B,size:"16"}),e("p",{children:r.readOnlyOrderGiftPrintCardTitle}),e("p",{children:r.readOnlyOrderGiftPrintCardText})]}):null,p!=null&&p.selected?a("div",{className:"cart-gift-options-readonly__checkboxes cart-gift-options-readonly__checkboxes--gift-wrap",children:[e(I,{source:B,size:"16"}),a("p",{children:[r.readOnlyOrderGiftWrapTitle," (+",e(U,{amount:p.price.value,currency:p.price.currency,weight:"normal"}),")"]}),e("p",{children:`${r.readOnlyOrderGiftWrapOptionsText} ${p==null?void 0:p.design}`})]}):null]}),L=w?a("div",{className:"cart-gift-options-readonly__form",children:[e("div",{children:r.readOnlyOrderFormTitle}),a("div",{children:[a("p",{children:[e("span",{children:r.readOnlyOrderFormTo}),e("span",{children:l})]}),a("p",{children:[e("span",{children:r.readOnlyOrderFormFrom}),e("span",{children:C})]})]}),a("div",{children:[e("p",{children:r.readOnlyOrderFormMessageTitle}),e("p",{children:f})]})]}):null;return a(ne,{variant:i,children:[_,T,L]})}return null},Ye=({item:t,view:n,loading:c,giftOptions:i,showModal:r,isEditable:l,errorsField:C,updateLoading:f,cartData:s,fieldsDisabled:m,isGiftOptionsApplied:x,giftWrappingConfig:p,readOnlyFormOrderView:w,isGiftMessageVisible:g,areGiftOptionsVisible:k,imageSwatchImageNode:y,onBlur:v,setShowModal:_,updateGiftOptions:T,onInputChange:L,handleFormMouseLeave:M})=>{const E=S({accordionHeading:`Cart.GiftOptions.${n}.accordionHeading`}),V=xe(u=>e(F,{"data-testid":"gift-options-product",iconClose:A,iconOpen:z,actionIconPosition:"right",children:e(Z,{title:a("div",{className:"cart-gift-options-view__icon--success",children:[e("span",{children:E.accordionHeading}),x?e(I,{source:B,size:"16"}):null]}),ariaLabelTitle:E.accordionHeading,showIconLeft:!0,iconLeft:J,defaultOpen:x||u,renderContentWhenClosed:!1,children:a(O,{children:[e(Se,{className:"cart-gift-options-view__top",view:n,item:t,giftOptions:i,disabled:m,onInputChange:L,cartData:s,giftWrappingConfig:p,setShowModal:_,areGiftOptionsVisible:k}),e("form",{className:"cart-gift-options-view__footer",onMouseLeave:M,children:e(Me,{view:n,giftOptions:i,disabled:m,errorMessage:C,onInputChange:L,onBlur:v,isGiftMessageVisible:g})})]})})}),[n,t,s,E,C,i,m,p,k,x,g,v,_,L,M]);return!k.isGiftOptionsVisible&&!g?null:a("div",{id:"cart-gift-options-view",className:d(["cart-gift-options-view",`cart-gift-options-view--${n}`,["cart-gift-options-view--loading",f]]),children:[f?e(ie,{className:"cart-gift-options-view__spinner"}):null,c?e(Ne,{}):a(O,{children:[l?null:e("div",{className:"cart-gift-options-view--readonly",children:e(Pe,{view:n,giftOptions:i,giftWrappingConfig:p,readOnlyFormOrderView:w})}),r?e(Ee,{view:n,productName:t&&"name"in t?t==null?void 0:t.name:"",showModal:r,giftWrappingConfig:p,setShowModal:()=>_(!1),updateGiftOptions:T,imageSwatchImageNode:y}):null,n==="product"&&l?V(!1):null,n==="order"&&l?e(ne,{variant:"secondary",children:V(!0)}):null]})]})},et=({entries:t,className:n,...c})=>{const i=S({itemLabel:"Cart.CartSummaryTable.item",priceLabel:"Cart.CartSummaryTable.price",qtyLabel:"Cart.CartSummaryTable.qty",subtotalLabel:"Cart.CartSummaryTable.subtotal",mobilePriceLabel:"Cart.CartSummaryTable.mobilePrice",mobileQtyLabel:"Cart.CartSummaryTable.mobileQty",mobileSubtotalLabel:"Cart.CartSummaryTable.mobileSubtotal"});return a("div",{...c,className:d(["cart-cart-summary-table",n]),children:[a("div",{className:"cart-cart-summary-table__header",children:[e("div",{className:"cart-cart-summary-table__header-item",children:i.itemLabel}),e("div",{className:"cart-cart-summary-table__header-price",children:i.priceLabel}),e("div",{className:"cart-cart-summary-table__header-qty",children:i.qtyLabel}),e("div",{className:"cart-cart-summary-table__header-subtotal",children:i.subtotalLabel})]}),e("div",{className:"cart-cart-summary-table__body",children:t.map(r=>r.undoBanner?e("div",{className:"cart-cart-summary-table__undo-banner",children:e(h,{node:r.undoBanner})},r.uid):a("div",{className:d(["cart-cart-summary-table__row",["cart-cart-summary-table__row--updating",r.updating],["cart-cart-summary-table__row--error",r.hasError]]),children:[e("div",{className:"cart-cart-summary-table__cell-item",children:e(h,{node:r.item})}),a("div",{className:"cart-cart-summary-table__cell-price",children:[e("span",{className:"cart-cart-summary-table__mobile-label",children:i.mobilePriceLabel}),e(h,{node:r.price})]}),a("div",{className:"cart-cart-summary-table__cell-qty",children:[e("span",{className:"cart-cart-summary-table__mobile-label",children:i.mobileQtyLabel}),e(h,{node:r.quantity})]}),a("div",{className:"cart-cart-summary-table__cell-subtotal",children:[e("span",{className:"cart-cart-summary-table__mobile-label",children:i.mobileSubtotalLabel}),e(h,{node:r.subtotal})]}),e("div",{className:"cart-cart-summary-table__item-footer",children:e(h,{className:"cart-cart-summary-table__item-actions",node:r.actions})})]},r.uid))})]})},N=({label:t,price:n,classSuffixes:c=[],labelClassSuffix:i,testId:r,children:l,...C})=>e(Oe,{...C,label:t,price:n,classSuffixes:c,labelClassSuffix:i,testId:r,children:l}),tt=({className:t,productTitle:n,sku:c,image:i,configurations:r,alert:l,warning:C,...f})=>a("div",{...f,className:d(["cart-cart-summary-table__item",t]),children:[i&&e(h,{className:"cart-cart-summary-table__item-image-wrapper",node:i}),a("div",{className:"cart-cart-summary-table__item-details",children:[n&&e(h,{className:"cart-cart-summary-table__item-name",node:n}),c&&e(h,{className:"cart-cart-summary-table__sku",node:c}),l&&e(h,{className:"cart-cart-summary-table__item-quantity-alert-text",node:l}),C&&e(h,{className:"cart-cart-summary-table__item-quantity-warning-text",node:C}),r&&e(h,{className:"cart-cart-summary-table__item-configurations",node:r})]})]});export{Xe as C,je as E,Ye as G,tt as I,Ke as M,N as O,Be as S,Re as a,Qe as b,Ue as c,ke as d,De as e,ze as f,z as g,$e as h,Je as i,qe as j,Ae as k,et as l};
-//# sourceMappingURL=components.js.map
