+++ scripts/__dropins__/storefront-cart/chunks/refreshCart.js
@@ -1,0 +1,100 @@
+/*! Copyright 2026 Adobe
+All Rights Reserved. */
+import{s as o,d as X,f as I,h as U}from"./resetCart.js";import{events as g}from"@dropins/tools/event-bus.js";import{Initializer as K,merge as j}from"@dropins/tools/lib.js";import{c as Y}from"./persisted-data.js";import{CART_FRAGMENT as A}from"../fragments.js";const z=new K({init:async r=>{const n={disableGuestCart:!1,...r};z.config.setConfig(n),D().catch(console.error)},listeners:()=>[g.on("authenticated",r=>{o.authenticated&&!r?g.emit("cart/reset",void 0):r&&!o.authenticated&&(o.authenticated=r,D().catch(console.error))},{eager:!0}),g.on("locale",async r=>{r!==o.locale&&(o.locale=r,D().catch(console.error))}),g.on("cart/reset",()=>{X().catch(console.error),g.emit("cart/data",null)}),g.on("cart/data",r=>{Y(r)}),g.on("checkout/updated",r=>{!r||(r==null?void 0:r.type)==="quote"||$().catch(console.error)}),g.on("requisitionList/alert",()=>{$().catch(console.error)})]}),B=z.config;function k(r){var c,e,u,l,i,_,a,f,y,d,m,h,b,x,t,T;if(!r)return null;const n={appliedGiftCards:((c=r==null?void 0:r.applied_gift_cards)==null?void 0:c.map(s=>{var w,R,M;const p={code:s.code??"",appliedBalance:{value:s.applied_balance.value??0,currency:s.applied_balance.currency??"USD"},currentBalance:{value:s.current_balance.value??0,currency:s.current_balance.currency??"USD"},expirationDate:s.expiration_date??""},v=(w=p==null?void 0:p.currentBalance)==null?void 0:w.value,C=(R=p==null?void 0:p.appliedBalance)==null?void 0:R.value,S=(M=p==null?void 0:p.currentBalance)==null?void 0:M.currency,L=v-C>0?v-C:0;return{...p,giftCardBalance:{value:L,currency:S}}}))??[],id:r.id,totalQuantity:tr(r),totalUniqueItems:r.itemsV2.items.length,totalGiftOptions:O((e=r==null?void 0:r.prices)==null?void 0:e.gift_options),giftReceiptIncluded:(r==null?void 0:r.gift_receipt_included)??!1,printedCardIncluded:(r==null?void 0:r.printed_card_included)??!1,cartGiftWrapping:((u=r==null?void 0:r.available_gift_wrappings)==null?void 0:u.map(s=>{var p,v,C,S,E;return{design:s.design??"",uid:s.uid,selected:((p=r==null?void 0:r.gift_wrapping)==null?void 0:p.uid)===s.uid,image:{url:((v=s==null?void 0:s.image)==null?void 0:v.url)??"",label:((C=s.image)==null?void 0:C.label)??""},price:{currency:((S=s==null?void 0:s.price)==null?void 0:S.currency)??"USD",value:((E=s==null?void 0:s.price)==null?void 0:E.value)??0}}}))??[],giftMessage:{senderName:((l=r==null?void 0:r.gift_message)==null?void 0:l.from)??"",recipientName:((i=r==null?void 0:r.gift_message)==null?void 0:i.to)??"",message:((_=r==null?void 0:r.gift_message)==null?void 0:_.message)??""},errors:cr(r==null?void 0:r.itemsV2),items:N(r==null?void 0:r.itemsV2),miniCartMaxItems:N(r==null?void 0:r.itemsV2).slice(0,((a=o.config)==null?void 0:a.miniCartMaxItemsDisplay)??10),total:{includingTax:{value:r.prices.grand_total.value,currency:r.prices.grand_total.currency},excludingTax:{value:r.prices.grand_total_excluding_tax.value,currency:r.prices.grand_total_excluding_tax.currency}},discount:P(r.prices.discounts,r.prices.grand_total.currency),subtotal:{excludingTax:{value:(f=r.prices.subtotal_excluding_tax)==null?void 0:f.value,currency:(y=r.prices.subtotal_excluding_tax)==null?void 0:y.currency},includingTax:{value:(d=r.prices.subtotal_including_tax)==null?void 0:d.value,currency:(m=r.prices.subtotal_including_tax)==null?void 0:m.currency},includingDiscountOnly:{value:(h=r.prices.subtotal_with_discount_excluding_tax)==null?void 0:h.value,currency:(b=r.prices.subtotal_with_discount_excluding_tax)==null?void 0:b.currency}},appliedTaxes:F(r.prices.applied_taxes),totalTax:P(r.prices.applied_taxes,r.prices.grand_total.currency),appliedDiscounts:F(r.prices.discounts),isVirtual:r.is_virtual,addresses:{shipping:r.shipping_addresses&&or(r)},isGuestCart:!o.authenticated,hasOutOfStockItems:_r(r),hasFullyOutOfStockItems:pr(r),appliedCoupons:r.applied_coupons};return j(n,(T=(t=(x=B.getConfig().models)==null?void 0:x.CartModel)==null?void 0:t.transformer)==null?void 0:T.call(t,r))}function O(r){var n,c,e,u,l,i,_,a,f,y,d,m;return{giftWrappingForItems:{value:((n=r==null?void 0:r.gift_wrapping_for_items)==null?void 0:n.value)??0,currency:((c=r==null?void 0:r.gift_wrapping_for_items)==null?void 0:c.currency)??"USD"},giftWrappingForItemsInclTax:{value:((e=r==null?void 0:r.gift_wrapping_for_items_incl_tax)==null?void 0:e.value)??0,currency:((u=r==null?void 0:r.gift_wrapping_for_items_incl_tax)==null?void 0:u.currency)??"USD"},giftWrappingForOrder:{value:((l=r==null?void 0:r.gift_wrapping_for_order)==null?void 0:l.value)??0,currency:((i=r==null?void 0:r.gift_wrapping_for_order)==null?void 0:i.currency)??"USD"},giftWrappingForOrderInclTax:{value:((_=r==null?void 0:r.gift_wrapping_for_order_incl_tax)==null?void 0:_.value)??0,currency:((a=r==null?void 0:r.gift_wrapping_for_order_incl_tax)==null?void 0:a.currency)??"USD"},printedCard:{value:((f=r==null?void 0:r.printed_card)==null?void 0:f.value)??0,currency:((y=r==null?void 0:r.printed_card)==null?void 0:y.currency)??"USD"},printedCardInclTax:{value:((d=r==null?void 0:r.printed_card_incl_tax)==null?void 0:d.value)??0,currency:((m=r==null?void 0:r.printed_card_incl_tax)==null?void 0:m.currency)??"USD"}}}function P(r,n){return r!=null&&r.length?r.reduce((c,e)=>({value:c.value+e.amount.value,currency:e.amount.currency}),{value:0,currency:n}):{value:0,currency:n}}function H(r,n){var c,e,u,l;return{src:r!=null&&r.useConfigurableParentThumbnail?n.product.thumbnail.url:((e=(c=n.configured_variant)==null?void 0:c.thumbnail)==null?void 0:e.url)||n.product.thumbnail.url,alt:r!=null&&r.useConfigurableParentThumbnail?n.product.thumbnail.label:((l=(u=n.configured_variant)==null?void 0:u.thumbnail)==null?void 0:l.label)||n.product.thumbnail.label}}function J(r){var n,c,e,u;return r.__typename==="ConfigurableCartItem"?{value:(c=(n=r.configured_variant)==null?void 0:n.price_range)==null?void 0:c.maximum_price.regular_price.value,currency:(u=(e=r.configured_variant)==null?void 0:e.price_range)==null?void 0:u.maximum_price.regular_price.currency}:r.__typename==="GiftCardCartItem"?{value:r.prices.price.value,currency:r.prices.price.currency}:{value:r.prices.original_item_price.value,currency:r.prices.original_item_price.currency}}function V(r,n){return r!=null&&r.length&&[...r].sort((e,u)=>u.quantity-e.quantity).find(e=>n>=e.quantity)||null}function Z(r){var i,_;const n=r.quantity,c=r.__typename==="ConfigurableCartItem",e=c?(i=r.configured_variant)==null?void 0:i.price_tiers:r.product.price_tiers,u=c?(_=r.configured_variant)==null?void 0:_.price_range:r.product.price_range,l=V(e,n);return l?l.discount.amount_off>0:(u==null?void 0:u.maximum_price.discount.amount_off)>0}function rr(r){var n,c,e;return{senderName:((n=r==null?void 0:r.gift_message)==null?void 0:n.from)??"",recipientName:((c=r==null?void 0:r.gift_message)==null?void 0:c.to)??"",message:((e=r==null?void 0:r.gift_message)==null?void 0:e.message)??""}}function er(r){return{currency:(r==null?void 0:r.currency)??"USD",value:(r==null?void 0:r.value)??0}}function N(r){var c;if(!((c=r==null?void 0:r.items)!=null&&c.length))return[];const n=o.config;return r.items.map(e=>{var u,l,i,_,a,f,y,d,m,h,b,x;return{giftWrappingAvailable:((u=e==null?void 0:e.product)==null?void 0:u.gift_wrapping_available)??!1,giftWrappingPrice:er((l=e==null?void 0:e.product)==null?void 0:l.gift_wrapping_price),giftMessage:rr(e),productGiftWrapping:((i=e==null?void 0:e.available_gift_wrapping)==null?void 0:i.map(t=>{var T,s,p,v,C;return{design:t.design??"",uid:t.uid,selected:((T=e.gift_wrapping)==null?void 0:T.uid)===t.uid,image:{url:((s=t==null?void 0:t.image)==null?void 0:s.url)??"",label:((p=t.image)==null?void 0:p.label)??""},price:{currency:((v=t==null?void 0:t.price)==null?void 0:v.currency)??"USD",value:((C=t==null?void 0:t.price)==null?void 0:C.value)??0}}}))??[],itemType:e.__typename,uid:e.uid,giftMessageAvailable:nr(e.product.gift_message_available),url:{urlKey:e.product.url_key,categories:e.product.categories.map(t=>t.url_key)},canonicalUrl:e.product.canonical_url,categories:e.product.categories.map(t=>t.name),priceTiers:e.__typename==="ConfigurableCartItem"?((a=(_=e.configured_variant)==null?void 0:_.price_tiers)==null?void 0:a.map(t=>t))||[]:((f=e.product.price_tiers)==null?void 0:f.map(t=>t))||[],quantity:e.quantity,sku:fr(e),topLevelSku:e.product.sku,name:e.product.name,image:H(n,e),price:{value:e.prices.price.value,currency:e.prices.price.currency},taxedPrice:{value:e.prices.price_including_tax.value,currency:e.prices.price_including_tax.currency},fixedProductTaxes:e.prices.fixed_product_taxes,rowTotal:{value:e.prices.row_total.value,currency:e.prices.row_total.currency},rowTotalIncludingTax:{value:e.prices.row_total_including_tax.value,currency:e.prices.row_total_including_tax.currency},links:sr(e.links),total:{value:(y=e.prices.original_row_total)==null?void 0:y.value,currency:(d=e.prices.original_row_total)==null?void 0:d.currency},discount:{value:e.prices.total_item_discount.value,currency:e.prices.total_item_discount.currency,label:(m=e.prices.discounts)==null?void 0:m.map(t=>t.label)},regularPrice:J(e),discounted:Z(e),bundleOptions:e.__typename==="BundleCartItem"?ur(e.bundle_options):null,bundleOptionsUIDs:e.__typename==="BundleCartItem"?lr(e.bundle_options):null,selectedOptions:(h=q(e.configurable_options))==null?void 0:h.options,selectedOptionsUIDs:(b=q(e.configurable_options))==null?void 0:b.uids,customizableOptions:ir(e.customizable_options),sender:e.__typename==="GiftCardCartItem"?e.sender_name:null,senderEmail:e.__typename==="GiftCardCartItem"?e.sender_email:null,recipient:e.__typename==="GiftCardCartItem"?e.recipient_name:null,recipientEmail:e.__typename==="GiftCardCartItem"?e.recipient_email:null,message:e.__typename==="GiftCardCartItem"?e.message:null,discountedTotal:{value:e.prices.row_total.value,currency:e.prices.row_total.currency},onlyXLeftInStock:e.__typename==="ConfigurableCartItem"?(x=e.configured_variant)==null?void 0:x.only_x_left_in_stock:e.product.only_x_left_in_stock,lowInventory:e.is_available&&e.product.only_x_left_in_stock!==null,insufficientQuantity:(e.__typename==="ConfigurableCartItem"?e.configured_variant:e.product).stock_status==="IN_STOCK"&&!e.is_available,outOfStock:e.product.stock_status==="OUT_OF_STOCK",stockLevel:ar(e),discountPercentage:gr(e),savingsAmount:yr(e),productAttributes:dr(e)}})}function nr(r){switch(+r){case 0:return!1;case 1:return!0;case 2:return null;default:return!1}}function cr(r){var c;const n=(c=r==null?void 0:r.items)==null?void 0:c.reduce((e,u)=>{var l;return(l=u.errors)==null||l.forEach(i=>{e.push({uid:u.uid,text:i.message})}),e},[]);return n!=null&&n.length?n:null}function F(r){return r!=null&&r.length?r.map(n=>({amount:{value:n.amount.value,currency:n.amount.currency},label:n.label,coupon:n.coupon})):[]}function ur(r){const n=r==null?void 0:r.map(e=>({uid:e.uid,label:e.label,value:e.values.map(u=>u.label).join(", ")})),c={};return n==null||n.forEach(e=>{c[e.label]=e.value}),Object.keys(c).length>0?c:null}function lr(r){if(!(r!=null&&r.length))return null;const n=[];return r.forEach(c=>{var e;if((e=c.values)!=null&&e.length){const u=c.values.map(l=>l.uid);n.push(...u)}}),n.length>0?n:null}function q(r){const n=r==null?void 0:r.map(u=>({uid:u.configurable_product_option_uid,label:u.option_label,value:u.value_label,valueUid:u.configurable_product_option_value_uid})),c={},e={};return n==null||n.forEach(u=>{c[u.label]=u.value,e[u.label]=u.valueUid}),{options:Object.keys(c).length>0?c:null,uids:Object.keys(e).length>0?e:null}}function ir(r){const n=r==null?void 0:r.map(e=>({uid:e.customizable_option_uid,label:e.label,type:e.type,values:e.values.map(u=>({uid:u.customizable_option_value_uid,label:u.label,value:u.value}))})),c={};return n==null||n.forEach(e=>{var u;switch(e.type){case"field":case"area":case"date_time":c[e.label]=e.values[0].value;break;case"radio":case"drop_down":c[e.label]=e.values[0].label;break;case"multiple":case"checkbox":c[e.label]=e.values.reduce((l,i)=>l?`${l}, ${i.label}`:i.label,"");break;case"file":{const l=new DOMParser,i=e.values[0].value,a=((u=l.parseFromString(i,"text/html").querySelector("a"))==null?void 0:u.textContent)||"";c[e.label]=a;break}}}),c}function tr(r){var n,c;return((n=o.config)==null?void 0:n.cartSummaryDisplayTotal)===0?r.itemsV2.items.length:((c=o.config)==null?void 0:c.cartSummaryDisplayTotal)===1?r.total_quantity:r.itemsV2.items.length}function sr(r){return(r==null?void 0:r.length)>0?{count:r.length,result:r.map(n=>n.title).join(", ")}:null}function or(r){var n,c,e,u;return(n=r.shipping_addresses)!=null&&n.length?(c=r.shipping_addresses)==null?void 0:c.map(l=>({countryCode:l.country.code,zipCode:l.postcode,regionCode:l.region.code})):(e=r.addresses)!=null&&e.length?(u=r.addresses)==null?void 0:u.filter(l=>l.default_shipping).map(l=>{var i;return l.default_shipping&&{countryCode:l.country_code,zipCode:l.postcode,regionCode:(i=l.region)==null?void 0:i.region_code}}):null}function _r(r){var n,c;return(c=(n=r==null?void 0:r.itemsV2)==null?void 0:n.items)==null?void 0:c.some(e=>{var u;return((u=e==null?void 0:e.product)==null?void 0:u.stock_status)==="OUT_OF_STOCK"||e.product.stock_status==="IN_STOCK"&&!e.is_available})}function ar(r){return r.not_available_message?r.product.quantity!=null?r.product.quantity:"noNumber":null}function pr(r){var n,c;return(c=(n=r==null?void 0:r.itemsV2)==null?void 0:n.items)==null?void 0:c.some(e=>{var u;return((u=e==null?void 0:e.product)==null?void 0:u.stock_status)==="OUT_OF_STOCK"})}function gr(r){var u,l,i,_,a,f,y,d;const n=r.quantity,c=V(r.product.price_tiers,n);if(c)return Math.round(c.discount.percent_off);let e;if(r.__typename==="ConfigurableCartItem")e=(_=(i=(l=(u=r==null?void 0:r.configured_variant)==null?void 0:u.price_range)==null?void 0:l.maximum_price)==null?void 0:i.discount)==null?void 0:_.percent_off;else{if(r.__typename==="BundleCartItem")return;e=(d=(y=(f=(a=r==null?void 0:r.product)==null?void 0:a.price_range)==null?void 0:f.maximum_price)==null?void 0:y.discount)==null?void 0:d.percent_off}if(e!==0)return Math.round(e)}function fr(r){var n;return r.__typename==="ConfigurableCartItem"?r.configured_variant.sku:((n=r.product)==null?void 0:n.variantSku)||r.product.sku}function yr(r){var e,u,l,i,_,a;let n,c;if(n=((u=(e=r==null?void 0:r.prices)==null?void 0:e.original_row_total)==null?void 0:u.value)-((i=(l=r==null?void 0:r.prices)==null?void 0:l.row_total)==null?void 0:i.value),c=(a=(_=r==null?void 0:r.prices)==null?void 0:_.row_total)==null?void 0:a.currency,n!==0)return{value:n,currency:c}}function dr(r){var n,c,e;return(e=(c=(n=r==null?void 0:r.product)==null?void 0:n.custom_attributesV2)==null?void 0:c.items)==null?void 0:e.map(u=>{const l=u.code.split("_").map(i=>i.charAt(0).toUpperCase()+i.slice(1)).join(" ");return{...u,code:l}})}function mr(r){var e,u;if(!r)return null;const n=l=>{switch(l){case 1:return"EXCLUDING_TAX";case 2:return"INCLUDING_TAX";case 3:return"INCLUDING_EXCLUDING_TAX";default:return"EXCLUDING_TAX"}},c=l=>{switch(+l){case 0:return!1;case 1:return!0;case 2:return null;default:return!1}};return{displayMiniCart:r.minicart_display,miniCartMaxItemsDisplay:r.minicart_max_items,cartExpiresInDays:r.cart_expires_in_days,cartSummaryDisplayTotal:r.cart_summary_display_quantity,cartSummaryMaxItems:r.max_items_in_order_summary,defaultCountry:r.default_country,categoryFixedProductTaxDisplaySetting:r.category_fixed_product_tax_display_setting,productFixedProductTaxDisplaySetting:r.product_fixed_product_tax_display_setting,salesFixedProductTaxDisplaySetting:r.sales_fixed_product_tax_display_setting,shoppingCartDisplaySetting:{zeroTax:r.shopping_cart_display_zero_tax,subtotal:n(r.shopping_cart_display_subtotal),price:n(r.shopping_cart_display_price),shipping:n(r.shopping_cart_display_shipping),fullSummary:r.shopping_cart_display_full_summary,grandTotal:r.shopping_cart_display_grand_total,taxGiftWrapping:r.shopping_cart_display_tax_gift_wrapping},useConfigurableParentThumbnail:r.configurable_thumbnail_source==="parent",allowGiftWrappingOnOrder:c(r==null?void 0:r.allow_gift_wrapping_on_order),allowGiftWrappingOnOrderItems:c(r==null?void 0:r.allow_gift_wrapping_on_order_items),allowGiftMessageOnOrder:c(r==null?void 0:r.allow_order),allowGiftMessageOnOrderItems:c(r==null?void 0:r.allow_items),allowGiftReceipt:!!+(r==null?void 0:r.allow_gift_receipt),allowPrintedCard:!!+(r==null?void 0:r.allow_printed_card),printedCardPrice:{currency:((e=r==null?void 0:r.printed_card_priceV2)==null?void 0:e.currency)??"",value:((u=r==null?void 0:r.printed_card_priceV2)==null?void 0:u.value)!=null?+r.printed_card_priceV2.value:0},cartGiftWrapping:n(+r.cart_gift_wrapping),cartPrintedCard:n(+r.cart_printed_card)}}const vr=`
+  fragment CUSTOMER_FRAGMENT on Customer {
+    addresses {
+      default_shipping
+      country_code
+      postcode
+      region {
+        region
+        region_code
+        region_id
+      }
+    }
+  }
+`,Cr=`
+  query GUEST_CART_QUERY(
+      $cartId: String!,
+      $pageSize: Int! = 100,
+      $currentPage: Int! = 1,
+      $itemsSortInput: QuoteItemsSortInput! = {field: CREATED_AT, order: DESC}
+    ) {
+
+    cart(cart_id: $cartId){
+      ...CART_FRAGMENT
+    }
+  }
+
+  ${A}
+`,hr=`
+  query CUSTOMER_CART_QUERY(
+      $pageSize: Int! = 100,
+      $currentPage: Int! = 1,
+      $itemsSortInput: QuoteItemsSortInput! = {field: CREATED_AT, order: DESC}
+    ) {
+     
+    customer {
+      ...CUSTOMER_FRAGMENT
+    }
+
+    cart: customerCart {
+      ...CART_FRAGMENT
+    }
+  }
+
+  ${vr}
+  ${A}
+`,G=async()=>{const r=o.authenticated,n=o.cartId;if(r)return I(hr,{method:"POST"}).then(({errors:c,data:e})=>{if(c)return U(c);const u={...e.cart,...e.customer};return k(u)});if(!n)throw new Error("No cart ID found");return I(Cr,{method:"POST",cache:"no-cache",variables:{cartId:n}}).then(({errors:c,data:e})=>c?U(c):k(e.cart))},br=`
+  mutation MERGE_CARTS_MUTATION(
+      $guestCartId: String!, 
+      $customerCartId: String!,
+      $pageSize: Int! = 100,
+      $currentPage: Int! = 1,
+      $itemsSortInput: QuoteItemsSortInput! = {field: CREATED_AT, order: DESC}
+    ) {
+      mergeCarts(
+        source_cart_id: $guestCartId,
+        destination_cart_id: $customerCartId
+      ) {
+        ...CART_FRAGMENT 
+      }
+  }
+
+  ${A}
+`,D=async()=>{if(o.initializing)return null;o.initializing=!0,o.config||(o.config=await Tr());const r=o.authenticated?await Q():await W();return g.emit("cart/initialized",r),g.emit("cart/data",r),o.initializing=!1,r};async function Q(){const r=o.cartId,n=await G();return n?(o.cartId=n.id,!r||n.id===r?n:await I(br,{variables:{guestCartId:r,customerCartId:n.id}}).then(()=>G()).then(c=>{const e={oldCartItems:n.items,newCart:c};return g.emit("cart/merged",e),c}).catch(()=>(console.error("Could not merge carts"),n))):null}async function W(){if(B.getConfig().disableGuestCart===!0||!o.cartId)return null;try{return await G()}catch(r){return console.error(r),null}}const xr=`
+query STORE_CONFIG_QUERY {
+  storeConfig {
+    minicart_display
+    minicart_max_items
+    cart_expires_in_days
+    cart_summary_display_quantity
+    max_items_in_order_summary
+    default_country
+    category_fixed_product_tax_display_setting
+    product_fixed_product_tax_display_setting
+    sales_fixed_product_tax_display_setting
+    shopping_cart_display_full_summary
+    shopping_cart_display_grand_total
+    shopping_cart_display_price
+    shopping_cart_display_shipping
+    shopping_cart_display_subtotal
+    shopping_cart_display_tax_gift_wrapping
+    shopping_cart_display_zero_tax
+    configurable_thumbnail_source
+    allow_gift_wrapping_on_order
+    allow_gift_wrapping_on_order_items
+    allow_order
+    allow_items
+    allow_gift_receipt
+    allow_printed_card
+    printed_card_priceV2 {
+      currency
+      value
+    }
+    cart_gift_wrapping
+    cart_printed_card
+  }
+}
+`,Tr=async()=>I(xr,{method:"GET",cache:"force-cache"}).then(({errors:r,data:n})=>r?U(r):mr(n.storeConfig)),$=async()=>{const r=o.authenticated?await Q():await W();return g.emit("cart/data",r),r};export{D as a,Q as b,B as c,W as d,Tr as e,G as g,z as i,$ as r,k as t};
+//# sourceMappingURL=refreshCart.js.map
