+++ scripts/__dropins__/storefront-cart/containers/CartSummaryList/CartSummaryList.d.ts
@@ -39,12 +39,6 @@ export interface CartSummaryListProps extends HTMLAttributes<HTMLDivElement> {
             onUndo: () => void;
             onDismiss: () => void;
         }>;
-        ConfirmDeleteBanner?: SlotProps<{
-            item: CartModel['items'][0];
-            loading: boolean;
-            onConfirm: () => void;
-            onCancel: () => void;
-        }>;
         ItemTitle?: SlotProps<{
             item: CartModel['items'][number];
         }>;
@@ -53,9 +47,7 @@ export interface CartSummaryListProps extends HTMLAttributes<HTMLDivElement> {
         }>;
         ItemQuantity?: SlotProps<{
             item: CartModel['items'][number];
-            enableUpdateItemQuantity: boolean | {
-                removeOnZero?: boolean;
-            };
+            enableUpdateItemQuantity: boolean;
             handleItemQuantityUpdate: (item: CartModel['items'][number], quantity: number) => void;
             itemsLoading: Set<string>;
             handleItemsError: (uid: string, message?: string) => void;
@@ -83,9 +75,7 @@ export interface CartSummaryListProps extends HTMLAttributes<HTMLDivElement> {
         }>;
     };
     enableRemoveItem?: boolean;
-    enableUpdateItemQuantity?: boolean | {
-        removeOnZero?: boolean;
-    };
+    enableUpdateItemQuantity?: boolean;
     onItemsErrorsChange?: (errors: Map<string, string>) => void;
     accordion?: boolean;
     variant?: 'primary' | 'secondary';
@@ -98,7 +88,6 @@ export interface CartSummaryListProps extends HTMLAttributes<HTMLDivElement> {
         text: string;
     }[];
     undo?: boolean;
-    confirmBeforeDelete?: boolean;
     includeOutOfStockItems?: boolean;
     /**
      * TEST ONLY: Allows test to inject recentlyRemovedItems for coverage
