+++ scripts/__dropins__/storefront-cart/containers/EmptyCart.js
@@ -1,4 +1,4 @@
 /*! Copyright 2026 Adobe
 All Rights Reserved. */
-import{E as u,E as x}from"../chunks/EmptyCart.js";import"@dropins/tools/preact-jsx-runtime.js";import"../chunks/components.js";import"@dropins/tools/lib.js";import"@dropins/tools/i18n.js";import"@dropins/tools/preact-compat.js";import"@dropins/tools/components.js";import"@dropins/tools/preact-hooks.js";import"../api.js";import"@dropins/tools/event-bus.js";import"../fragments.js";import"@dropins/tools/fetch-graphql.js";export{u as EmptyCart,x as default};
+import{E as s,E as d}from"../chunks/EmptyCart.js";import"@dropins/tools/preact-jsx-runtime.js";import"../chunks/EmptyCart2.js";import"@dropins/tools/lib.js";import"@dropins/tools/components.js";/* empty css                        */import"@dropins/tools/preact-compat.js";import"@dropins/tools/i18n.js";import"@dropins/tools/preact-hooks.js";export{s as EmptyCart,d as default};
 //# sourceMappingURL=EmptyCart.js.map
