+++ scripts/__dropins__/storefront-cart/containers/EstimateShipping.js
@@ -1,4 +1,4 @@
 /*! Copyright 2026 Adobe
 All Rights Reserved. */
-import{jsx as a}from"@dropins/tools/preact-jsx-runtime.js";import"@dropins/tools/lib.js";import{Button as X,Input as R,Picker as U,Price as N}from"@dropins/tools/components.js";import{e as k}from"../chunks/components.js";import{useState as p,useEffect as D,useCallback as M}from"@dropins/tools/preact-compat.js";import"@dropins/tools/preact-hooks.js";import{getCountries as j,getRegions as $,getEstimateShipping as q,a as A,s as G}from"../api.js";import{events as B}from"@dropins/tools/event-bus.js";import{useText as w}from"@dropins/tools/i18n.js";import"../fragments.js";import"@dropins/tools/fetch-graphql.js";const H=()=>{const[b,F]=p(!1),[T,P]=p([]),[_,m]=p("US"),[h,S]=p(""),[n,y]=p(""),[f,L]=p([]),[g,d]=p(!1),[x,v]=p(),[C,t]=p(),[o,E]=p(""),[e,u]=p(!1),I=()=>{m("US"),S(""),y(""),v(null),t(null),E(""),u(!1)},Z=async c=>{const{shippingCountry:i,shippingState:r="",shippingZip:s=""}=c,V={countryCode:i,postcode:s,region:{region:r}};return F(!0),q(V).then(l=>(l&&(v({amount:l.amount.value,currency:l.amount.currency,priceIncludingtax:{amount:l.price_incl_tax.value,currency:l.price_incl_tax.currency},priceExcludingtax:{amount:l.price_excl_tax.value,currency:l.price_excl_tax.currency}}),t({carrier_code:l.carrier_code,method_code:l.method_code}),m(i),S(r),y(s),E(r||s||i),u(!0)),m(i),S(r),y(s),E(r||s||i),l)).finally(()=>{F(!1)})},z=c=>{c.preventDefault(),S(""),y("");const i=c.target.value;m(i)};return D(()=>{j().then(c=>{let i="US";const r=c.map(s=>(s.isDefaultCountry&&(i=s.id),{text:s.label,value:s.id}));P(r),m(i)})},[]),D(()=>{d(!0),$(_).then(c=>{const i=c.map(r=>({text:r.name,value:r.code}));L(i)}).finally(()=>{d(!1)})},[_,d]),{loading:b,regionsLoading:g,estimatedDestinationText:o,countries:T,selectedCountry:_,selectedRegion:h,selectedZip:n,regions:f,estimatedShippingPrice:x,estimatedShippingMethod:C,shippingEstimated:e,handleEstimateShipping:Z,handleCountrySelected:z,resetValues:I,setPriceSummaryLoading:F}},se=({showDefaultEstimatedShippingCost:b})=>{var E;const[F,T]=p(!1),{loading:P,countries:_,regions:m,selectedCountry:h,estimatedDestinationText:S,estimatedShippingPrice:n,handleCountrySelected:y,handleEstimateShipping:f,regionsLoading:L,selectedRegion:g,selectedZip:d,shippingEstimated:x,resetValues:v}=H(),C=M(e=>{f(e).then(()=>{A(e)})},[f]);D(()=>{const e=B.on("cart/data",u=>{var i,r,s;T((u==null?void 0:u.isVirtual)||!1);const I=(r=(i=u==null?void 0:u.addresses)==null?void 0:i.shipping)==null?void 0:r[0];if(b&&!I&&f({shippingCountry:((s=G.config)==null?void 0:s.defaultCountry)??""}),!I)return;const{countryCode:Z,regionCode:z,zipCode:c}=I;C({shippingCountry:Z,shippingState:z,shippingZip:c})},{eager:!0});return()=>{e==null||e.off()}},[]),D(()=>{const e=B.on("cart/updated",()=>{x&&f({shippingCountry:h,shippingState:g,shippingZip:d})});return()=>{e==null||e.off()}},[x,h,g,d,f]),D(()=>{const e=B.on("cart/reset",()=>{v(),A(null)});return()=>{e==null||e.off()}},[v]),D(()=>{const e=B.on("cart/merged",()=>{x&&C({shippingCountry:h,shippingState:g,shippingZip:d})});return()=>{e==null||e.off()}},[x,h,g,d,C]);const t=w({applyButton:"Cart.PriceSummary.estimatedShippingForm.apply.label",countryField:"Cart.PriceSummary.estimatedShippingForm.country.placeholder",freeShipping:"Cart.PriceSummary.freeShipping",stateField:"Cart.PriceSummary.estimatedShippingForm.state.placeholder",taxToBeDetermined:"Cart.PriceSummary.taxToBeDetermined",zipField:"Cart.PriceSummary.estimatedShippingForm.zip.placeholder"});if(F)return null;const o=(E=G.config)==null?void 0:E.shoppingCartDisplaySetting;return a(k,{loading:P,taxIncluded:(o==null?void 0:o.shipping)==="INCLUDING_TAX",taxExcluded:(o==null?void 0:o.shipping)==="INCLUDING_EXCLUDING_TAX",price:(n==null?void 0:n.amount)==0?a("span",{"data-testId":"free-shipping",children:t.freeShipping}):(o==null?void 0:o.shipping)==="INCLUDING_TAX"&&n?a(N,{"data-testid":"shipping",...n.priceIncludingtax}):n?a(N,{...n}):a("span",{children:t.taxToBeDetermined}),estimated:!0,priceExcludingTax:n!=null&&n.priceExcludingtax?a(N,{"data-testid":"shipping-excluding-tax",...n.priceExcludingtax}):a("span",{children:t.taxToBeDetermined}),countryField:a(U,{name:"shippingCountry",placeholder:t.countryField,value:h,variant:"primary",options:_,handleSelect:y,"data-testid":"estimate-shipping-country-selector"}),stateField:m.length>0?a(U,{name:"shippingState",placeholder:t.stateField,variant:"primary",options:m,value:g,"data-testid":"estimate-shipping-state-selector",disabled:L}):a(R,{"aria-label":t.stateField,name:"shippingState",placeholder:t.stateField,variant:"primary",value:g,disabled:L,"data-testid":"estimate-shipping-state-input",maxLength:50}),zipField:a(R,{"aria-label":t.zipField,name:"shippingZip",placeholder:t.zipField,variant:"primary","data-testid":"estimate-shipping-zip-input",value:d,maxLength:12}),estimateButton:a(X,{variant:"secondary","data-testid":"estimate-shipping-apply-button","aria-label":t.applyButton,children:t.applyButton}),destinationText:S||t.taxToBeDetermined,onEstimate:C})};export{se as EstimateShipping,se as default};
+import{jsxs as T,jsx as e,Fragment as V}from"@dropins/tools/preact-jsx-runtime.js";import{classes as D,VComponent as B,getFormValues as $}from"@dropins/tools/lib.js";import{Button as q,Input as w,Picker as G,Price as U}from"@dropins/tools/components.js";/* empty css                        */import{useRef as H,useState as s,useEffect as Z,useCallback as J}from"@dropins/tools/preact-compat.js";import{useText as M,Text as k}from"@dropins/tools/i18n.js";import"@dropins/tools/preact-hooks.js";import{s as X}from"../chunks/resetCart.js";import{events as R}from"@dropins/tools/event-bus.js";import{s as j}from"../chunks/persisted-data.js";import{a as O,b as Q,g as W}from"../chunks/getEstimateShipping.js";import"@dropins/tools/fetch-graphql.js";const Y=({countryField:z,destinationText:f,estimateButton:I,estimated:A,loading:N,onEstimate:l,price:S,priceExcludingTax:F,priceIncludingTax:n,stateField:C,taxExcluded:y,taxIncluded:P,zipField:m})=>{const d=H(null),[L,x]=s(!0),[E,i]=s("zip"),r=M({editZipAction:"Cart.EstimateShipping.editZipAction",destinationLinkAriaLabel:"Cart.EstimateShipping.destinationLinkAriaLabel",shippingLabel:"Cart.EstimateShipping.label",zipPlaceholder:"Cart.EstimateShipping.zipPlaceholder"}),_=p=>{p.preventDefault(),x(b=>!b)},t=p=>{p.preventDefault(),x(!0),i(b=>b==="zip"?"state":"zip")},h=p=>{p.preventDefault(),x(!1);const b=$(d.current);l==null||l(b)},v=r.destinationLinkAriaLabel.replace("{destination}",f);return T("div",{"data-testid":"estimate-shipping",className:D(["cart-estimate-shipping",["cart-estimate-shipping--loading",N]]),children:[e("span",{className:"cart-estimate-shipping__label",children:A?f?T(V,{children:[e(k,{id:"Cart.EstimateShipping.estimatedDestination"})," ",e("a",{className:"cart-estimate-shippingLink",role:"button",href:"",onClick:_,onKeyDown:p=>{(p.key==="Enter"||p.key===" ")&&_(p)},tabIndex:0,"aria-label":v,"data-testid":"shipping-destination-link",children:f})]}):e(k,{id:"Cart.EstimateShipping.estimated"}):e(k,{id:"Cart.EstimateShipping.label"})}),e(B,{node:S,className:"cart-estimate-shipping__price"}),A&&T(V,{children:[e("div",{className:D(["cart-estimate-shipping__caption"]),children:e("a",{href:"#",className:"cart-estimate-shipping__link",onClick:t,"data-testid":"shipping-alternate-field-link",children:E==="zip"?e(k,{id:"Cart.EstimateShipping.alternateField.state"}):e(k,{id:"Cart.EstimateShipping.alternateField.zip"})})}),T("form",{className:D(["cart-estimate-shipping--edit",["cart-estimate-shipping--hide",!L]]),ref:d,"data-testid":"shipping-estimate-form",children:[z&&e(B,{node:z,className:D(["cart-estimate-shipping--country"])}),E==="state"?C&&e(B,{node:C,className:D(["cart-estimate-shipping--state"])}):m&&e(B,{node:m,className:D(["cart-estimate-shipping--zip"])}),I&&e(B,{node:I,className:D(["cart-estimate-shipping--action"]),onClick:h,type:"submit"})]})]}),P&&e("div",{"data-testid":"shipping-tax-included",className:D(["cart-estimate-shipping__caption"]),children:T("span",{children:[n," ",e(k,{id:"Cart.EstimateShipping.withTaxes"})]})}),y?e("div",{"data-testid":"shipping-tax-included-excluded",className:D(["cart-estimate-shipping__caption"]),children:T("span",{children:[F," ",e(k,{id:"Cart.EstimateShipping.withoutTaxes"})]})}):void 0]})},ee=()=>{const[z,f]=s(!1),[I,A]=s([]),[N,l]=s("US"),[S,F]=s(""),[n,C]=s(""),[y,P]=s([]),[m,d]=s(!1),[L,x]=s(),[E,i]=s(),[r,_]=s(""),[t,h]=s(!1),v=()=>{l("US"),F(""),C(""),x(null),i(null),_(""),h(!1)},p=async u=>{const{shippingCountry:a,shippingState:o="",shippingZip:c=""}=u,K={countryCode:a,postcode:c,region:{region:o}};return f(!0),W(K).then(g=>(g&&(x({amount:g.amount.value,currency:g.amount.currency,priceIncludingtax:{amount:g.price_incl_tax.value,currency:g.price_incl_tax.currency},priceExcludingtax:{amount:g.price_excl_tax.value,currency:g.price_excl_tax.currency}}),i({carrier_code:g.carrier_code,method_code:g.method_code}),l(a),F(o),C(c),_(o||c||a),h(!0)),l(a),F(o),C(c),_(o||c||a),g)).finally(()=>{f(!1)})},b=u=>{u.preventDefault(),F(""),C("");const a=u.target.value;l(a)};return Z(()=>{O().then(u=>{let a="US";const o=u.map(c=>(c.isDefaultCountry&&(a=c.id),{text:c.label,value:c.id}));A(o),l(a)})},[]),Z(()=>{d(!0),Q(N).then(u=>{const a=u.map(o=>({text:o.name,value:o.code}));P(a)}).finally(()=>{d(!1)})},[N,d]),{loading:z,regionsLoading:m,estimatedDestinationText:r,countries:I,selectedCountry:N,selectedRegion:S,selectedZip:n,regions:y,estimatedShippingPrice:L,estimatedShippingMethod:E,shippingEstimated:t,handleEstimateShipping:p,handleCountrySelected:b,resetValues:v,setPriceSummaryLoading:f}},he=({showDefaultEstimatedShippingCost:z})=>{var _;const[f,I]=s(!1),{loading:A,countries:N,regions:l,selectedCountry:S,estimatedDestinationText:F,estimatedShippingPrice:n,handleCountrySelected:C,handleEstimateShipping:y,regionsLoading:P,selectedRegion:m,selectedZip:d,shippingEstimated:L,resetValues:x}=ee(),E=J(t=>{y(t).then(()=>{j(t)})},[y]);Z(()=>{const t=R.on("cart/data",h=>{var a,o,c;I((h==null?void 0:h.isVirtual)||!1);const v=(o=(a=h==null?void 0:h.addresses)==null?void 0:a.shipping)==null?void 0:o[0];if(z&&!v&&y({shippingCountry:((c=X.config)==null?void 0:c.defaultCountry)??""}),!v)return;const{countryCode:p,regionCode:b,zipCode:u}=v;E({shippingCountry:p,shippingState:b,shippingZip:u})},{eager:!0});return()=>{t==null||t.off()}},[]),Z(()=>{const t=R.on("cart/updated",()=>{L&&y({shippingCountry:S,shippingState:m,shippingZip:d})});return()=>{t==null||t.off()}},[L,S,m,d,y]),Z(()=>{const t=R.on("cart/reset",()=>{x(),j(null)});return()=>{t==null||t.off()}},[x]),Z(()=>{const t=R.on("cart/merged",()=>{L&&E({shippingCountry:S,shippingState:m,shippingZip:d})});return()=>{t==null||t.off()}},[L,S,m,d,E]);const i=M({applyButton:"Cart.PriceSummary.estimatedShippingForm.apply.label",countryField:"Cart.PriceSummary.estimatedShippingForm.country.placeholder",freeShipping:"Cart.PriceSummary.freeShipping",stateField:"Cart.PriceSummary.estimatedShippingForm.state.placeholder",taxToBeDetermined:"Cart.PriceSummary.taxToBeDetermined",zipField:"Cart.PriceSummary.estimatedShippingForm.zip.placeholder"});if(f)return null;const r=(_=X.config)==null?void 0:_.shoppingCartDisplaySetting;return e(Y,{loading:A,taxIncluded:(r==null?void 0:r.shipping)==="INCLUDING_TAX",taxExcluded:(r==null?void 0:r.shipping)==="INCLUDING_EXCLUDING_TAX",price:(n==null?void 0:n.amount)==0?e("span",{"data-testId":"free-shipping",children:i.freeShipping}):(r==null?void 0:r.shipping)==="INCLUDING_TAX"&&n?e(U,{"data-testid":"shipping",...n.priceIncludingtax}):n?e(U,{...n}):e("span",{children:i.taxToBeDetermined}),estimated:!0,priceExcludingTax:n!=null&&n.priceExcludingtax?e(U,{"data-testid":"shipping-excluding-tax",...n.priceExcludingtax}):e("span",{children:i.taxToBeDetermined}),countryField:e(G,{name:"shippingCountry",placeholder:i.countryField,value:S,variant:"primary",options:N,handleSelect:C,"data-testid":"estimate-shipping-country-selector"}),stateField:l.length>0?e(G,{name:"shippingState",placeholder:i.stateField,variant:"primary",options:l,value:m,"data-testid":"estimate-shipping-state-selector",disabled:P}):e(w,{"aria-label":i.stateField,name:"shippingState",placeholder:i.stateField,variant:"primary",value:m,disabled:P,"data-testid":"estimate-shipping-state-input",maxLength:50}),zipField:e(w,{"aria-label":i.zipField,name:"shippingZip",placeholder:i.zipField,variant:"primary","data-testid":"estimate-shipping-zip-input",value:d,maxLength:12}),estimateButton:e(q,{variant:"secondary","data-testid":"estimate-shipping-apply-button","aria-label":i.applyButton,children:i.applyButton}),destinationText:F||i.taxToBeDetermined,onEstimate:E})};export{he as EstimateShipping,he as default};
 //# sourceMappingURL=EstimateShipping.js.map
