+++ scripts/__dropins__/storefront-cart/containers/MiniCart.js
@@ -1,4 +1,4 @@
 /*! Copyright 2026 Adobe
 All Rights Reserved. */
-import{jsx as r,jsxs as w}from"@dropins/tools/preact-jsx-runtime.js";import{useState as B,useEffect as H,useCallback as X}from"@dropins/tools/preact-compat.js";import{Slot as o}from"@dropins/tools/lib.js";import{s as m,p as O,updateProductsFromCart as j,getCartDataFromCache as Y}from"../api.js";import{events as V}from"@dropins/tools/event-bus.js";import{Button as h,Price as C}from"@dropins/tools/components.js";import{M as $}from"../chunks/components.js";import"@dropins/tools/preact-hooks.js";import"../chunks/CartSummaryGrid.js";import{C as q}from"../chunks/CartSummaryList.js";import"../chunks/OrderSummary.js";import"../chunks/CartSummaryTable.js";import{useText as z}from"@dropins/tools/i18n.js";import"../fragments.js";import"@dropins/tools/fetch-graphql.js";import"../chunks/EmptyCart.js";const J=100,K=({children:W,initialData:x=null,hideFooter:L=!0,slots:t,routeProduct:g,routeCart:d,routeCheckout:a,routeEmptyCartCTA:v,showDiscount:D,showSavings:k,enableItemRemoval:P=!0,enableQuantityUpdate:T=!1,hideHeading:S=!1,undo:A=!1,confirmBeforeDelete:F=!1,...N})=>{var y,b,f;const[e,Q]=B(x),n=(y=m.config)==null?void 0:y.shoppingCartDisplaySetting;H(()=>{const i=V.on("cart/data",u=>{Q(u)},{eager:!0});return()=>{i==null||i.off()}},[]);const p=z({cartLink:"Cart.MiniCart.cartLink",checkoutLink:"Cart.MiniCart.checkoutLink"}),I=(i,u)=>j([{uid:i,quantity:u}]),U=i=>I(i,0),c=e==null?void 0:e.hasOutOfStockItems,M=X(i=>{e&&!c&&(i.preventDefault(),O(e,m.locale),setTimeout(()=>{const u=a==null?void 0:a();u&&(window.location.href=u)},J))},[e,c,a]),E=r(o,{name:"ProductList",slot:t==null?void 0:t.ProductList,context:{itemQuantityUpdateHandler:I,itemRemoveHandler:U,totalQuantity:e==null?void 0:e.totalQuantity},children:r(q,{"data-testid":"default-cart-summary-list",routeProduct:g,routeEmptyCartCTA:v,initialData:e,maxItems:(b=m.config)==null?void 0:b.miniCartMaxItemsDisplay,showMaxItems:!!((f=m.config)!=null&&f.miniCartMaxItemsDisplay),hideHeading:S||!(e!=null&&e.totalQuantity),hideFooter:L,enableRemoveItem:P,enableUpdateItemQuantity:T,showDiscount:D,showSavings:k,undo:A,confirmBeforeDelete:F,slots:{Thumbnail:t==null?void 0:t.Thumbnail,Heading:t==null?void 0:t.Heading,EmptyCart:t==null?void 0:t.EmptyCart,Footer:t==null?void 0:t.Footer,RowTotalFooter:t==null?void 0:t.RowTotalFooter,ProductAttributes:t==null?void 0:t.ProductAttributes,CartSummaryFooter:t==null?void 0:t.CartSummaryFooter,CartItem:t==null?void 0:t.CartItem,UndoBanner:t==null?void 0:t.UndoBanner,ConfirmDeleteBanner:t==null?void 0:t.ConfirmDeleteBanner,ItemTitle:t==null?void 0:t.ItemTitle,ItemPrice:t==null?void 0:t.ItemPrice,ItemQuantity:t==null?void 0:t.ItemQuantity,ItemTotal:t==null?void 0:t.ItemTotal,ItemSku:t==null?void 0:t.ItemSku,ItemRemoveAction:t==null?void 0:t.ItemRemoveAction}})}),_=i=>r(o,{name:"ProductListFooter",slot:t==null?void 0:t.ProductListFooter,context:{data:i}}),R=i=>r(o,{name:"PreCheckoutSection",slot:t==null?void 0:t.PreCheckoutSection,context:{data:i}}),G=()=>(n==null?void 0:n.subtotal)==="INCLUDING_TAX"||(n==null?void 0:n.subtotal)==="INCLUDING_EXCLUDING_TAX"?{amount:e==null?void 0:e.subtotal.includingTax.value,currency:e==null?void 0:e.subtotal.includingTax.currency,"data-testid":"subtotal-including-tax",style:{font:"inherit"}}:{amount:e==null?void 0:e.subtotal.excludingTax.value,currency:e==null?void 0:e.subtotal.excludingTax.currency,"data-testid":"subtotal-excluding-tax",style:{font:"inherit"}};return r($,{...N,productListFooter:_(e),subtotal:e!=null&&e.totalQuantity?(e==null?void 0:e.subtotal)&&r(C,{...G()}):void 0,subtotalExcludingTaxes:e!=null&&e.totalQuantity?(e==null?void 0:e.subtotal)&&((n==null?void 0:n.subtotal)==="INCLUDING_EXCLUDING_TAX"?r(C,{amount:e==null?void 0:e.subtotal.excludingTax.value,currency:e==null?void 0:e.subtotal.excludingTax.currency,"data-testid":"subtotal-including-excluding-tax",style:{font:"inherit"}}):void 0):void 0,preCheckoutSection:R(e),ctas:e!=null&&e.totalQuantity?w("div",{children:[a&&r(h,{"data-testid":"route-checkout-button",variant:"primary",href:c?void 0:a(),disabled:c,"aria-disabled":c,onClick:M,children:p.checkoutLink}),d&&r(h,{"data-testid":"route-cart-button",variant:"tertiary",href:d(),children:p.cartLink})]}):void 0,products:E})};K.getInitialData=async function(){return Y()};export{K as MiniCart,K as default};
+import{jsx as e,jsxs as d,Fragment as B}from"@dropins/tools/preact-jsx-runtime.js";import{useState as O,useEffect as j,useCallback as V}from"@dropins/tools/preact-compat.js";import{classes as h,VComponent as o,Slot as _}from"@dropins/tools/lib.js";import{g as Y}from"../chunks/persisted-data.js";import{events as $}from"@dropins/tools/event-bus.js";import{Button as T,Price as k}from"@dropins/tools/components.js";/* empty css                        */import{useText as P}from"@dropins/tools/i18n.js";import"@dropins/tools/preact-hooks.js";import{s as f}from"../chunks/resetCart.js";import{a as q}from"../chunks/acdl.js";import{u as z}from"../chunks/updateProductsFromCart.js";import"../chunks/CartSummaryGrid.js";import{C as J}from"../chunks/CartSummaryList.js";import"../chunks/OrderSummary.js";import"../chunks/CartSummaryTable2.js";import"@dropins/tools/fetch-graphql.js";import"../chunks/refreshCart.js";import"../fragments.js";import"../chunks/EmptyCart2.js";import"../chunks/createGuestCart.js";import"../chunks/WarningWithCircle.js";import"../chunks/ChevronDown.js";import"../chunks/getEstimatedTotals.js";import"../chunks/OrderSummaryLine.js";import"../chunks/ChevronUp.js";import"../chunks/Coupon.js";import"../chunks/GiftCard.js";import"../chunks/EmptyCart.js";import"../chunks/Close.js";const K=({className:x,products:p,productListFooter:b,subtotal:t,subtotalExcludingTaxes:I,preCheckoutSection:m,ctas:a,...l})=>{const y=P({subtotal:"Cart.MiniCart.subtotal",subtotalExcludingTaxes:"Cart.MiniCart.subtotalExcludingTaxes"});return e("div",{...l,className:h(["cart-mini-cart",x]),children:p&&d(B,{children:[e("div",{className:"cart-mini-cart__products","data-testid":"mini-cart-products-wrapper",children:p}),b&&e("div",{className:h(["cart-mini-cart__productListFooter"]),"data-testid":"mini-cart-product-list-footer",children:e(o,{node:b})}),d("div",{className:"cart-mini-cart__footer","data-testid":"mini-cart-subtotals",children:[t&&d("div",{className:"cart-mini-cart__footer__estimated-total","data-testid":"mini-cart-subtotal",children:[y.subtotal,e(o,{node:t})]}),I&&d("div",{className:"cart-mini-cart__footer__estimated-total-excluding-taxes","data-testid":"mini-cart-subtotal-excluding-taxes",children:[y.subtotalExcludingTaxes,e(o,{node:I,className:h(["dropin-price-summary__price","dropin-price-summary__price--muted"])})]}),m&&e("div",{className:h(["cart-mini-cart__preCheckoutSection"]),"data-testid":"mini-cart-pre-checkout-section",children:e(o,{node:m})}),a&&e(o,{node:a,className:"cart-mini-cart__footer__ctas"})]})]})})},W=100,Z=({children:x,initialData:p=null,hideFooter:b=!0,slots:t,routeProduct:I,routeCart:m,routeCheckout:a,routeEmptyCartCTA:l,showDiscount:y,showSavings:S,enableItemRemoval:D=!0,enableQuantityUpdate:A=!1,hideHeading:F=!1,undo:Q=!1,...U})=>{var C,L,N;const[i,E]=O(p),n=(C=f.config)==null?void 0:C.shoppingCartDisplaySetting;j(()=>{const r=$.on("cart/data",c=>{E(c)},{eager:!0});return()=>{r==null||r.off()}},[]);const g=P({cartLink:"Cart.MiniCart.cartLink",checkoutLink:"Cart.MiniCart.checkoutLink"}),v=(r,c)=>z([{uid:r,quantity:c}]),M=r=>v(r,0),u=i==null?void 0:i.hasOutOfStockItems,R=V(r=>{i&&!u&&(r.preventDefault(),q(i,f.locale),setTimeout(()=>{const c=a==null?void 0:a();c&&(window.location.href=c)},W))},[i,u,a]),w=e(_,{name:"ProductList",slot:t==null?void 0:t.ProductList,context:{itemQuantityUpdateHandler:v,itemRemoveHandler:M,totalQuantity:i==null?void 0:i.totalQuantity},children:e(J,{"data-testid":"default-cart-summary-list",routeProduct:I,routeEmptyCartCTA:l,initialData:i,maxItems:(L=f.config)==null?void 0:L.miniCartMaxItemsDisplay,showMaxItems:!!((N=f.config)!=null&&N.miniCartMaxItemsDisplay),hideHeading:F||!(i!=null&&i.totalQuantity),hideFooter:b,enableRemoveItem:D,enableUpdateItemQuantity:A,showDiscount:y,showSavings:S,undo:Q,slots:{Thumbnail:t==null?void 0:t.Thumbnail,Heading:t==null?void 0:t.Heading,EmptyCart:t==null?void 0:t.EmptyCart,Footer:t==null?void 0:t.Footer,RowTotalFooter:t==null?void 0:t.RowTotalFooter,ProductAttributes:t==null?void 0:t.ProductAttributes,CartSummaryFooter:t==null?void 0:t.CartSummaryFooter,CartItem:t==null?void 0:t.CartItem,UndoBanner:t==null?void 0:t.UndoBanner,ItemTitle:t==null?void 0:t.ItemTitle,ItemPrice:t==null?void 0:t.ItemPrice,ItemQuantity:t==null?void 0:t.ItemQuantity,ItemTotal:t==null?void 0:t.ItemTotal,ItemSku:t==null?void 0:t.ItemSku,ItemRemoveAction:t==null?void 0:t.ItemRemoveAction}})}),G=r=>e(_,{name:"ProductListFooter",slot:t==null?void 0:t.ProductListFooter,context:{data:r}}),H=r=>e(_,{name:"PreCheckoutSection",slot:t==null?void 0:t.PreCheckoutSection,context:{data:r}}),X=()=>(n==null?void 0:n.subtotal)==="INCLUDING_TAX"||(n==null?void 0:n.subtotal)==="INCLUDING_EXCLUDING_TAX"?{amount:i==null?void 0:i.subtotal.includingTax.value,currency:i==null?void 0:i.subtotal.includingTax.currency,"data-testid":"subtotal-including-tax",style:{font:"inherit"}}:{amount:i==null?void 0:i.subtotal.excludingTax.value,currency:i==null?void 0:i.subtotal.excludingTax.currency,"data-testid":"subtotal-excluding-tax",style:{font:"inherit"}};return e(K,{...U,productListFooter:G(i),subtotal:i!=null&&i.totalQuantity?(i==null?void 0:i.subtotal)&&e(k,{...X()}):void 0,subtotalExcludingTaxes:i!=null&&i.totalQuantity?(i==null?void 0:i.subtotal)&&((n==null?void 0:n.subtotal)==="INCLUDING_EXCLUDING_TAX"?e(k,{amount:i==null?void 0:i.subtotal.excludingTax.value,currency:i==null?void 0:i.subtotal.excludingTax.currency,"data-testid":"subtotal-including-excluding-tax",style:{font:"inherit"}}):void 0):void 0,preCheckoutSection:H(i),ctas:i!=null&&i.totalQuantity?d("div",{children:[a&&e(T,{"data-testid":"route-checkout-button",variant:"primary",href:u?void 0:a(),disabled:u,"aria-disabled":u,onClick:R,children:g.checkoutLink}),m&&e(T,{"data-testid":"route-cart-button",variant:"tertiary",href:m(),children:g.cartLink})]}):void 0,products:w})};Z.getInitialData=async function(){return Y()};export{Z as MiniCart,Z as default};
 //# sourceMappingURL=MiniCart.js.map
