+++ scripts/__dropins__/storefront-cart/containers/MiniCart/MiniCart.d.ts
@@ -32,12 +32,6 @@ export interface MiniCartProps extends HTMLAttributes<HTMLDivElement> {
             onUndo: () => void;
             onDismiss: () => void;
         }>;
-        ConfirmDeleteBanner?: SlotProps<{
-            item: CartModel['items'][0];
-            loading: boolean;
-            onConfirm: () => void;
-            onCancel: () => void;
-        }>;
         ItemTitle?: SlotProps<{
             item: CartModel['items'][number];
         }>;
@@ -46,9 +40,7 @@ export interface MiniCartProps extends HTMLAttributes<HTMLDivElement> {
         }>;
         ItemQuantity?: SlotProps<{
             item: CartModel['items'][number];
-            enableUpdateItemQuantity: boolean | {
-                removeOnZero?: boolean;
-            };
+            enableUpdateItemQuantity: boolean;
             handleItemQuantityUpdate: (item: CartModel['items'][number], quantity: number) => void;
             itemsLoading: Set<string>;
             handleItemsError: (uid: string, message?: string) => void;
@@ -80,12 +72,9 @@ export interface MiniCartProps extends HTMLAttributes<HTMLDivElement> {
     showDiscount?: boolean;
     showSavings?: boolean;
     enableItemRemoval?: boolean;
-    enableQuantityUpdate?: boolean | {
-        removeOnZero?: boolean;
-    };
+    enableQuantityUpdate?: boolean;
     hideHeading?: boolean;
     undo?: boolean;
-    confirmBeforeDelete?: boolean;
 }
 export declare const MiniCart: Container<MiniCartProps, CartModel | null>;
 //# sourceMappingURL=MiniCart.d.ts.map
